Cluster genomic regions by distance from region center

Cluster genomic regions by distance from region center

clusterGPosition(
  dt,
  distance = 10000000,
  no_annotation = FALSE,
  genome_build = c("hg38", "hg19"),
  simplify = TRUE
)

Arguments

dt

a genomic region data with first 3 columns for chromosome, start and end.

distance

a distance value for clustering, regions with average distance lower than this value would be clustered into one. Should not greater than 1e8.

no_annotation

if TRUE, use builtin data for annotation. This is useful when input a list of genes.

genome_build

genome build version.

simplify

if FALSE, return a data.table instead of a vector.

Value

cluster list.

Examples

library(data.table)
region <- data.table(
  chr = c("chr1", "chr2", "chr1"),
  start = c(1032, 992, 10000), end = c(4242, 1e6, 400023)
)
clusterGPosition(region, distance = 1000)
clusterGPosition(region)
clusterGPosition(data.table(gene_id = c("KRAS", "EGFR", "MYC", "ERBB2", "GRB7")),
  no_annotation = TRUE, simplify = FALSE
)
clusterGPosition(data.table(gene_id = c("KRAS", "EGFR", "MYC", "ERBB2", "GRB7")),
  no_annotation = TRUE, simplify = TRUE
)