R6 class representing focal copy number amplification list predicted from a cohort

Contains fields storing data and methods to get, process and visualize fCNA information. Examples please see gcap.ASCNworkflow().

Public fields

data

a data.table storing fCNA list, which typically contains following columns:

sample sample or case ID.

band chromosome cytoband.

gene_id gene ID, typically Ensembl ID. You can convert the ID with R package IDConverter.

total_cn total copy number value.

minor_cn copy number value for minor allele.

background_cn background copy number value (for circular amplicon), calculated as pmax(blood_cn_top5 + tightness * blood_cn_top5_sd, 2) * ploidy / 2. (Optional) At default, blood_cn_top5(_sd) is obtained from SNP array data of ~2000 TCGA diploidy blood samples, they are mean and sd of top 5% copy number values for each gene (in ~2000 samples); ploidy is tumor ploidy.

prob the probability the gene located in circular DNA.

amplicon_type the type of DNA amplicon.

sample_summary

a data.table storing sample summary data, which typically contains at least the following columns:

sample sample or case ID. Should only include cases have been called with GCAP workflow, otherwise the extra cases would be automatically classified as 'nofocal' (i.e. NA in sample_summary field) class.

purity, ploidy for tumor purity or ploidy.

AScore aneuploidy score.

pLOH genome percentage harboring LOH events.

CN1 ... CN19 activity of copy number signatures.

class the sample class based on amplicon type.

ec_genes number of genes predicted as located on circular DNA.

ec_possibly_genes same with ec_genes but with less confidence.

ec_cytobands number of cytobands predicted as located on circular DNA. (the regions of ec_possibly_genes are not included in computation)

gene_summary

a data.table storing gene summary data.

cytoband_summary

a data.table storing cytoband summary data.

Active bindings

min_n

check $new() method for details. If you updated this value, a function will be called to update the sample summary.

Methods

Public methods

fCNA$new()

fCNA$subset()

fCNA$getSampleSummary()

fCNA$getGeneSummary()

fCNA$getCytobandSummary()

fCNA$saveToFiles()

fCNA$convertGeneID()

fCNA$print()

Method new()

Create a fCNA object. Typically, you can obtain this object from gcap.workflow() or gcap.ASCNworkflow().

Usage

fCNA$new(fcna, pdata = fcna[, "sample", drop = FALSE], min_n = 1L)

Arguments

fcna

a data.frame storing focal copy number amplicon list.

pdata

a data.frame storing phenotype or sample-level related data. (Optional)

min_n

a minimal cytoband number (default is 1) to determine sample class. e.g., sample with at least 1 cytoband harboring circular genes would be labelled as "circular". This only affect sample_summary field.

Method subset()

Return a subset fCNA object

Usage

fCNA$subset(..., on = c("data", "sample_summary"))

Arguments

...

subset expressions on fCNA$data or fCNA$sample_summary.

on

if it is "data", subset operations are on data field of fCNA object, same for "sample_summary".

Returns

a fCNA

Method getSampleSummary()

Get sample summary of fCNA

Usage

fCNA$getSampleSummary()

Returns

a data.table

Method getGeneSummary()

Get gene level summary of fCNA type

Usage

fCNA$getGeneSummary()

Returns

a data.table

Method getCytobandSummary()

Get cytoband level summary of fCNA type

Usage

fCNA$getCytobandSummary()

Returns

a data.table

Method saveToFiles()

Save the key data to local files

Usage

fCNA$saveToFiles(dirpath, fileprefix = "fCNA")

Arguments

dirpath

directory path storing output files.

fileprefix

file prefix. Two result files shall be generated.

Method convertGeneID()

Convert Gene IDs between Ensembl and Hugo Symbol System

Usage

fCNA$convertGeneID(
  type = c("ensembl", "symbol"),
  genome_build = c("hg38", "hg19", "mm10", "mm9")
)

Arguments

type

type of input IDs, could be 'ensembl' or 'symbol'.

genome_build

reference genome build.

Method print()

print the fCNA object

Usage

fCNA$print(...)

Arguments

...

unused.