Draw Gene Amplicon Distribution By Gene or User Specified Class

See gcap.plotProfile() for examples.

gcap.plotDistribution(
  fCNA,
  x = NULL,
  x_size = 8,
  by = c("gene", "sample"),
  merge_circular = TRUE,
  fill = TRUE,
  palette = c("#CCCCCC", "#0066CC", "#FFCCCC", "#CC0033"),
  ...
)

Arguments

fCNA

a fCNA object.

x

a list of gene or a column name in fCNA$sample_summary.

x_size

font size of x axis text.

by

"gene" (default) or "sample" mode. When "sample" is set, it uses the column specified by x to grouping samples.

merge_circular

if TRUE, merge 'possibly_circular' into 'circular' type.

fill

if TRUE, show the percentage instead of count.

palette

color palette.

...

other parameters passing to ggplot2::geom_bar.

Value

a ggplot object.

See also