Draw Forest Plot for fCNA+/- Comparison by Sample or Gene

See gcap.plotProfile() for examples.

gcap.plotForest(
  fCNA,
  response_data,
  f = c("coxph", "binomial", "gaussian", "Gamma", "inverse.gaussian", "poisson",
    "quasi", "quasibinomial", "quasipoisson"),
  x = "class",
  covars = NULL,
  y = if (is.data.frame(response_data)) {     colnames(response_data)[-1] } else {    
    response_data },
  merge_circular = TRUE,
  x_is_gene = FALSE,
  gene_focus = c("fCNA", "circular", "vs"),
  optimize_model = FALSE,
  ending_time = NULL,
  parallel = FALSE,
  exp = NULL,
  ref_line = NULL,
  xlim = NULL,
  ...
)

Arguments

fCNA

a fCNA object.

response_data

a data.frame or specified column names in fCNA$sample_summary prepared regression modeling, default is CoxPH model, you can use same data as gcap.plotKMcurve.

f

a length-1 string specifying modeling function or family of glm(), default is 'coxph'. Other options are members of GLM family, see stats::family(). 'binomial' is logistic, and 'gaussian' is linear.

x

focal variables (terms), the column names of fCNA$sample_summary or a list of genes (vector).

covars

covariables in fCNA$sample_summary.

y

predicted variables or expression in string format, e.g., "sex" or "factor(sex)" is acceptable.

merge_circular

if TRUE, merge 'possibly_circular' into 'circular' type.

x_is_gene

if TRUE, treat x as a list of genes.

gene_focus

focal amplication type you focus on. Can be 'fCNA' or 'circular' and 'vs'. If 'fCNA' selected, noncircular and circular genes are included to classify samples. 'vs' for comparing circular and noncircular. (at least 2 samples in each group)

optimize_model

if TRUE, optimize the variable list with stepAIC in 'both' direction.

ending_time

survival analysis ending time. If a numeric ending is typed, all survival data longer than the ending time will be rewritten.

parallel

if TRUE, use N-1 cores to run the task.

exp

logical, indicating whether or not to exponentiate the the coefficients.

ref_line

reference line, default is 1 for HR.

xlim

limits of x axis.

...

other plot options passing to ?forestploter::forest(). Also check https://github.com/adayim/forestploter to see more complex adjustment of the result plot.

Value

a list containing model and forest plot.