Draw Oncoprint for fCNA Profile Visualization

See oncoprint and heatmap to modify the result plot in depth.

gcap.plotProfile(
  fCNA,
  genes = NULL,
  samples = NULL,
  merge_circular = FALSE,
  show_column_names = TRUE,
  remove_empty_columns = FALSE,
  remove_empty_rows = TRUE,
  heatmap_legend_param = list(title = "fCNA"),
  col = c(noncircular = "#0066CC", circular = "#CC0033", possibly_circular = "#FFCCCC"),
  ...
)

Arguments

fCNA

a fCNA object.

genes

a list of genes. Must be subsets of fCNA$data$gene_id.

samples

a vector to subset samples.

merge_circular

if TRUE, merge 'possibly_circular' into 'circular' type.

show_column_names

see ?ComplexHeatmap::oncoPrint.

remove_empty_columns

see ?ComplexHeatmap::oncoPrint.

remove_empty_rows

see ?ComplexHeatmap::oncoPrint.

heatmap_legend_param

see ?ComplexHeatmap::oncoPrint.

col

see ?ComplexHeatmap::oncoPrint.

...

Other parameters passing to ComplexHeatmap::oncoPrint.

Value

a Heatmap object, you can go further modify it.

See also

Examples

# \donttest{
if (require("ComplexHeatmap") && require("IDConverter")) {
  data("ascn")
  data <- ascn

  # Create fake data
  set.seed(1234)
  data$sample <- sample(LETTERS[1:10], nrow(data), replace = TRUE)
  rv <- gcap.ASCNworkflow(data, outdir = tempdir(), model = "XGB11")

  gcap.plotProfile(rv, samples = c("B", "A", "C"))

  rv$convertGeneID()
  ht <- gcap.plotProfile(rv,
    samples = c("B", "A", "C", "D", "F"),
    genes = unique(rv$data$gene_id)[1:10],
    top_annotation = ComplexHeatmap::HeatmapAnnotation(
      cbar = ComplexHeatmap::anno_oncoprint_barplot(),
      foo1 = c("g1", "g1", "g2", "g2", "g3"),
      bar1 = ComplexHeatmap::anno_points(1:5)
    )
  )
  ht
  ComplexHeatmap::draw(ht, merge_legends = TRUE)

  #----------------
  # Plot KM curve
  #----------------
  surv_data <- data.frame(
    sample = rv$sample_summary$sample,
    time = 3000 * abs(rnorm(nrow(rv$sample_summary))),
    status = sample(c(0, 1), nrow(rv$sample_summary), replace = TRUE)
  )
  surv_data
  p <- gcap.plotKMcurve(rv, surv_data)
  p

  p2 <- gcap.plotKMcurve(rv, surv_data, genes = "MYC")
  p2

  # ---------------
  # Plot forest
  # ---------------
  p3 <- gcap.plotForest(rv, surv_data)

  # Fake some data for gene analysis
  rv$data[amplicon_type == "circular"]$gene_id[1:10] <- "MYC"
  rv$data[amplicon_type == "circular"]$gene_id[11:20] <- "EGFR"
  rv$data[amplicon_type == "circular"]$gene_id[31:35] <- "GSX2"
  p4 <- gcap.plotForest(rv, surv_data,
    x = c("TP53", "MYC", "ABC", "EGFR", "GSX2"), x_is_gene = TRUE,
    ref_line = 1, xlim = c(0, 10)
  )

  # -----------------
  # Plot distribution
  # -----------------
  gcap.plotDistribution(rv)
  p5 <- gcap.plotDistribution(rv,
    x = c("MYC", "EGFR", "CDK4", "AKT3"),
    width = 0.5, x_size = 5, fill = FALSE
  )
  p5
  rv$sample_summary[, ploidy_class := ifelse(ploidy > 2.5, "2+", "2")]
  p6 <- gcap.plotDistribution(rv, x = "ploidy_class", by = "sample")
  p6
}
# }