Run ASCAT on tumor-normal pair WES data files

A wrapper calling ASCAT on WES data on one or more tumor-normal pairs. Note, for multiple tumor-normal pairs, the first 5 arguments should be a vector with same length.

gcap.runASCAT(
  tumourseqfile,
  normalseqfile,
  tumourname,
  normalname,
  jobname = tumourname,
  outdir = getwd(),
  allelecounter_exe = "~/miniconda3/envs/cancerit/bin/alleleCounter",
  g1000allelesprefix = file.path("~/data/snp/1000G_loci_hg38",
    "1kg.phase3.v5a_GRCh38nounref_allele_index_chr"),
  g1000lociprefix = file.path("~/data/snp/1000G_loci_hg38",
    "1kg.phase3.v5a_GRCh38nounref_loci_chrstring_chr"),
  GCcontentfile = "~/data/snp/GC_correction_hg38.txt",
  replictimingfile = "~/data/snp/RT_correction_hg38.txt",
  nthreads = 22,
  minCounts = 10,
  BED_file = NA,
  probloci_file = NA,
  chrom_names = 1:22,
  gender = "XX",
  min_base_qual = 20,
  min_map_qual = 35,
  penalty = 70,
  skip_finished_ASCAT = FALSE
)

Arguments

tumourseqfile

Full path to the tumour BAM file.

normalseqfile

Full path to the normal BAM file.

tumourname

Identifier to be used for tumour output files.

normalname

Identifier to be used for normal output files.

jobname

job name, typically an unique name for a tumor-normal pair.

outdir

result output path.

allelecounter_exe

Path to the allele counter executable.

g1000allelesprefix

Prefix path to the 1000 Genomes alleles reference files.

g1000lociprefix

Prefix path to the 1000 Genomes SNP reference files.

GCcontentfile

File containing the GC content around every SNP for increasing window sizes

replictimingfile

File containing replication timing at every SNP for various cell lines (optional)

nthreads

The number of parallel processes for getting allele counts (optional, default=1).

minCounts

Minimum depth required in the normal for a SNP to be considered (optional, default=10).

BED_file

A BED file for only looking at SNPs within specific intervals (optional, default=NA).

probloci_file

A file (chromosome <tab> position; no header) containing specific loci to ignore (optional, default=NA).

chrom_names

A vector containing the names of chromosomes to be considered (optional, default=1:22).

gender

a vector of gender for each cases ("XX" or "XY"). Default = all female ("XX"). Ignore this if you don't include sex chromosomes.

min_base_qual

Minimum base quality required for a read to be counted (optional, default=20).

min_map_qual

Minimum mapping quality required for a read to be counted (optional, default=35).

penalty

penalty of introducing an additional ASPCF breakpoint (expert parameter, don't adapt unless you know what you're doing)

skip_finished_ASCAT

if TRUE, skipped finished ASCAT calls to save time.

Value

Nothing. Check the outdir for results.