Run gene-level circular prediction

Run gene-level circular prediction

gcap.runPrediction(data, model = "XGB32")

Arguments

data

data to predict (data.frame/matrix format), from gcap.collapse2Genes() in general.

model

model name ("XGB11", "XGB32", "XGB56") or a custom model from input. 'toy' can be used for test.

Value

a numeric vector representing prob.

Examples

data("ec")
# Use toy model for illustration
y_pred <- gcap.runPrediction(ec, "toy")
#> <gcap> 2022-05-24 18:14:39 info [gcap.runPrediction]: selecting necessary features from input data
#> <gcap> 2022-05-24 18:14:39 info [gcap.runPrediction]: running prediction
#> [18:14:39] WARNING: amalgamation/../src/learner.cc:1040: 
#>   If you are loading a serialized model (like pickle in Python, RDS in R) generated by
#>   older XGBoost, please export the model by calling `Booster.save_model` from that version
#>   first, then load it back in current version. See:
#> 
#>     https://xgboost.readthedocs.io/en/latest/tutorials/saving_model.html
#> 
#>   for more details about differences between saving model and serializing.
#> 
#> [18:14:39] WARNING: amalgamation/../src/learner.cc:749: Found JSON model saved before XGBoost 1.6, please save the model using current version again. The support for old JSON model will be discontinued in XGBoost 2.3.
#> [18:14:39] WARNING: amalgamation/../src/learner.cc:438: 
#>   If you are loading a serialized model (like pickle in Python, RDS in R) generated by
#>   older XGBoost, please export the model by calling `Booster.save_model` from that version
#>   first, then load it back in current version. See:
#> 
#>     https://xgboost.readthedocs.io/en/latest/tutorials/saving_model.html
#> 
#>   for more details about differences between saving model and serializing.
#> 
y_pred
#>    [1] 9.996558e-01 9.997931e-01 9.998011e-01 9.951738e-01 9.982932e-01
#>    [6] 9.989906e-01 9.998568e-01 9.990216e-01 9.995852e-01 9.986591e-01
#>   [11] 9.995577e-01 9.995648e-01 9.997943e-01 9.996861e-01 9.995279e-01
#>   [16] 9.999089e-01 9.996369e-01 9.982839e-01 9.998569e-01 9.996058e-01
#>   [21] 1.008624e-04 3.668110e-04 1.842791e-05 2.203745e-05 3.549933e-05
#>   [26] 3.373874e-05 2.526317e-04 1.408498e-04 1.004817e-03 3.374122e-05
#>   [31] 2.814008e-05 7.825880e-05 7.023397e-05 1.964915e-04 1.523355e-05
#>   [36] 5.664105e-05 9.694960e-05 6.834344e-05 3.798360e-05 4.306786e-05
#>   [41] 1.174688e-05 5.302919e-04 7.955192e-05 1.593249e-04 1.896029e-05
#>   [46] 1.091219e-04 9.939062e-05 3.464393e-04 2.436079e-03 4.662870e-04
#>   [51] 8.906637e-04 8.093182e-05 2.668810e-05 3.718610e-04 1.042741e-03
#>   [56] 1.278685e-05 7.778391e-05 7.945405e-05 1.470105e-03 1.020623e-02
#>   [61] 5.735652e-05 2.864957e-05 1.411447e-05 5.983174e-04 4.883202e-05
#>   [66] 2.287755e-04 3.873918e-03 6.487765e-05 6.201187e-05 3.570123e-05
#>   [71] 1.334806e-04 1.182496e-03 9.103127e-05 6.428416e-05 5.488676e-05
#>   [76] 8.357849e-05 5.505367e-05 8.455266e-06 7.180345e-02 7.444093e-05
#>   [81] 3.021428e-05 3.273776e-04 3.360522e-05 1.646324e-05 2.172004e-04
#>   [86] 4.169708e-04 4.730370e-04 2.241187e-05 1.228682e-04 9.729239e-04
#>   [91] 3.208002e-05 5.144923e-05 1.203633e-04 6.233105e-04 2.682600e-04
#>   [96] 7.914749e-04 2.421910e-04 3.184480e-05 3.048251e-04 5.011217e-04
#>  [101] 1.457718e-05 4.811778e-05 7.515139e-05 3.702239e-03 7.586185e-05
#>  [106] 1.236432e-02 4.908020e-05 3.910048e-05 8.200319e-05 9.900864e-06
#>  [111] 2.436079e-03 6.094390e-05 1.956471e-05 5.268015e-05 1.767937e-05
#>  [116] 1.621150e-04 1.300570e-04 2.776479e-04 3.157789e-05 2.606016e-05
#>  [121] 1.935648e-05 9.900864e-06 1.846070e-05 2.303637e-05 5.911879e-05
#>  [126] 6.268323e-05 9.198735e-04 3.751297e-05 2.151884e-05 2.039436e-04
#>  [131] 5.792620e-05 4.905120e-05 3.240935e-05 7.694656e-05 7.515139e-05
#>  [136] 2.870451e-05 1.105406e-03 5.624699e-05 1.278792e-04 2.401499e-05
#>  [141] 1.006192e-04 9.926661e-04 3.176223e-04 1.068797e-04 9.564730e-05
#>  [146] 4.516960e-05 1.050177e-04 6.215152e-05 1.682637e-04 4.267090e-05
#>  [151] 2.767549e-04 2.569734e-04 6.959819e-05 1.109273e-05 6.427443e-06
#>  [156] 1.226377e-04 1.816616e-04 2.621806e-05 5.534597e-05 2.737916e-04
#>  [161] 2.172177e-05 6.417422e-05 2.204414e-05 1.724512e-04 9.466540e-06
#>  [166] 3.373874e-05 1.037861e-04 8.215801e-05 7.051650e-04 4.267090e-05
#>  [171] 2.303637e-05 1.031555e-04 4.793009e-04 2.452128e-04 1.503541e-05
#>  [176] 2.655167e-05 3.516987e-04 1.809493e-05 5.735652e-05 5.456508e-05
#>  [181] 1.282476e-05 7.227578e-05 4.556659e-05 2.452128e-04 1.000741e-04
#>  [186] 2.397382e-05 7.358720e-05 1.935152e-05 2.002694e-03 1.075991e-04
#>  [191] 4.877278e-05 1.852458e-05 7.888052e-05 7.358720e-05 4.552055e-05
#>  [196] 6.240595e-04 1.307091e-04 9.901110e-05 7.592597e-05 5.906857e-05
#>  [201] 5.007918e-04 8.168434e-06 3.373874e-05 9.984526e-05 1.229549e-04
#>  [206] 2.204414e-05 2.928787e-05 1.226377e-04 4.100412e-04 4.525117e-05
#>  [211] 8.743510e-05 2.100166e-05 1.476146e-05 1.401372e-04 2.040934e-05
#>  [216] 6.491361e-05 1.389196e-04 4.286741e-05 7.844477e-05 5.890253e-05
#>  [221] 1.091219e-04 1.061116e-04 1.226377e-04 8.019967e-06 4.913648e-04
#>  [226] 7.216517e-05 3.373874e-05 5.011217e-04 4.603920e-05 6.194544e-05
#>  [231] 8.320006e-05 9.821219e-05 9.939355e-05 4.213571e-04 3.653556e-05
#>  [236] 1.182496e-03 6.256546e-05 5.454965e-04 9.052098e-06 4.261819e-04
#>  [241] 2.849694e-04 2.629086e-05 8.617508e-06 2.154306e-04 1.207030e-05
#>  [246] 2.505852e-04 3.373874e-05 1.244604e-04 8.183726e-05 2.668810e-05
#>  [251] 5.467112e-03 2.037865e-04 2.305751e-05 2.199827e-04 1.336096e-04
#>  [256] 1.547216e-04 3.478333e-04 4.964584e-05 9.103127e-05 2.310114e-04
#>  [261] 2.605650e-04 2.991142e-05 2.697349e-04 2.123363e-05 3.208002e-05
#>  [266] 2.343759e-05 3.531719e-03 4.079356e-04 2.146076e-05 1.407112e-04
#>  [271] 9.513811e-04 1.442985e-04 2.146076e-05 1.073590e-03 3.012147e-05
#>  [276] 9.354052e-05 8.785382e-05 4.211424e-02 1.186338e-04 1.387668e-04
#>  [281] 7.515139e-05 2.671937e-05 1.993488e-04 2.029018e-05 7.099401e-05
#>  [286] 5.069188e-04 6.577076e-04 4.627336e-05 1.935152e-05 1.389196e-04
#>  [291] 5.126406e-04 1.838163e-04 4.732080e-05 2.363041e-03 7.927276e-03
#>  [296] 1.169647e-04 1.354596e-05 1.944047e-04 2.031692e-04 9.458434e-04
#>  [301] 1.476139e-04 1.320094e-04 7.714771e-04 1.067624e-04 3.273776e-04
#>  [306] 4.369023e-05 8.704259e-04 5.430451e-04 3.879023e-04 8.357849e-05
#>  [311] 9.764169e-04 6.466253e-04 1.025393e-05 3.563198e-05 2.177839e-05
#>  [316] 4.883112e-04 1.698105e-04 9.691040e-05 4.883202e-05 3.299894e-05
#>  [321] 3.004001e-05 3.309948e-05 3.265141e-03 1.908972e-05 4.352196e-04
#>  [326] 4.240523e-05 5.155581e-03 6.190020e-05 6.371557e-06 2.173982e-05
#>  [331] 4.957691e-04 2.363290e-05 7.955192e-05 2.984005e-04 3.974078e-05
#>  [336] 1.697159e-03 2.173785e-05 6.844336e-05 1.613241e-05 8.512859e-05
#>  [341] 4.783205e-04 2.255793e-04 8.638787e-05 4.881712e-05 5.461177e-05
#>  [346] 1.269328e-04 6.867091e-04 1.935152e-05 1.780739e-04 1.808209e-04
#>  [351] 5.810357e-05 1.895997e-04 3.859851e-05 3.740628e-04 8.657409e-04
#>  [356] 1.423094e-03 3.516987e-04 1.408498e-04 5.098388e-02 1.401372e-04
#>  [361] 2.032293e-05 5.087260e-05 1.222519e-04 3.184480e-05 1.899995e-04
#>  [366] 1.215799e-04 5.636023e-05 1.547216e-04 7.552476e-05 8.256469e-05
#>  [371] 1.000639e-04 2.323199e-05 1.203633e-04 1.145947e-04 1.240456e-05
#>  [376] 6.818466e-05 5.811931e-05 4.473368e-05 8.617508e-06 2.567203e-05
#>  [381] 3.861505e-05 2.556956e-05 6.051640e-04 2.028231e-05 7.309790e-05
#>  [386] 7.810107e-06 2.316056e-05 1.772014e-04 6.834344e-05 9.758797e-06
#>  [391] 3.345599e-05 1.286828e-04 1.849033e-05 1.908972e-05 1.058072e-05
#>  [396] 5.144923e-05 8.127211e-05 2.508654e-05 7.005769e-04 1.574118e-05
#>  [401] 2.190819e-03 1.895997e-04 1.643685e-05 6.544217e-05 2.154306e-04
#>  [406] 2.242264e-04 2.605650e-04 1.605766e-05 2.363290e-05 1.191737e-04
#>  [411] 5.894287e-04 1.131648e-04 1.245738e-04 8.626094e-04 1.748803e-04
#>  [416] 1.791042e-03 1.017706e-03 5.394263e-05 7.707829e-04 1.439868e-05
#>  [421] 6.137632e-04 6.086178e-05 9.508058e-05 4.046112e-05 1.591376e-01
#>  [426] 5.664105e-05 5.480428e-05 2.363290e-05 6.373237e-05 3.616645e-04
#>  [431] 5.871403e-05 6.292261e-05 1.842791e-05 7.282504e-06 1.945965e-05
#>  [436] 7.750915e-04 2.725841e-05 2.931791e-04 3.646946e-05 1.040662e-04
#>  [441] 6.417422e-05 1.212841e-04 1.044518e-05 3.739703e-05 4.649962e-05
#>  [446] 5.157725e-05 3.514456e-04 5.302780e-06 1.105311e-04 4.899944e-04
#>  [451] 9.614522e-05 1.313113e-05 1.174688e-05 1.096098e-04 2.498486e-05
#>  [456] 3.184480e-05 2.253250e-04 2.655402e-04 3.238481e-05 7.579207e-05
#>  [461] 1.475281e-05 1.142490e-04 4.945518e-05 4.044649e-04 1.676880e-05
#>  [466] 7.579207e-05 1.178107e-04 5.906857e-05 1.904812e-04 1.407112e-04
#>  [471] 1.048306e-02 2.901680e-04 1.362381e-04 4.267090e-05 2.311031e-04
#>  [476] 1.720204e-05 9.900864e-06 1.058072e-05 3.158755e-05 2.939538e-05
#>  [481] 5.461177e-05 1.799000e-04 7.694656e-05 1.824652e-05 1.621150e-04
#>  [486] 1.976140e-05 1.311678e-02 1.832072e-04 1.621665e-04 2.838358e-05
#>  [491] 2.629086e-05 3.158710e-05 5.313855e-04 9.649636e-05 8.612973e-06
#>  [496] 8.446713e-05 1.846070e-05 8.612973e-06 1.753850e-04 6.660852e-05
#>  [501] 2.404538e-05 4.393648e-05 2.576028e-04 6.316402e-03 1.681441e-04
#>  [506] 6.005336e-04 4.910034e-05 1.502498e-05 6.233517e-05 8.334037e-05
#>  [511] 1.696816e-05 2.873335e-03 1.677786e-03 2.912445e-04 6.579552e-05
#>  [516] 9.694960e-05 2.854888e-04 8.318768e-06 2.413609e-04 3.189610e-05
#>  [521] 2.925904e-04 5.735652e-05 6.094390e-05 1.382234e-04 4.286741e-05
#>  [526] 6.411250e-05 1.094924e-02 1.715868e-03 1.086169e-05 1.776299e-04
#>  [531] 1.203633e-04 6.682322e-04 4.831716e-04 2.172177e-05 1.157277e-04
#>  [536] 1.762236e-04 1.044518e-05 1.173079e-04 2.852641e-04 1.076598e-04
#>  [541] 3.417571e-05 9.939062e-05 3.410495e-04 1.962606e-05 4.825177e-05
#>  [546] 2.777706e-02 5.465121e-04 4.267090e-05 1.275025e-04 7.282504e-06
#>  [551] 1.154375e-04 2.307431e-04 2.569753e-04 8.256469e-05 1.506147e-05
#>  [556] 3.374122e-05 2.196786e-05 3.261954e-03 8.168434e-06 4.556659e-05
#>  [561] 3.342145e-05 2.608290e-04 5.800160e-04 1.966295e-04 4.221463e-05
#>  [566] 4.527897e-05 5.107672e-05 3.112393e-05 2.099095e-04 5.146311e-05
#>  [571] 1.040294e-04 2.725841e-05 6.660852e-05 2.725841e-05 7.579207e-05
#>  [576] 9.061107e-05 4.649962e-05 3.035482e-04 8.256469e-05 3.432803e-04
#>  [581] 3.462247e-04 1.652665e-05 8.168756e-05 4.944631e-05 1.566068e-04
#>  [586] 1.244604e-04 1.564268e-05 5.796054e-03 5.945388e-05 2.388585e-05
#>  [591] 2.463232e-05 3.020412e-05 1.698544e-04 6.945644e-05 4.556659e-05
#>  [596] 1.074894e-04 1.895997e-04 1.864377e-02 7.047783e-04 2.854888e-04
#>  [601] 8.200319e-05 1.243266e-04 6.288128e-04 1.849033e-05 5.919822e-04
#>  [606] 8.168434e-06 2.370730e-05 2.751204e-05 2.547692e-05 5.669261e-05
#>  [611] 7.955192e-05 8.638787e-05 5.707939e-05 1.050177e-04 7.881150e-05
#>  [616] 6.529063e-05 9.252372e-04 5.534597e-05 8.657391e-05 8.743510e-05
#>  [621] 9.726054e-05 1.864046e-05 2.020803e-05 2.740993e-05 1.045351e-04
#>  [626] 1.776299e-04 4.980367e-03 1.243266e-04 8.283724e-05 1.356078e-05
#>  [631] 1.508707e-03 1.261697e-04 4.883202e-05 9.958766e-05 9.855049e-05
#>  [636] 1.332688e-03 9.194995e-05 3.448693e-05 2.531932e-05 8.961021e-03
#>  [641] 1.608515e-05 6.639113e-05 6.436452e-05 6.715216e-05 1.340203e-04
#>  [646] 7.420152e-05 1.300570e-04 3.155557e-04 6.094390e-05 2.725841e-05
#>  [651] 1.557180e-04 1.689856e-04 2.232152e-05 1.017706e-03 5.906857e-05
#>  [656] 8.256469e-05 5.302780e-06 2.785462e-04 1.364532e-04 2.943077e-05
#>  [661] 7.023397e-05 2.585315e-05 2.668810e-05 1.066842e-05 8.638787e-05
#>  [666] 5.707939e-05 5.832134e-05 1.951980e-05 1.284741e-04 1.283053e-04
#>  [671] 1.203655e-04 9.901110e-05 2.450064e-04 1.321193e-05 7.719172e-04
#>  [676] 6.547700e-05 2.747877e-05 6.487765e-05 6.938270e-05 2.539541e-04
#>  [681] 2.154306e-04 2.740899e-04 6.376903e-05 1.552548e-04 2.727370e-05
#>  [686] 1.838163e-04 1.983117e-05 9.644678e-05 5.986297e-05 2.972922e-05
#>  [691] 5.404102e-05 8.638787e-05 3.598663e-04 2.724828e-04 9.200846e-05
#>  [696] 1.050177e-04 3.174204e-04 9.061107e-05 4.267090e-05 1.791042e-03
#>  [701] 4.649962e-05 5.258112e-05 1.613589e-05 1.608515e-05 1.507128e-04
#>  [706] 1.328823e-05 2.584950e-05 1.613241e-05 1.334806e-04 4.468550e-05
#>  [711] 9.061107e-05 2.241843e-04 9.466540e-06 4.218936e-05 1.139278e-04
#>  [716] 3.411982e-05 5.146311e-05 1.252214e-03 3.682225e-05 3.056921e-04
#>  [721] 3.824069e-04 1.899288e-04 1.487578e-04 6.277051e-05 3.540845e-05
#>  [726] 1.621100e-05 7.852305e-05 1.454075e-04 1.182496e-03 6.287157e-05
#>  [731] 3.004001e-05 8.532144e-05 1.197763e-04 7.341938e-06 3.898851e-05
#>  [736] 2.133476e-05 3.061686e-04 1.910802e-05 5.970972e-05 2.390323e-04
#>  [741] 6.436452e-05 4.232807e-05 3.100448e-04 2.463413e-05 1.066842e-05
#>  [746] 5.532920e-05 7.833698e-05 2.464882e-04 1.910802e-05 2.215624e-05
#>  [751] 8.282732e-04 1.812341e-04 8.559455e-03 4.037598e-05 6.986760e-04
#>  [756] 5.302780e-06 1.179514e-04 3.448693e-05 4.800774e-05 6.037038e-05
#>  [761] 9.174212e-06 1.941807e-03 1.476146e-05 1.537356e-04 1.384615e-03
#>  [766] 3.925321e-05 9.926200e-05 1.954608e-04 1.571908e-05 9.726054e-05
#>  [771] 4.516960e-05 5.906857e-05 1.304904e-04 2.599501e-05 9.174212e-06
#>  [776] 1.173079e-04 7.455908e-04 5.506166e-04 1.376113e-04 7.066999e-05
#>  [781] 1.859052e-04 1.290871e-03 9.778475e-05 7.023397e-05 4.109513e-05
#>  [786] 2.521587e-03 1.935648e-05 3.373874e-05 4.267090e-05 7.461071e-05
#>  [791] 3.682225e-05 9.564730e-05 1.058234e-04 9.694960e-05 4.599638e-04
#>  [796] 1.109176e-04 4.495504e-05 7.244773e-05 1.128081e-05 3.395566e-04
#>  [801] 8.357849e-05 1.823522e-03 2.187592e-03 4.800774e-05 1.358806e-04
#>  [806] 9.882328e-06 2.632863e-05 1.386044e-05 3.498546e-05 2.473786e-05
#>  [811] 1.076685e-04 4.841933e-05 3.709490e-05 1.846070e-05 5.247453e-05
#>  [816] 2.445597e-04 1.887295e-05 3.416896e-04 3.506065e-04 3.004001e-05
#>  [821] 1.397753e-05 8.617508e-06 2.027198e-04 1.336096e-04 4.652273e-05
#>  [826] 7.955192e-05 9.112777e-05 8.565714e-05 7.282504e-06 7.023397e-05
#>  [831] 3.998247e-05 1.380745e-03 9.866661e-05 2.572403e-04 9.027342e-05
#>  [836] 5.550300e-05 8.709940e-05 4.380429e-04 3.273776e-04 3.172241e-04
#>  [841] 7.020919e-05 1.025742e-04 3.190206e-05 8.854289e-06 1.753850e-04
#>  [846] 8.156474e-05 6.639431e-04 4.800774e-05 5.113686e-05 1.929361e-04
#>  [851] 2.343864e-05 3.008955e-05 1.356078e-05 1.041109e-04 5.302780e-06
#>  [856] 1.497185e-03 4.647528e-05 7.854080e-05 5.921927e-05 1.185066e-04
#>  [861] 4.627336e-05 3.054692e-04 8.239920e-04 2.629086e-05 4.713794e-05
#>  [866] 2.463232e-05 1.125488e-03 2.585315e-05 5.360714e-04 4.129005e-05
#>  [871] 2.763450e-05 2.529074e-04 3.652398e-03 1.335749e-05 8.217166e-04
#>  [876] 4.666476e-04 6.644853e-04 3.028375e-05 1.011382e-04 7.630836e-04
#>  [881] 4.905120e-05 1.121519e-04 7.800228e-05 2.671127e-05 1.849386e-04
#>  [886] 1.842791e-05 7.591012e-05 7.819883e-05 4.819723e-05 2.725841e-05
#>  [891] 7.942700e-05 1.843978e-05 1.000741e-04 2.100166e-05 4.712477e-05
#>  [896] 2.228111e-02 1.076598e-04 1.511549e-03 2.582700e-04 3.179312e-05
#>  [901] 1.091219e-04 2.363290e-05 1.325660e-04 6.455311e-04 4.649962e-05
#>  [906] 6.094390e-05 1.967097e-03 2.622216e-05 7.822620e-05 5.712366e-05
#>  [911] 8.446713e-05 2.098470e-05 3.108504e-04 1.908972e-05 5.254353e-04
#>  [916] 3.222634e-03 6.358692e-05 2.029018e-05 7.699632e-05 1.608515e-05
#>  [921] 3.708914e-05 1.959551e-04 1.000639e-04 6.243851e-05 2.984005e-04
#>  [926] 7.504119e-05 1.040294e-04 3.004001e-05 1.476146e-05 1.222519e-04
#>  [931] 8.498048e-05 1.937751e-03 4.964584e-05 7.822620e-05 3.708224e-05
#>  [936] 7.151422e-06 2.797004e-04 7.955192e-05 3.108504e-04 1.086169e-05
#>  [941] 4.541402e-05 8.093182e-05 6.003259e-05 2.161790e-05 3.785993e-04
#>  [946] 6.417422e-05 1.404056e-01 7.413645e-03 3.091761e-04 3.175591e-05
#>  [951] 3.373874e-05 5.984802e-05 8.512859e-05 5.681218e-03 3.465180e-05
#>  [956] 3.628460e-04 3.158755e-05 4.939913e-02 1.646324e-05 5.962978e-05
#>  [961] 1.528551e-05 8.327102e-06 3.215579e-05 9.694960e-05 8.743510e-05
#>  [966] 3.620370e-06 1.762236e-04 1.191737e-04 1.269794e-04 1.280921e-04
#>  [971] 1.784529e-05 2.963148e-04 1.233926e-03 9.172261e-06 1.730116e-05
#>  [976] 1.300269e-04 3.583688e-03 7.957210e-05 3.057454e-05 4.881712e-05
#>  [981] 4.762225e-04 3.616645e-04 5.430451e-04 9.152210e-05 6.886399e-05
#>  [986] 2.938205e-03 7.050958e-05 6.945644e-05 8.704259e-04 1.025742e-04
#>  [991] 2.270994e-04 6.844336e-05 8.736735e-04 2.159047e-03 2.705557e-05
#>  [996] 8.607466e-05 1.016822e-03 1.547216e-04 4.652273e-05 1.729477e-01
#> [1001] 4.026762e-05 2.197148e-04 5.832134e-05 1.089310e-04 1.542305e-04
#> [1006] 7.309790e-05 1.846070e-05 1.698105e-04 7.464857e-05 8.357849e-05
#> [1011] 4.468550e-05 1.859052e-04 8.246476e-05 4.999637e-05 2.075605e-05
#> [1016] 1.650910e-05 1.067624e-04 1.659894e-04 3.269971e-05 2.751204e-05
#> [1021] 1.846070e-05 2.204414e-05 3.690651e-05 1.704412e-03 3.368187e-05
#> [1026] 9.140382e-05 3.303947e-05 7.033394e-04 8.073441e-05 2.537978e-05
#> [1031] 5.112384e-04 1.669661e-05 4.753972e-05 5.643601e-05 1.040631e-04
#> [1036] 4.786164e-05 9.743906e-05 4.724893e-05 2.216296e-05 7.159556e-05
#> [1041] 4.732080e-05 1.831155e-04 1.221933e-04 7.033394e-04 4.999637e-05
#> [1046] 9.497056e-04 1.745618e-05 4.640540e-05 4.286483e-04 6.094390e-05
#> [1051] 3.285078e-04 9.939062e-05 2.436863e-05 1.040631e-04 6.544217e-05
#> [1056] 9.151754e-06 9.958766e-05 5.152072e-04 2.871254e-05 2.943077e-05
#> [1061] 1.747312e-03 9.743906e-05 8.200319e-05 8.446713e-05 2.343864e-05
#> [1066] 3.299894e-05 2.127346e-05 9.184330e-05 2.519354e-05 4.944631e-05
#> [1071] 9.898401e-05 1.000639e-04 1.552465e-04 5.152072e-04 5.940127e-04
#> [1076] 3.373874e-05 1.357176e-05 9.152210e-05 5.504842e-05 1.374857e-05
#> [1081] 6.055293e-05 1.217184e-04 4.908020e-05 5.214162e-05 1.280921e-04
#> [1086] 4.883202e-05 2.310114e-04 3.269971e-05 1.921092e-02 5.129237e-05
#> [1091] 5.443961e-04 1.057296e-03 1.008624e-04 1.512683e-05 2.548041e-03
#> [1096] 4.947180e-04 5.144923e-05 8.020236e-04 4.109513e-05 1.421509e-04
#> [1101] 1.523051e-05 4.999637e-05 4.883202e-05 4.165775e-05 4.604709e-04
#> [1106] 2.984005e-04 5.683969e-05 1.925522e-03 4.729692e-04 3.557840e-05
#> [1111] 1.490955e-03 2.436079e-03 2.343864e-05 2.027322e-05 6.233517e-05
#> [1116] 7.226385e-05 1.000741e-04 1.050177e-04 2.549569e-02 3.434188e-04
#> [1121] 5.146311e-05 1.485900e-04 6.547700e-05 5.841881e-05 3.708991e-04
#> [1126] 8.298628e-05 1.460931e-04 8.566885e-06 8.298628e-05 1.682637e-04
#> [1131] 8.498048e-05 4.877278e-05 1.903333e-04 3.498546e-05 3.004001e-05
#> [1136] 1.548480e-05 5.616180e-06 7.712242e-05 2.715127e-04 6.917993e-05
#> [1141] 2.558076e-04 3.449209e-05 1.008624e-04 1.336037e-04 1.091219e-04
#> [1146] 6.159670e-03 3.176223e-04 7.552476e-05 3.449209e-05 8.298628e-05
#> [1151] 1.507059e-04 4.954987e-04 9.103127e-05 2.154306e-04 2.645325e-04
#> [1156] 6.094390e-05 2.751204e-05 2.599501e-05 5.323508e-04 1.411447e-05
#> [1161] 4.883202e-05 1.959992e-04 7.865503e-05 1.257287e-04 2.854888e-04
#> [1166] 5.730370e-05 5.841881e-05 4.945518e-05 5.027794e-05 1.121324e-03
#> [1171] 1.652665e-05 1.173079e-04 6.511567e-05 3.157789e-05 3.465180e-05
#> [1176] 4.791393e-05 1.224668e-04 8.401174e-05 7.866769e-04 2.088906e-05
#> [1181] 3.562964e-05 2.725841e-05 7.159556e-05 2.039436e-04 2.547692e-05
#> [1186] 1.064848e-04 4.811778e-05 3.360186e-05 8.446713e-05 1.194994e-04
#> [1191] 1.000639e-04 7.442448e-04 4.653514e-04 2.363290e-05 4.927605e-05
#> [1196] 1.682637e-04 1.608515e-05 7.515139e-05 2.215624e-05 3.546231e-04
#> [1201] 6.078347e-05 1.025742e-04 1.669711e-05 2.885784e-04 1.653503e-04
#> [1206] 8.701315e-04 5.136819e-05 4.590443e-05 2.169075e-03 1.057036e-04
#> [1211] 1.374718e-04 2.063249e-04 5.911879e-05 5.792620e-05 1.037577e-04
#> [1216] 1.733786e-05 7.105893e-06 6.871574e-05 5.474912e-05 1.107477e-04
#> [1221] 2.546888e-05 4.604709e-04 1.548480e-05 6.716022e-05 9.103127e-05
#> [1226] 3.089770e-05 1.104846e-03 2.830798e-04 4.861532e-04 1.964915e-04
#> [1231] 1.880544e-05 5.557111e-05 2.098749e-04 1.874943e-04 1.213969e-04
#> [1236] 1.180172e-03 3.528444e-04 2.105433e-04 4.218936e-05 4.362109e-04
#> [1241] 2.146076e-05 2.939538e-05 2.514220e-05 1.646324e-05 1.178197e-03
#> [1246] 6.871574e-05 3.229188e-04 3.063382e-05 3.536223e-04 1.811982e-04
#> [1251] 4.880759e-05 7.515139e-05 2.232152e-05 1.222803e-04 1.212800e-04
#> [1256] 1.186886e-04 1.017706e-03 1.876892e-05 7.854882e-05 6.229916e-05
#> [1261] 6.945644e-05 3.885730e-05 4.426849e-05 5.646712e-05 1.046629e-04
#> [1266] 5.906857e-05 4.267090e-05 2.123363e-05 3.908419e-05 1.648352e-04
#> [1271] 5.775352e-04 1.665811e-05 2.751204e-05 3.004001e-05 7.099401e-05
#> [1276] 1.859052e-04 2.928787e-05 2.984005e-04 2.027322e-05 2.875334e-05
#> [1281] 1.908972e-05 1.659544e-03 2.939538e-05 7.541127e-05 7.515139e-05
#> [1286] 1.184680e-04 6.313069e-05 3.346578e-05 2.370881e-04 4.168164e-05
#> [1291] 8.617508e-06 1.215194e-04 2.315584e-03 1.257287e-04 2.232152e-05
#> [1296] 1.825710e-04 2.025621e-03 4.292349e-05 3.690651e-05 2.621806e-05
#> [1301] 2.225474e-04 5.911879e-05 1.213969e-04 1.269328e-04 4.732080e-05
#> [1306] 1.207481e-04 4.417615e-03 3.606046e-04 3.682225e-05 1.174688e-05
#> [1311] 1.008624e-04 1.187728e-03 3.642418e-04 2.457310e-04 1.449588e-04
#> [1316] 1.612896e-05 1.613241e-05 1.419296e-03 8.838295e-05 1.334806e-04
#> [1321] 1.408498e-04 2.950317e-05 1.084071e-03 1.859052e-04 4.800774e-05
#> [1326] 3.739703e-05 3.283260e-05 1.340689e-04 5.857232e-05 2.162152e-04
#> [1331] 5.783552e-04 5.707939e-05 3.849601e-04 1.142490e-04 3.859851e-05
#> [1336] 1.454075e-04 1.487578e-04 7.066999e-05 1.025742e-04 1.772014e-04
#> [1341] 1.547216e-04 2.629086e-05 1.899590e-05 4.486899e-05 3.391393e-04
#> [1346] 7.420152e-05 2.032293e-05 3.280950e-05 9.984526e-05 1.429642e-04
#> [1351] 5.302780e-06 5.504842e-05 3.218899e-05 1.018374e-03 1.846070e-05
#> [1356] 8.743510e-05 2.232152e-05 6.963503e-05 5.480428e-05 1.351874e-05
#> [1361] 1.149189e-05 4.787091e-05 1.008308e-05 5.287127e-04 1.613589e-05
#> [1366] 4.232807e-05 4.333966e-05 5.803834e-04 2.039436e-04 3.374122e-05
#> [1371] 5.410878e-05 2.463232e-05 1.166114e-04 1.910216e-05 1.356078e-05
#> [1376] 1.698544e-04 5.550300e-05 6.660852e-05 4.997855e-04 1.743529e-04
#> [1381] 1.425940e-05 6.607042e-05 7.700542e-05 6.229916e-05 1.244604e-04
#> [1386] 6.834344e-05 2.376158e-03 1.196016e-04 2.742556e-04 7.099401e-05
#> [1391] 9.564730e-05 1.031555e-04 7.629482e-05 6.086178e-05 6.417422e-05
#> [1396] 9.135962e-03 5.161060e-04 9.882328e-06 2.482097e-04 1.949660e-05
#> [1401] 7.173778e-05 4.597641e-04 3.772459e-04 1.173079e-04 2.368856e-05
#> [1406] 6.233517e-05 2.144020e-05 7.917269e-05 2.463232e-05 5.295602e-05
#> [1411] 2.203745e-05 3.373874e-05 2.255793e-04 1.971833e-05 3.391645e-05
#> [1416] 1.282476e-05 9.678581e-06 2.927475e-05 2.370730e-05 2.963148e-04
#> [1421] 1.125371e-05 6.938270e-05 3.626629e-05 5.701279e-02 7.955192e-05
#> [1426] 5.911879e-05 8.589755e-05 1.058234e-04 7.591012e-05 5.906857e-05
#> [1431] 7.023397e-05 1.935152e-05 9.644678e-05 3.099245e-05 1.008624e-04
#> [1436] 6.094390e-05 2.901994e-05 3.287840e-02 1.408498e-04 4.095984e-05
#> [1441] 3.158755e-05 4.084777e-05 2.436079e-03 2.608794e-03 5.461177e-05
#> [1446] 9.694960e-05 9.283309e-05 2.173982e-05 1.137862e-04 2.232152e-05
#> [1451] 6.683420e-03 1.261891e-04 7.815335e-05 7.023397e-05 1.810185e-03
#> [1456] 8.200319e-05 7.272339e-04 5.911879e-05 4.649962e-05 5.504842e-05
#> [1461] 7.699632e-05 7.631407e-04 3.121935e-05 4.232807e-05 9.194995e-05
#> [1466] 6.715216e-05 1.706071e-04 2.615072e-03 1.191737e-04 1.946487e-03
#> [1471] 8.334037e-05 1.016751e-04 4.266806e-03 5.734919e-05 5.595101e-04
#> [1476] 1.025742e-04 6.935067e-06 2.632863e-05 5.292236e-04 1.023675e-03
#> [1481] 3.031184e-04 2.556956e-05 2.363290e-05 5.609430e-04 3.849020e-04
#> [1486] 1.025742e-04 2.234882e-04 1.168657e-04 4.988330e-04 1.638589e-05
#> [1491] 9.200846e-05 2.204414e-05 1.838163e-04 5.504842e-05 3.643386e-05
#> [1496] 1.121324e-03 8.377038e-05 5.474912e-05 1.168450e-03 5.443961e-04
#> [1501] 6.875297e-05 5.052147e-05 4.877278e-05 7.042846e-03 5.456508e-05
#> [1506] 3.772459e-04 3.373874e-05 2.215624e-05 6.164149e-04 1.174688e-05
#> [1511] 4.081965e-05 1.334806e-04 5.488676e-05 4.883202e-05 1.776299e-04
#> [1516] 9.151754e-06 2.850753e-05 5.302780e-06 7.541127e-05 5.664105e-05
#> [1521] 3.240935e-05 1.356078e-05 6.511567e-05 2.225474e-04 1.401372e-04
#> [1526] 2.076727e-04 5.707939e-05 5.664105e-05 6.068363e-05 4.340927e-05
#> [1531] 4.564510e-04 1.006192e-04 7.955192e-05 9.564730e-05 3.053242e-04
#> [1536] 7.694656e-05 1.956471e-05 1.698298e-04 1.729415e-04 4.106877e-05
#> [1541] 9.726054e-05 1.071876e-04 9.984526e-05 4.732080e-05 3.501367e-04
#> [1546] 2.683024e-05 1.449642e-04 5.292219e-05 1.901383e-04 1.244604e-04
#> [1551] 1.222133e-04 1.408498e-04 1.705931e-03 3.840769e-04 7.667387e-04
#> [1556] 1.582192e-03 8.785382e-05 8.357849e-05 2.838358e-05 5.525499e-06
#> [1561] 4.945518e-05 2.901085e-03 9.644678e-05 2.233220e-04 2.798185e-04
#> [1566] 7.034685e-05 1.174688e-05 4.240523e-05 7.634606e-05 2.870451e-05
#> [1571] 1.711564e-03 1.614834e-04 1.806441e-04 6.373237e-05 4.232807e-05
#> [1576] 7.175168e-04 3.552849e-05 8.334037e-05 4.928959e-05 1.443413e-04
#> [1581] 9.726054e-05 2.640494e-05 1.079362e-03 1.177320e-03 3.397515e-05
#> [1586] 1.288627e-04 8.628423e-03 4.026762e-05 1.362393e-03 5.664105e-05
#> [1591] 8.168434e-06 3.715140e-05 3.449528e-05 4.717953e-04 9.939062e-05
#> [1596] 7.957210e-05 9.799207e-05 1.812401e-05 2.323199e-05 8.498048e-05
#> [1601] 2.303566e-05 2.622216e-05 7.054133e-05 2.310114e-04 1.804026e-04
#> [1606] 5.857232e-05 1.582520e-04 5.841881e-05 3.918940e-04 4.193773e-05
#> [1611] 5.302780e-06 7.207903e-04 9.984526e-05 2.398620e-04 2.463413e-05
#> [1616] 4.876901e-05 2.629086e-05 2.172177e-05 9.443596e-05 1.000639e-04
#> [1621] 1.547216e-04 2.640494e-05 1.354113e-03 4.944631e-05 1.191737e-04
#> [1626] 3.840769e-04 3.373874e-05 8.200319e-05 1.523051e-05 5.488676e-05
#> [1631] 2.442638e-05 6.607042e-05 3.708224e-05 8.589755e-05 8.233706e-05
#> [1636] 1.336037e-04 3.581581e-05 3.589419e-03 8.638787e-05 5.861646e-05
#> [1641] 1.866812e-04 1.280921e-04 2.436079e-03 1.366932e-05 5.828301e-06
#> [1646] 2.735320e-04 4.645876e-04 2.123363e-05 5.353339e-04 4.905120e-05
#> [1651] 4.503967e-04 1.034493e-04 1.269328e-04 4.473368e-05 3.800217e-04
#> [1656] 2.128995e-04 4.267090e-05 2.200151e-05 1.551208e-04 2.509949e-05
#> [1661] 1.476011e-05 1.955675e-04 4.826314e-05 4.889857e-05 6.834344e-05
#> [1666] 5.157725e-05 1.243266e-04 2.697349e-04 2.747877e-05 3.479226e-05
#> [1671] 2.273538e-04 9.859363e-05 9.345519e-05 1.790257e-03 6.041203e-05
#> [1676] 5.504842e-05 1.015228e-04 9.154977e-05 3.299894e-05 4.066566e-04
#> [1681] 1.523051e-05 1.191737e-04 9.199003e-04 1.176404e-04 2.432725e-05
#> [1686] 6.078347e-05 5.203308e-05 9.040892e-05 1.004433e-03 2.701910e-04
#> [1691] 3.945109e-05 2.657379e-05 9.876113e-05 1.203633e-04 9.958766e-05
#> [1696] 1.730116e-05 3.562964e-05 9.726054e-05 5.921927e-05 9.061107e-05
#> [1701] 3.812874e-04 1.064848e-04 5.362105e-05 1.373749e-04 4.086732e-04
#> [1706] 1.730014e-04 6.358528e-05 4.516960e-05 1.536274e-05 2.303566e-05
#> [1711] 3.981844e-05 2.161790e-05 9.742753e-05 4.883202e-05 2.098749e-04
#> [1716] 9.958766e-05 4.859631e-05 4.046112e-05 1.017706e-03 5.706985e-04
#> [1721] 3.125762e-04 9.151754e-06 5.438476e-05 4.666476e-04 2.526317e-04
#> [1726] 5.632498e-05 5.911879e-05 9.347613e-04 5.947418e-04 2.463232e-05
#> [1731] 3.484934e-05 1.705855e-03 7.591012e-05 2.814008e-05 3.682126e-05
#> [1736] 1.151548e-03 2.323199e-05 8.156474e-05 2.203745e-05 2.249243e-04
#> [1741] 5.152072e-04 6.213005e-04 2.316056e-05 3.409452e-05 1.291102e-05
#> [1746] 3.708224e-05 9.490974e-05 6.567091e-05 8.200319e-05 4.905120e-05
#> [1751] 1.203633e-04 3.004356e-04 1.998526e-04 1.592353e-04 8.893463e-04
#> [1756] 5.911641e-05 1.476146e-05 4.232807e-05 6.784303e-06 8.327102e-06
#> [1761] 7.888052e-05 2.821414e-04 4.427330e-04 1.523051e-05 5.152072e-04
#> [1766] 9.564730e-05 8.377038e-05 1.720204e-05 5.277484e-06 1.219570e-04
#> [1771] 1.108584e-04 4.432864e-05 2.731732e-04 5.302780e-06 9.788899e-04
#> [1776] 5.446433e-05 2.767788e-05 1.791042e-03 4.803847e-04 1.842791e-05
#> [1781] 6.352443e-05 9.898401e-05 4.886715e-05 1.025393e-05 8.626094e-04
#> [1786] 6.538023e-05 1.236950e-04 1.600931e-03 1.791042e-03 7.888052e-05
#> [1791] 7.722894e-06 1.222519e-04 1.291102e-05 1.356078e-05 3.158755e-05
#> [1796] 6.937601e-04 4.232807e-05 1.425940e-05 2.573767e-04 2.875139e-05
#> [1801] 8.168434e-06 1.017706e-03 9.289773e-05 2.849924e-05 7.516447e-04
#> [1806] 2.928723e-04 8.781186e-05 3.918940e-04 2.144652e-05 2.207938e-04
#> [1811] 8.048512e-05 4.473368e-05 3.409452e-05 1.652665e-05 3.552567e-04
#> [1816] 1.574118e-05 6.086178e-05 4.669435e-04 1.954757e-04 5.192938e-05
#> [1821] 8.298628e-05 1.574118e-05 1.790640e-04 6.358692e-05 2.708774e-05
#> [1826] 2.682423e-04 3.453053e-05 5.707939e-05 4.790846e-04 5.254353e-04
#> [1831] 1.293106e-05 2.782086e-04 1.556827e-05 1.495690e-04 9.901110e-05
#> [1836] 2.371410e-05 5.254559e-05 1.137037e-05 1.707203e-04 3.606046e-04
#> [1841] 2.203745e-05 1.762236e-04 1.022724e-04 2.725841e-05 1.048348e-04
#> [1846] 3.449209e-05 1.443413e-04 1.236925e-04 4.047057e-03 1.413620e-04
#> [1851] 4.414639e-04 1.240456e-05 4.354319e-05 6.963503e-05 5.007028e-05
#> [1856] 8.357849e-05 7.009333e-05 5.166624e-04 1.006192e-04 2.215624e-05
#> [1861] 8.256469e-05 7.885487e-05 2.106285e-04 5.857232e-05 1.935648e-05
#> [1866] 5.144923e-05 1.791042e-03 3.968769e-05 2.144020e-05 5.585306e-05
#> [1871] 6.896562e-04 1.646324e-05 5.994352e-05 2.854888e-04 8.668158e-06
#> [1876] 4.732080e-05 2.020901e-03 6.428416e-05 1.850805e-04 8.919839e-05
#> [1881] 2.370881e-04 1.107477e-04 1.465199e-05 2.519892e-04 2.032293e-05
#> [1886] 3.693211e-05 5.430451e-04 5.809600e-03 2.353284e-03 2.736438e-05
#> [1891] 5.338900e-05 6.840656e-05 6.834344e-05 4.677838e-03 1.776299e-04
#> [1896] 1.147596e-04 2.343864e-05 5.430451e-04 1.082941e-04 3.419764e-04
#> [1901] 8.357849e-05 7.366893e-04 2.794765e-05 7.066999e-05 9.458434e-04
#> [1906] 3.350899e-05 4.556659e-05 4.652273e-05 3.377334e-05 4.388213e-06
#> [1911] 1.557063e-05 2.107677e-05 1.156139e-04 8.589755e-05 1.044518e-05
#> [1916] 7.552476e-05 1.430917e-04 7.591012e-05 5.019086e-04 7.093790e-04
#> [1921] 5.280272e-05 1.776299e-04 3.184480e-05 1.071876e-04 4.521071e-05
#> [1926] 8.298628e-05 1.721817e-03 7.358720e-05 2.601227e-05 1.022452e-05
#> [1931] 2.203745e-05 3.103864e-04 2.875334e-05 1.644720e-05 5.962978e-05
#> [1936] 5.857232e-05 1.282476e-05 9.061107e-05 1.439868e-05 2.255793e-04
#> [1941] 2.595447e-04 9.229255e-05 2.984005e-04 7.023397e-05 1.953449e-05
#> [1946] 2.685525e-05 1.362381e-04 3.449209e-05 4.489155e-05 1.000639e-04
#> [1951] 7.955192e-05 4.280356e-05 3.885730e-05 9.694960e-05 1.000741e-04
#> [1956] 5.906857e-05 4.980847e-03 3.739703e-05 7.888052e-05 7.420697e-05
#> [1961] 3.836347e-05 2.034874e-05 9.678581e-06 1.363487e-05 1.244604e-04
#> [1966] 4.028600e-03 5.911879e-05 1.908972e-05 1.859052e-04 8.854289e-06
#> [1971] 1.298595e-04 2.810277e-05 2.029018e-05 7.504119e-05 2.730792e-05
#> [1976] 2.436863e-05 2.762740e-03 7.591012e-05 6.068363e-05 2.645227e-04
#> [1981] 2.309639e-04 2.344656e-02 8.025214e-05 3.989944e-05 1.734398e-05
#> [1986] 2.216458e-04 5.410878e-05 1.006637e-04 8.838295e-05 9.793073e-03
#> [1991] 1.401372e-04 2.100166e-05 4.732080e-05 2.027322e-05 1.880544e-05
#> [1996] 1.574118e-05 1.762236e-04 2.545579e-05 2.947600e-05 8.060564e-05
#> [2001] 1.865876e-04 7.309790e-05 1.433365e-02 2.116373e-05 1.217184e-04
#> [2006] 7.556497e-04 2.370896e-04 4.008651e-05 9.750127e-04 5.504842e-05
#> [2011] 5.157725e-05 5.504842e-05 2.487928e-03 2.085035e-05 4.428407e-05
#> [2016] 9.003983e-05 3.445703e-03 2.601227e-05 3.445703e-03 5.268015e-05