Summarize prediction result into gene/sample-level

Summarize prediction result into gene/sample-level

gcap.runScoring(data, genome_build = "hg38", min_n = 1L, tightness = 1L)

Arguments

data

a data.table containing result from gcap.runPrediction.

genome_build

genome build version, should be one of 'hg38', 'hg19'.

min_n

a minimal cytoband number (default is 1) to determine sample class. e.g., sample with at least 1 cytoband harboring circular genes would be labelled as "circular".

tightness

a value to control the tightness to be a circular amplicon. If the value is larger, it is more likely a fCNA assigned to noncircular instead of circular.

Value

a list of data.table.

Examples

data("ec")
ec2 <- ec
ec2$prob <- gcap.runPrediction(ec)
#> <gcap> 2022-03-30 12:59:03 info [gcap.runPrediction]: using model file XGB_NF32.rds
#> <gcap> 2022-03-30 12:59:03 info [gcap.runPrediction]: selecting necessary features from input data
#> <gcap> 2022-03-30 12:59:03 info [gcap.runPrediction]: running prediction
#> [12:59:03] WARNING: amalgamation/../src/c_api/c_api.cc:718: `ntree_limit` is deprecated, use `iteration_range` instead.
score <- gcap.runScoring(ec2)
#> <gcap> 2022-03-30 12:59:03 info [gcap.runScoring]: checking input data type
#> <gcap> 2022-03-30 12:59:03 info [gcap.runScoring]: checking columns
#> <gcap> 2022-03-30 12:59:03 info [gcap.runScoring]: No 'sample' column found, mutating one
#> <gcap> 2022-03-30 12:59:03 info [gcap.runScoring]: filtering out records without prob result
#> <gcap> 2022-03-30 12:59:03 info [gcap.runScoring]: joining extra annotation data
#> <gcap> 2022-03-30 12:59:05 info [gcap.runScoring]: summarizing in gene level
#> <gcap> 2022-03-30 12:59:06 info [gcap.runScoring]: summarizing in sample level
#> <gcap> 2022-03-30 12:59:06 info [gcap.runScoring]: done
score
#> $data
#>       sample         band         gene_id total_cn minor_cn purity   ploidy
#>    1: sample  chr10:p11.1 ENSG00000075407        2        1   0.45 1.992596
#>    2: sample chr10:p11.21 ENSG00000108100        4        2   0.75 3.780270
#>    3: sample chr10:p11.21 ENSG00000148498        2        1   0.89 2.172488
#>    4: sample chr10:p11.21 ENSG00000256892        7        3   0.58 4.191208
#>    5: sample  chr10:p12.1 ENSG00000099256        4        2   0.57 3.828760
#>   ---                                                                      
#> 2016: sample   chr9:q34.3 ENSG00000181090        4        2   0.76 3.335332
#> 2017: sample   chr9:q34.3 ENSG00000186193        1        0   0.53 1.910773
#> 2018: sample   chr9:q34.3 ENSG00000188229        3        0   0.88 3.516363
#> 2019: sample   chr9:q34.3 ENSG00000261793        6        2   0.94 4.279899
#> 2020: sample   chr9:q34.3 ENSG00000284976        7        2   0.83 4.102196
#>       AScore       pLOH cna_burden       CN1       CN2       CN3       CN4
#>    1:      7 0.09848282      0.159 32.855780  5.239920  0.000000  0.000000
#>    2:     28 0.23170477      0.762  0.247997 12.031210 63.823823 19.494553
#>    3:     10 0.05947269      0.235 24.552253  7.052684  0.000000  0.413440
#>    4:     25 0.13763327      0.868  0.000000  0.000000 96.509495 31.445068
#>    5:     23 0.15308192      0.840  4.018197 42.331531  4.384916  0.034201
#>   ---                                                                     
#> 2016:     20 0.17734065      0.751  2.492506 29.382937  0.000000  5.358470
#> 2017:     17 0.25444529      0.294 22.395815  1.410976  0.000000  0.674042
#> 2018:     20 0.14110685      0.773  3.533483 34.650972  0.000000  0.000000
#> 2019:     11 0.02393007      0.974  0.000000 51.339374  0.000000  3.234558
#> 2020:     25 0.44861184      0.758  0.105810 15.238764  4.794849  3.367103
#>            CN5      CN6       CN7        CN8       CN9     CN10     CN11
#>    1:  0.00000 0.000000  1.721568   0.000000  1.182732  0.00000  0.00000
#>    2:  0.00000 0.000000  0.000000  28.241614 22.624531 60.73726  0.00000
#>    3:  0.00000 0.000000  0.000000   3.691159 14.290464  0.00000  0.00000
#>    4: 36.26579 0.000000 31.936986 212.594639  0.000000  0.00000  0.00000
#>    5:  0.00000 0.000000  9.690830  21.555225  0.899325  0.00000  0.00000
#>   ---                                                                   
#> 2016:  0.00000 8.704632 25.672591   6.724989 22.020808  0.00000  0.00000
#> 2017:  0.00000 0.000000 10.435391   0.045490 35.790772  0.00000  0.00000
#> 2018:  0.00000 0.326792  0.000000   0.000000  0.000000  0.00000  0.00000
#> 2019:  0.00000 0.000000  0.855554   0.000000  0.000000  0.00000  0.00000
#> 2020: 10.43265 0.000000 23.778204   0.000000  2.185567  0.00000 26.67354
#>           CN12     CN13     CN14      CN15     CN16      CN17       CN18
#>    1:  0.00000 0.000000  0.00000  0.000000 0.000000  0.000000   0.000000
#>    2:  0.00000 8.500558  0.00000  0.000000 2.811555  7.130902 108.355998
#>    3:  0.00000 0.000000  0.00000  0.000000 0.000000  0.000000   0.000000
#>    4:  0.00000 0.000000  0.00000 93.661275 0.000000 52.467891 188.118858
#>    5:  0.00000 0.000000  0.00000 24.085775 0.000000  0.000000   0.000000
#>   ---                                                                   
#> 2016:  0.00000 1.243929  0.00000 12.700900 2.671725  0.000000   0.000000
#> 2017:  0.00000 3.237960  0.39964  0.000000 0.000000  0.000000   7.644897
#> 2018:  0.00000 0.617547  0.00000  2.842098 0.000000  0.000000   2.464904
#> 2019:  0.00000 0.000000  0.00000  0.000000 0.000000  0.000000   1.462469
#> 2020: 20.59824 0.000000 10.06959 13.189780 9.440001  8.720157  36.405743
#>            CN19 age gender type_BLCA type_BRCA type_CESC type_COAD type_ESCA
#>    1:  0.000000  72      0         0         0         0         0         0
#>    2:  0.000000  68      1         0         0         0         0         0
#>    3:  0.000000  47      1         0         0         0         0         0
#>    4:  0.000000  84      0         0         1         0         0         0
#>    5:  0.000000  55      1         0         0         0         0         1
#>   ---                                                                       
#> 2016:  3.026513  50      0         0         1         0         0         0
#> 2017: 17.965018  84      0         0         0         0         0         1
#> 2018:  4.564205  70      1         1         0         0         0         0
#> 2019:  5.108045  71      0         0         0         0         0         0
#> 2020:  0.000000  78      0         0         1         0         0         0
#>       type_GBM type_HNSC type_KICH type_KIRP type_LGG type_LIHC type_LUAD
#>    1:        1         0         0         0        0         0         0
#>    2:        0         0         0         0        0         0         0
#>    3:        1         0         0         0        0         0         0
#>    4:        0         0         0         0        0         0         0
#>    5:        0         0         0         0        0         0         0
#>   ---                                                                    
#> 2016:        0         0         0         0        0         0         0
#> 2017:        0         0         0         0        0         0         0
#> 2018:        0         0         0         0        0         0         0
#> 2019:        0         0         0         0        0         0         0
#> 2020:        0         0         0         0        0         0         0
#>       type_LUSC type_OV type_PRAD type_SARC type_SKCM type_STAD type_UCEC
#>    1:         0       0         0         0         0         0         0
#>    2:         1       0         0         0         0         0         0
#>    3:         0       0         0         0         0         0         0
#>    4:         0       0         0         0         0         0         0
#>    5:         0       0         0         0         0         0         0
#>   ---                                                                    
#> 2016:         0       0         0         0         0         0         0
#> 2017:         0       0         0         0         0         0         0
#> 2018:         0       0         0         0         0         0         0
#> 2019:         0       0         0         0         0         0         0
#> 2020:         0       0         0         0         0         0         0
#>       type_UVM freq_Linear   freq_BFB freq_Circular   freq_HR y_Circular
#>    1:        0   1.2453300  0.6226650     2.4906600 2.1793275          0
#>    2:        0   0.3113325  0.0000000     1.5566625 1.2453300          0
#>    3:        0   0.3113325  0.0000000     0.9339975 3.1133250          0
#>    4:        0   5.9153176 12.4533001     4.3586550 5.2926526          0
#>    5:        0   0.3113325  0.3113325     1.2453300 1.2453300          0
#>   ---                                                                   
#> 2016:        0   1.5566625  0.3113325     2.1793275 0.9339975          0
#> 2017:        0   0.0000000  0.6226650     4.0473225 2.4906600          0
#> 2018:        0   1.8679950  0.3113325     2.4906600 2.8019925          0
#> 2019:        1   4.9813200 12.1419676     4.0473225 4.6699875          0
#> 2020:        0   0.3113325  2.4906600     0.6226650 2.1793275          0
#>              prob background_cn amplicon_type
#>    1: 0.001358518      1.992596       nofocal
#>    2: 0.001366050      3.780270       nofocal
#>    3: 0.001358518      2.172488       nofocal
#>    4: 0.007997432      4.191208       nofocal
#>    5: 0.001366050      3.828760       nofocal
#>   ---                                        
#> 2016: 0.001489777      3.335332       nofocal
#> 2017: 0.001358518      1.910773       nofocal
#> 2018: 0.001358518      3.516363       nofocal
#> 2019: 0.005260550      4.279899       nofocal
#> 2020: 0.014405195      4.102196       nofocal
#> 
#> $gene
#>             gene_id     amplicon_type N total_cn minor_cn cytoband_cn
#>  1: ENSG00000166796          circular 1       26        4         3.0
#>  2: ENSG00000162139 possibly_circular 1       15        3         4.5
#>  3: ENSG00000023697          circular 1        8        2         4.0
#>  4: ENSG00000111729          circular 1       12        2         2.0
#>  5: ENSG00000175189          circular 1       31        2         3.0
#>  6: ENSG00000125246 possibly_circular 1        9        3         4.0
#>  7: ENSG00000103657          circular 1       10        2         2.0
#>  8: ENSG00000197562          circular 1       11        2         2.0
#>  9: ENSG00000090659          circular 1       12        2         3.0
#> 10: ENSG00000065717          circular 1       17        3         3.0
#> 11: ENSG00000083817          circular 1       10        3         3.0
#> 12: ENSG00000184588          circular 1       19        2         4.0
#> 13: ENSG00000243749          circular 1       13        2         2.0
#> 14: ENSG00000159086          circular 1       14        3         3.0
#> 15: ENSG00000160799          circular 1       17        2         3.0
#> 16: ENSG00000205678 possibly_circular 1       11        2        11.0
#> 17: ENSG00000164610          circular 1       12        2         3.0
#> 18: ENSG00000266265 possibly_circular 1        8        1         5.5
#>     support_IDs         band
#>  1:      sample  chr11:p15.1
#>  2:      sample  chr11:q13.4
#>  3:      sample  chr12:p12.3
#>  4:      sample chr12:p13.31
#>  5:      sample  chr12:q13.3
#>  6:      sample  chr13:q32.3
#>  7:      sample chr15:q22.31
#>  8:      sample  chr16:p13.3
#>  9:      sample  chr19:p13.2
#> 10:      sample  chr19:p13.3
#> 11:      sample chr19:q13.43
#> 12:      sample   chr1:p31.3
#> 13:      sample   chr1:p34.3
#> 14:      sample chr21:q22.11
#> 15:      sample  chr3:p21.31
#> 16:      sample   chr4:q13.1
#> 17:      sample   chr7:p14.3
#> 18:      sample   chr7:q32.2
#> 
#> $sample
#>    sample ec_genes ec_cytobands
#> 1: sample       14           14
#>                                                                                                                                                         ec_cytobands_detail
#> 1: chr11:p15.1,chr12:p12.3,chr12:p13.31,chr12:q13.3,chr15:q22.31,chr16:p13.3,chr19:p13.2,chr19:p13.3,chr19:q13.43,chr1:p31.3,chr1:p34.3,chr21:q22.11,chr3:p21.31,chr7:p14.3
#>    ec_possibly_genes    class purity   ploidy AScore       pLOH cna_burden
#> 1:                 4 circular   0.45 1.992596      7 0.09848282      0.159
#>         CN1     CN2 CN3 CN4 CN5 CN6      CN7 CN8      CN9 CN10 CN11 CN12 CN13
#> 1: 32.85578 5.23992   0   0   0   0 1.721568   0 1.182732    0    0    0    0
#>    CN14 CN15 CN16 CN17 CN18 CN19
#> 1:    0    0    0    0    0    0
#>