GCAP workflow for gene-level amplicon prediction from ASCN input

Unlike gcap.workflow, this function directly uses the allele-specific copy number data along with some extra sample information to infer ecDNA genes.

gcap.ASCNworkflow(
  data,
  genome_build = c("hg38", "hg19"),
  target = c("circle", "nonLinear"),
  prob_cutoff = 0.9,
  outdir = getwd(),
  result_file_prefix = paste0("gcap_", uuid::UUIDgenerate(TRUE))
)

Arguments

data
a data.frame with following rows. The key columns can be obtained from common allele specific CNV calling software, e.g., ASCAT, Sequenza, FACETS.

chromosome: chromosome names starts with 'chr'.

start: start position of the segment.

end: end position of the segment.

total_cn: total integer copy number of the segment.

minor_cn: minor allele integer copy number of the segment.

sample: sample identifier.

purity: tumor purity of the sample. Set to 1 if you don't know.

ploidy (optinal): ploidy value of the sample tumor genome.

age (optional): age of the case, use along with gender.

gender (optional): gender of the case, use along with age.

genome_build
genome build version, should be one of 'hg38', 'hg19'.

target
'circle' and 'nonLinear' to select model if predict circle amplicon or non-linear amplicon. Can be one or both.

prob_cutoff
a value in [0-1] to determine the cutoff for labeling response. The default cutoff 0.9 is a balance between F1 score and amplicon load accuracy.

outdir
result output path.

result_file_prefix
file name prefix (without directory path) for storing final model prediction file in CSV format. Default a unique file name is generated by UUID approach.

Value

a list of invisible data.table and corresponding files saved to local machine.

Examples

data("ascn")
data <- ascn
rv <- gcap.ASCNworkflow(data, outdir = tempdir())
data$purity <- 1
rv2 <- gcap.ASCNworkflow(data, outdir = tempdir())
data$age <- 60
data$gender <- "XY"
rv3 <- gcap.ASCNworkflow(data, outdir = tempdir())