GCAP workflow for gene-level amplicon prediction from ASCN input

Unlike gcap.workflow, this function directly uses the allele-specific copy number data along with some extra sample information to infer ecDNA genes.

gcap.ASCNworkflow(
  data,
  genome_build = c("hg38", "hg19"),
  model = "XGB32",
  tightness = 1L,
  gap_cn = 4L,
  outdir = getwd(),
  result_file_prefix = paste0("gcap_", uuid::UUIDgenerate(TRUE))
)

Arguments

data

a data.frame with following columns. The key columns can be obtained from common allele specific CNV calling software, e.g., ASCAT, Sequenza, FACETS.

chromosome: chromosome names starts with 'chr'.

start: start position of the segment.

end: end position of the segment.

total_cn: total integer copy number of the segment.

minor_cn: minor allele integer copy number of the segment. Set it to NA if you don't have this data.

sample: sample identifier.

purity: tumor purity of the sample. Set to 1 if you don't know.

ploidy (optinal): ploidy value of the sample tumor genome.

age (optional): age of the case, use along with gender.

gender (optional): gender of the case, use along with age.

type (optional): cancer type of the case, use along with age and gender. Please refer to gcap.collapse2Genes to see the supported cancer types. This info is only used in 'XGB56' model. If you don't use this model, you don't need to set it.

genome_build

genome build version, should be one of 'hg38', 'hg19'.

model

model name ("XGB11", "XGB32", "XGB56") or a custom model from input. 'toy' can be used for test.

tightness

a value to control the tightness to be a circular amplicon. If the value is larger, it is more likely a fCNA assigned to noncircular instead of circular.

gap_cn

a gap copy number value, default is 4L refer to Kim 2020 Nat.Gen. A gene with copy number above ploidy + gap_cn would be treated as amplicon. Smaller, more amplicons.

outdir

result output path.

result_file_prefix

file name prefix (without directory path) for storing final model prediction file in CSV format. Default a unique file name is generated by UUID approach.

Value

a list of invisible data.table and corresponding files saved to local machine.

Examples

data("ascn")
data <- ascn
rv <- gcap.ASCNworkflow(data, outdir = tempdir(), model = "XGB11")
#> <gcap> 2022-03-30 17:15:32 info [gcap.ASCNworkflow]: =========================
#> <gcap> 2022-03-30 17:15:32 info [gcap.ASCNworkflow]:    GCAP ASCN WORKFLOW    
#> <gcap> 2022-03-30 17:15:32 info [gcap.ASCNworkflow]: =========================
#> 
#> <gcap> 2022-03-30 17:15:32 info [gcap.ASCNworkflow]: ============================================================
#> <gcap> 2022-03-30 17:15:32 info [gcap.ASCNworkflow]: Step 1: Extract features and collapse features to gene level
#> <gcap> 2022-03-30 17:15:32 info [gcap.ASCNworkflow]: ============================================================
#> <gcap> 2022-03-30 17:15:32 info [gcap.runASCATBuildflow]: extracting sample-level and region-level features
#> <gcap> 2022-03-30 17:15:32 info [gcap.extractFeatures]: > Extract features from input ascn_data<
#> <gcap> 2022-03-30 17:15:32 info [gcap.extractFeatures]: generating CopyNumber object in sigminer package
#> i [2022-03-30 17:15:32]: Started.
#> i [2022-03-30 17:15:32]: Genome build  : hg38.
#> i [2022-03-30 17:15:32]: Genome measure: called.
#> i [2022-03-30 17:15:32]: When add_loh is TRUE, use_all is forced to TRUE.
#> Please drop columns you don't want to keep before reading.
#> v [2022-03-30 17:15:32]: Chromosome size database for build obtained.
#> i [2022-03-30 17:15:32]: Reading input.
#> v [2022-03-30 17:15:32]: A data frame as input detected.
#> v [2022-03-30 17:15:32]: Column names checked.
#> v [2022-03-30 17:15:32]: Column order set.
#> v [2022-03-30 17:15:32]: Chromosomes unified.
#> v [2022-03-30 17:15:32]: Data imported.
#> i [2022-03-30 17:15:32]: Segments info:
#> i [2022-03-30 17:15:32]:     Keep - 129
#> i [2022-03-30 17:15:32]:     Drop - 0
#> v [2022-03-30 17:15:32]: Segments sorted.
#> i [2022-03-30 17:15:32]: Adding LOH labels...
#> i [2022-03-30 17:15:32]: Skipped joining adjacent segments with same copy number value.
#> v [2022-03-30 17:15:32]: Segmental table cleaned.
#> i [2022-03-30 17:15:32]: Annotating.
#> v [2022-03-30 17:15:32]: Annotation done.
#> i [2022-03-30 17:15:32]: Summarizing per sample.
#> v [2022-03-30 17:15:32]: Summarized.
#> i [2022-03-30 17:15:32]: Generating CopyNumber object.
#> v [2022-03-30 17:15:32]: Generated.
#> i [2022-03-30 17:15:32]: Validating object.
#> v [2022-03-30 17:15:32]: Done.
#> i [2022-03-30 17:15:32]: 0.218 secs elapsed.
#> <gcap> 2022-03-30 17:15:32 info [gcap.extractFeatures]: estimating ploidy from copy number data
#> <gcap> 2022-03-30 17:15:32 info [gcap.extractFeatures]: checking if input data contains ploidy and if there are NAs should be overwritten
#> <gcap> 2022-03-30 17:15:32 info [gcap.extractFeatures]: getting Aneuploidy score
#> <gcap> 2022-03-30 17:15:32 info [gcap.extractFeatures]: getting pLOH score
#> <gcap> 2022-03-30 17:15:32 info [gcap.extractFeatures]: getting CNA burden
#> <gcap> 2022-03-30 17:15:32 info [gcap.extractFeatures]: generating copy number catalog matrix for fitting signature activity
#> i [2022-03-30 17:15:32]: Started.
#> i [2022-03-30 17:15:32]: When you use method 'S', please make sure you have set 'join_adj_seg' to FALSE and 'add_loh' to TRUE in 'read_copynumber() in the previous step!
#> v [2022-03-30 17:15:32]: Matrix generated.
#> i [2022-03-30 17:15:32]: 0.077 secs elapsed.
#> <gcap> 2022-03-30 17:15:32 info [gcap.extractFeatures]: fitting copy number signature activity
#> i [2022-03-30 17:15:32]: Started.
#> v [2022-03-30 17:15:32]: Signature index detected.
#> i [2022-03-30 17:15:32]: Checking signature database in package.
#> i [2022-03-30 17:15:33]: Checking signature index.
#> i [2022-03-30 17:15:33]: Valid index for db 'CNS_TCGA':
#> CN1 CN2 CN3 CN4 CN5 CN6 CN7 CN8 CN9 CN10 CN11 CN12 CN13 CN14 CN15 CN16 CN17 CN18 CN19
#> v [2022-03-30 17:15:33]: Database and index checked.
#> v [2022-03-30 17:15:33]: Signature normalized.
#> i [2022-03-30 17:15:33]: Checking row number for catalog matrix and signature matrix.
#> v [2022-03-30 17:15:33]: Checked.
#> i [2022-03-30 17:15:33]: Checking rownames for catalog matrix and signature matrix.
#> v [2022-03-30 17:15:33]: Checked.
#> v [2022-03-30 17:15:33]: Method 'QP' detected.
#> v [2022-03-30 17:15:33]: Corresponding function generated.
#> i [2022-03-30 17:15:33]: Calling function.
#> i [2022-03-30 17:15:33]: Fitting sample: TCGA-02-2485-01
#> i [2022-03-30 17:15:33]: Fitting sample: TCGA-05-4422-01
#> v [2022-03-30 17:15:33]: Done.
#> i [2022-03-30 17:15:33]: Generating output signature exposures.
#> v [2022-03-30 17:15:33]: Done.
#> i [2022-03-30 17:15:33]: 0.062 secs elapsed.
#> <gcap> 2022-03-30 17:15:33 info [gcap.extractFeatures]: merging data
#> <gcap> 2022-03-30 17:15:33 info [gcap.extractFeatures]: feature extraction done
#> <gcap> 2022-03-30 17:15:33 info [gcap.extractFeatures]: now you can modify the result and append 'age' and 'gender' columns to the 'fts_sample' element of result list
#> <gcap> 2022-03-30 17:15:33 info [gcap.runASCATBuildflow]: checking if input age, gender (and type) columns
#> <gcap> 2022-03-30 17:15:33 info [gcap.runASCATBuildflow]: None exist, set to NULL.
#> <gcap> 2022-03-30 17:15:33 info [gcap.runASCATBuildflow]: collapsing all data into gene-level prediction input
#> <gcap> 2022-03-30 17:15:33 info [gcap.collapse2Genes]: please make sure the first 3 columns of `fts$fts_region` are for chr, start, end.
#> <gcap> 2022-03-30 17:15:33 info [gcap.collapse2Genes]: collapsing region-level features to gene-level
#> <gcap> 2022-03-30 17:15:33 info [collapse_to_genes]: checking input chromosome names
#> <gcap> 2022-03-30 17:15:33 info [collapse_to_genes]: reading reference file D:/Rlib/gcap/extdata/hg38_target_genes.rds
#> <gcap> 2022-03-30 17:15:33 info [collapse_to_genes]: finding overlaps
#> <gcap> 2022-03-30 17:15:33 info [collapse_to_genes]: calculating intersect size
#> <gcap> 2022-03-30 17:15:33 info [collapse_to_genes]: keeping records with >= 100% overlap ratio with a gene
#> <gcap> 2022-03-30 17:15:33 info [gcap.collapse2Genes]: merging gene-level and sample-level data
#> <gcap> 2022-03-30 17:15:33 info [gcap.collapse2Genes]: merging data and prior amplicon frequency data
#> <gcap> 2022-03-30 17:15:33 info [gcap.collapse2Genes]: done
#> <gcap> 2022-03-30 17:15:33 info [gcap.ASCNworkflow]: =======================
#> <gcap> 2022-03-30 17:15:33 info [gcap.ASCNworkflow]: Step 2: Run prediction
#> <gcap> 2022-03-30 17:15:33 info [gcap.ASCNworkflow]: =======================
#> <gcap> 2022-03-30 17:15:33 info [gcap.runPrediction]: using model file XGB_NF11.rds
#> <gcap> 2022-03-30 17:15:34 info [gcap.runPrediction]: selecting necessary features from input data
#> <gcap> 2022-03-30 17:15:34 info [gcap.runPrediction]: running prediction
#> [17:15:34] WARNING: amalgamation/../src/c_api/c_api.cc:718: `ntree_limit` is deprecated, use `iteration_range` instead.
#> <gcap> 2022-03-30 17:15:34 info [gcap.ASCNworkflow]: ====================================
#> <gcap> 2022-03-30 17:15:34 info [gcap.ASCNworkflow]: Step 3: Run scoring and summarizing
#> <gcap> 2022-03-30 17:15:34 info [gcap.ASCNworkflow]: ====================================
#> <gcap> 2022-03-30 17:15:34 info [gcap.runScoring]: checking input data type
#> <gcap> 2022-03-30 17:15:34 info [gcap.runScoring]: checking columns
#> <gcap> 2022-03-30 17:15:34 info [gcap.runScoring]: filtering out records without prob result
#> <gcap> 2022-03-30 17:15:34 info [gcap.runScoring]: joining extra annotation data
#> <gcap> 2022-03-30 17:15:37 info [gcap.runScoring]: summarizing in gene level
#> <gcap> 2022-03-30 17:15:40 info [gcap.runScoring]: summarizing in sample level
#> <gcap> 2022-03-30 17:15:40 info [gcap.runScoring]: done
#> <gcap> 2022-03-30 17:15:41 info [gcap.ASCNworkflow]: Saving raw prediction result to C:\Users\ADMINI~1\AppData\Local\Temp\Rtmpuer5z4/gcap_f86e1914-b009-11ec-8000-132265b5ee3a_prediction_result.rds
#> <gcap> 2022-03-30 17:15:41 info [gcap.ASCNworkflow]: Saving gene result to C:\Users\ADMINI~1\AppData\Local\Temp\Rtmpuer5z4/gcap_f86e1914-b009-11ec-8000-132265b5ee3a_by_gene.csv
#> <gcap> 2022-03-30 17:15:41 info [gcap.ASCNworkflow]: Saving sample result to C:\Users\ADMINI~1\AppData\Local\Temp\Rtmpuer5z4/gcap_f86e1914-b009-11ec-8000-132265b5ee3a_by_sample.csv
#> <gcap> 2022-03-30 17:15:41 info [gcap.ASCNworkflow]: ===========================================
#> <gcap> 2022-03-30 17:15:41 info [gcap.ASCNworkflow]:  Done! Thanks for using GCAP ASCN workflow
#> <gcap> 2022-03-30 17:15:41 info [gcap.ASCNworkflow]: ===========================================
data$purity <- 1
rv2 <- gcap.ASCNworkflow(data, outdir = tempdir(), model = "XGB11")
#> <gcap> 2022-03-30 17:15:41 info [gcap.ASCNworkflow]: =========================
#> <gcap> 2022-03-30 17:15:41 info [gcap.ASCNworkflow]:    GCAP ASCN WORKFLOW    
#> <gcap> 2022-03-30 17:15:41 info [gcap.ASCNworkflow]: =========================
#> 
#> <gcap> 2022-03-30 17:15:41 info [gcap.ASCNworkflow]: ============================================================
#> <gcap> 2022-03-30 17:15:41 info [gcap.ASCNworkflow]: Step 1: Extract features and collapse features to gene level
#> <gcap> 2022-03-30 17:15:41 info [gcap.ASCNworkflow]: ============================================================
#> <gcap> 2022-03-30 17:15:41 info [gcap.runASCATBuildflow]: extracting sample-level and region-level features
#> <gcap> 2022-03-30 17:15:41 info [gcap.extractFeatures]: > Extract features from input ascn_data<
#> <gcap> 2022-03-30 17:15:41 info [gcap.extractFeatures]: generating CopyNumber object in sigminer package
#> i [2022-03-30 17:15:41]: Started.
#> i [2022-03-30 17:15:41]: Genome build  : hg38.
#> i [2022-03-30 17:15:41]: Genome measure: called.
#> i [2022-03-30 17:15:41]: When add_loh is TRUE, use_all is forced to TRUE.
#> Please drop columns you don't want to keep before reading.
#> v [2022-03-30 17:15:41]: Chromosome size database for build obtained.
#> i [2022-03-30 17:15:41]: Reading input.
#> v [2022-03-30 17:15:41]: A data frame as input detected.
#> v [2022-03-30 17:15:41]: Column names checked.
#> v [2022-03-30 17:15:41]: Column order set.
#> v [2022-03-30 17:15:41]: Chromosomes unified.
#> v [2022-03-30 17:15:41]: Data imported.
#> i [2022-03-30 17:15:41]: Segments info:
#> i [2022-03-30 17:15:41]:     Keep - 129
#> i [2022-03-30 17:15:41]:     Drop - 0
#> v [2022-03-30 17:15:41]: Segments sorted.
#> i [2022-03-30 17:15:41]: Adding LOH labels...
#> i [2022-03-30 17:15:41]: Skipped joining adjacent segments with same copy number value.
#> v [2022-03-30 17:15:41]: Segmental table cleaned.
#> i [2022-03-30 17:15:41]: Annotating.
#> v [2022-03-30 17:15:41]: Annotation done.
#> i [2022-03-30 17:15:41]: Summarizing per sample.
#> v [2022-03-30 17:15:41]: Summarized.
#> i [2022-03-30 17:15:41]: Generating CopyNumber object.
#> v [2022-03-30 17:15:41]: Generated.
#> i [2022-03-30 17:15:41]: Validating object.
#> v [2022-03-30 17:15:41]: Done.
#> i [2022-03-30 17:15:41]: 0.085 secs elapsed.
#> <gcap> 2022-03-30 17:15:41 info [gcap.extractFeatures]: estimating ploidy from copy number data
#> <gcap> 2022-03-30 17:15:41 info [gcap.extractFeatures]: checking if input data contains ploidy and if there are NAs should be overwritten
#> <gcap> 2022-03-30 17:15:41 info [gcap.extractFeatures]: getting Aneuploidy score
#> <gcap> 2022-03-30 17:15:41 info [gcap.extractFeatures]: getting pLOH score
#> <gcap> 2022-03-30 17:15:41 info [gcap.extractFeatures]: getting CNA burden
#> <gcap> 2022-03-30 17:15:41 info [gcap.extractFeatures]: generating copy number catalog matrix for fitting signature activity
#> i [2022-03-30 17:15:41]: Started.
#> i [2022-03-30 17:15:41]: When you use method 'S', please make sure you have set 'join_adj_seg' to FALSE and 'add_loh' to TRUE in 'read_copynumber() in the previous step!
#> v [2022-03-30 17:15:41]: Matrix generated.
#> i [2022-03-30 17:15:41]: 0.034 secs elapsed.
#> <gcap> 2022-03-30 17:15:41 info [gcap.extractFeatures]: fitting copy number signature activity
#> i [2022-03-30 17:15:41]: Started.
#> v [2022-03-30 17:15:41]: Signature index detected.
#> i [2022-03-30 17:15:41]: Checking signature database in package.
#> i [2022-03-30 17:15:41]: Checking signature index.
#> i [2022-03-30 17:15:41]: Valid index for db 'CNS_TCGA':
#> CN1 CN2 CN3 CN4 CN5 CN6 CN7 CN8 CN9 CN10 CN11 CN12 CN13 CN14 CN15 CN16 CN17 CN18 CN19
#> v [2022-03-30 17:15:41]: Database and index checked.
#> v [2022-03-30 17:15:41]: Signature normalized.
#> i [2022-03-30 17:15:41]: Checking row number for catalog matrix and signature matrix.
#> v [2022-03-30 17:15:41]: Checked.
#> i [2022-03-30 17:15:41]: Checking rownames for catalog matrix and signature matrix.
#> v [2022-03-30 17:15:41]: Checked.
#> v [2022-03-30 17:15:41]: Method 'QP' detected.
#> v [2022-03-30 17:15:41]: Corresponding function generated.
#> i [2022-03-30 17:15:41]: Calling function.
#> i [2022-03-30 17:15:41]: Fitting sample: TCGA-02-2485-01
#> i [2022-03-30 17:15:41]: Fitting sample: TCGA-05-4422-01
#> v [2022-03-30 17:15:41]: Done.
#> i [2022-03-30 17:15:41]: Generating output signature exposures.
#> v [2022-03-30 17:15:41]: Done.
#> i [2022-03-30 17:15:41]: 0.037 secs elapsed.
#> <gcap> 2022-03-30 17:15:41 info [gcap.extractFeatures]: merging data
#> <gcap> 2022-03-30 17:15:41 info [gcap.extractFeatures]: feature extraction done
#> <gcap> 2022-03-30 17:15:41 info [gcap.extractFeatures]: now you can modify the result and append 'age' and 'gender' columns to the 'fts_sample' element of result list
#> <gcap> 2022-03-30 17:15:41 info [gcap.runASCATBuildflow]: checking if input age, gender (and type) columns
#> <gcap> 2022-03-30 17:15:41 info [gcap.runASCATBuildflow]: None exist, set to NULL.
#> <gcap> 2022-03-30 17:15:41 info [gcap.runASCATBuildflow]: collapsing all data into gene-level prediction input
#> <gcap> 2022-03-30 17:15:41 info [gcap.collapse2Genes]: please make sure the first 3 columns of `fts$fts_region` are for chr, start, end.
#> <gcap> 2022-03-30 17:15:41 info [gcap.collapse2Genes]: collapsing region-level features to gene-level
#> <gcap> 2022-03-30 17:15:41 info [collapse_to_genes]: checking input chromosome names
#> <gcap> 2022-03-30 17:15:41 info [collapse_to_genes]: reading reference file D:/Rlib/gcap/extdata/hg38_target_genes.rds
#> <gcap> 2022-03-30 17:15:41 info [collapse_to_genes]: finding overlaps
#> <gcap> 2022-03-30 17:15:41 info [collapse_to_genes]: calculating intersect size
#> <gcap> 2022-03-30 17:15:41 info [collapse_to_genes]: keeping records with >= 100% overlap ratio with a gene
#> <gcap> 2022-03-30 17:15:41 info [gcap.collapse2Genes]: merging gene-level and sample-level data
#> <gcap> 2022-03-30 17:15:41 info [gcap.collapse2Genes]: merging data and prior amplicon frequency data
#> <gcap> 2022-03-30 17:15:41 info [gcap.collapse2Genes]: done
#> <gcap> 2022-03-30 17:15:41 info [gcap.ASCNworkflow]: =======================
#> <gcap> 2022-03-30 17:15:41 info [gcap.ASCNworkflow]: Step 2: Run prediction
#> <gcap> 2022-03-30 17:15:41 info [gcap.ASCNworkflow]: =======================
#> <gcap> 2022-03-30 17:15:41 info [gcap.runPrediction]: using model file XGB_NF11.rds
#> <gcap> 2022-03-30 17:15:41 info [gcap.runPrediction]: selecting necessary features from input data
#> <gcap> 2022-03-30 17:15:41 info [gcap.runPrediction]: running prediction
#> [17:15:41] WARNING: amalgamation/../src/c_api/c_api.cc:718: `ntree_limit` is deprecated, use `iteration_range` instead.
#> <gcap> 2022-03-30 17:15:41 info [gcap.ASCNworkflow]: ====================================
#> <gcap> 2022-03-30 17:15:41 info [gcap.ASCNworkflow]: Step 3: Run scoring and summarizing
#> <gcap> 2022-03-30 17:15:41 info [gcap.ASCNworkflow]: ====================================
#> <gcap> 2022-03-30 17:15:41 info [gcap.runScoring]: checking input data type
#> <gcap> 2022-03-30 17:15:41 info [gcap.runScoring]: checking columns
#> <gcap> 2022-03-30 17:15:41 info [gcap.runScoring]: filtering out records without prob result
#> <gcap> 2022-03-30 17:15:41 info [gcap.runScoring]: joining extra annotation data
#> <gcap> 2022-03-30 17:15:44 info [gcap.runScoring]: summarizing in gene level
#> <gcap> 2022-03-30 17:15:46 info [gcap.runScoring]: summarizing in sample level
#> <gcap> 2022-03-30 17:15:46 info [gcap.runScoring]: done
#> <gcap> 2022-03-30 17:15:47 info [gcap.ASCNworkflow]: Saving raw prediction result to C:\Users\ADMINI~1\AppData\Local\Temp\Rtmpuer5z4/gcap_fc0fe55c-b009-11ec-8000-132265b5ee3a_prediction_result.rds
#> <gcap> 2022-03-30 17:15:47 info [gcap.ASCNworkflow]: Saving gene result to C:\Users\ADMINI~1\AppData\Local\Temp\Rtmpuer5z4/gcap_fc0fe55c-b009-11ec-8000-132265b5ee3a_by_gene.csv
#> <gcap> 2022-03-30 17:15:47 info [gcap.ASCNworkflow]: Saving sample result to C:\Users\ADMINI~1\AppData\Local\Temp\Rtmpuer5z4/gcap_fc0fe55c-b009-11ec-8000-132265b5ee3a_by_sample.csv
#> <gcap> 2022-03-30 17:15:47 info [gcap.ASCNworkflow]: ===========================================
#> <gcap> 2022-03-30 17:15:47 info [gcap.ASCNworkflow]:  Done! Thanks for using GCAP ASCN workflow
#> <gcap> 2022-03-30 17:15:47 info [gcap.ASCNworkflow]: ===========================================
data$age <- 60
data$gender <- "XY"
rv3 <- gcap.ASCNworkflow(data, outdir = tempdir(), model = "XGB32")
#> <gcap> 2022-03-30 17:15:47 info [gcap.ASCNworkflow]: =========================
#> <gcap> 2022-03-30 17:15:47 info [gcap.ASCNworkflow]:    GCAP ASCN WORKFLOW    
#> <gcap> 2022-03-30 17:15:47 info [gcap.ASCNworkflow]: =========================
#> 
#> <gcap> 2022-03-30 17:15:47 info [gcap.ASCNworkflow]: ============================================================
#> <gcap> 2022-03-30 17:15:47 info [gcap.ASCNworkflow]: Step 1: Extract features and collapse features to gene level
#> <gcap> 2022-03-30 17:15:47 info [gcap.ASCNworkflow]: ============================================================
#> <gcap> 2022-03-30 17:15:47 info [gcap.runASCATBuildflow]: extracting sample-level and region-level features
#> <gcap> 2022-03-30 17:15:47 info [gcap.extractFeatures]: > Extract features from input ascn_data<
#> <gcap> 2022-03-30 17:15:47 info [gcap.extractFeatures]: generating CopyNumber object in sigminer package
#> i [2022-03-30 17:15:47]: Started.
#> i [2022-03-30 17:15:47]: Genome build  : hg38.
#> i [2022-03-30 17:15:47]: Genome measure: called.
#> i [2022-03-30 17:15:47]: When add_loh is TRUE, use_all is forced to TRUE.
#> Please drop columns you don't want to keep before reading.
#> v [2022-03-30 17:15:47]: Chromosome size database for build obtained.
#> i [2022-03-30 17:15:47]: Reading input.
#> v [2022-03-30 17:15:47]: A data frame as input detected.
#> v [2022-03-30 17:15:47]: Column names checked.
#> v [2022-03-30 17:15:47]: Column order set.
#> v [2022-03-30 17:15:47]: Chromosomes unified.
#> v [2022-03-30 17:15:47]: Data imported.
#> i [2022-03-30 17:15:47]: Segments info:
#> i [2022-03-30 17:15:47]:     Keep - 129
#> i [2022-03-30 17:15:47]:     Drop - 0
#> v [2022-03-30 17:15:47]: Segments sorted.
#> i [2022-03-30 17:15:47]: Adding LOH labels...
#> i [2022-03-30 17:15:47]: Skipped joining adjacent segments with same copy number value.
#> v [2022-03-30 17:15:47]: Segmental table cleaned.
#> i [2022-03-30 17:15:47]: Annotating.
#> v [2022-03-30 17:15:47]: Annotation done.
#> i [2022-03-30 17:15:47]: Summarizing per sample.
#> v [2022-03-30 17:15:47]: Summarized.
#> i [2022-03-30 17:15:47]: Generating CopyNumber object.
#> v [2022-03-30 17:15:47]: Generated.
#> i [2022-03-30 17:15:47]: Validating object.
#> v [2022-03-30 17:15:47]: Done.
#> i [2022-03-30 17:15:47]: 0.126 secs elapsed.
#> <gcap> 2022-03-30 17:15:47 info [gcap.extractFeatures]: estimating ploidy from copy number data
#> <gcap> 2022-03-30 17:15:47 info [gcap.extractFeatures]: checking if input data contains ploidy and if there are NAs should be overwritten
#> <gcap> 2022-03-30 17:15:47 info [gcap.extractFeatures]: getting Aneuploidy score
#> <gcap> 2022-03-30 17:15:47 info [gcap.extractFeatures]: getting pLOH score
#> <gcap> 2022-03-30 17:15:47 info [gcap.extractFeatures]: getting CNA burden
#> <gcap> 2022-03-30 17:15:47 info [gcap.extractFeatures]: generating copy number catalog matrix for fitting signature activity
#> i [2022-03-30 17:15:47]: Started.
#> i [2022-03-30 17:15:47]: When you use method 'S', please make sure you have set 'join_adj_seg' to FALSE and 'add_loh' to TRUE in 'read_copynumber() in the previous step!
#> v [2022-03-30 17:15:47]: Matrix generated.
#> i [2022-03-30 17:15:47]: 0.052 secs elapsed.
#> <gcap> 2022-03-30 17:15:47 info [gcap.extractFeatures]: fitting copy number signature activity
#> i [2022-03-30 17:15:47]: Started.
#> v [2022-03-30 17:15:47]: Signature index detected.
#> i [2022-03-30 17:15:47]: Checking signature database in package.
#> i [2022-03-30 17:15:47]: Checking signature index.
#> i [2022-03-30 17:15:47]: Valid index for db 'CNS_TCGA':
#> CN1 CN2 CN3 CN4 CN5 CN6 CN7 CN8 CN9 CN10 CN11 CN12 CN13 CN14 CN15 CN16 CN17 CN18 CN19
#> v [2022-03-30 17:15:47]: Database and index checked.
#> v [2022-03-30 17:15:47]: Signature normalized.
#> i [2022-03-30 17:15:47]: Checking row number for catalog matrix and signature matrix.
#> v [2022-03-30 17:15:47]: Checked.
#> i [2022-03-30 17:15:47]: Checking rownames for catalog matrix and signature matrix.
#> v [2022-03-30 17:15:47]: Checked.
#> v [2022-03-30 17:15:47]: Method 'QP' detected.
#> v [2022-03-30 17:15:47]: Corresponding function generated.
#> i [2022-03-30 17:15:47]: Calling function.
#> i [2022-03-30 17:15:47]: Fitting sample: TCGA-02-2485-01
#> i [2022-03-30 17:15:47]: Fitting sample: TCGA-05-4422-01
#> v [2022-03-30 17:15:47]: Done.
#> i [2022-03-30 17:15:47]: Generating output signature exposures.
#> v [2022-03-30 17:15:47]: Done.
#> i [2022-03-30 17:15:47]: 0.066 secs elapsed.
#> <gcap> 2022-03-30 17:15:47 info [gcap.extractFeatures]: merging data
#> <gcap> 2022-03-30 17:15:47 info [gcap.extractFeatures]: feature extraction done
#> <gcap> 2022-03-30 17:15:47 info [gcap.extractFeatures]: now you can modify the result and append 'age' and 'gender' columns to the 'fts_sample' element of result list
#> <gcap> 2022-03-30 17:15:47 info [gcap.runASCATBuildflow]: checking if input age, gender (and type) columns
#> <gcap> 2022-03-30 17:15:47 info [gcap.runASCATBuildflow]: collapsing all data into gene-level prediction input
#> <gcap> 2022-03-30 17:15:47 info [gcap.collapse2Genes]: please make sure the first 3 columns of `fts$fts_region` are for chr, start, end.
#> <gcap> 2022-03-30 17:15:47 info [gcap.collapse2Genes]: checking data type for age column
#> <gcap> 2022-03-30 17:15:47 info [gcap.collapse2Genes]: checking data type for gender column
#> <gcap> 2022-03-30 17:15:47 warn [gcap.collapse2Genes]: non numeric type found, transforming XY to 1, otherwise 0
#> <gcap> 2022-03-30 17:15:47 info [gcap.collapse2Genes]: collapsing region-level features to gene-level
#> <gcap> 2022-03-30 17:15:47 info [collapse_to_genes]: checking input chromosome names
#> <gcap> 2022-03-30 17:15:47 info [collapse_to_genes]: reading reference file D:/Rlib/gcap/extdata/hg38_target_genes.rds
#> <gcap> 2022-03-30 17:15:47 info [collapse_to_genes]: finding overlaps
#> <gcap> 2022-03-30 17:15:47 info [collapse_to_genes]: calculating intersect size
#> <gcap> 2022-03-30 17:15:47 info [collapse_to_genes]: keeping records with >= 100% overlap ratio with a gene
#> <gcap> 2022-03-30 17:15:47 info [gcap.collapse2Genes]: merging gene-level and sample-level data
#> <gcap> 2022-03-30 17:15:47 info [gcap.collapse2Genes]: merging data and extra info
#> <gcap> 2022-03-30 17:15:47 info [gcap.collapse2Genes]: merging data and prior amplicon frequency data
#> <gcap> 2022-03-30 17:15:47 info [gcap.collapse2Genes]: done
#> <gcap> 2022-03-30 17:15:47 info [gcap.ASCNworkflow]: =======================
#> <gcap> 2022-03-30 17:15:47 info [gcap.ASCNworkflow]: Step 2: Run prediction
#> <gcap> 2022-03-30 17:15:47 info [gcap.ASCNworkflow]: =======================
#> <gcap> 2022-03-30 17:15:47 info [gcap.runPrediction]: using model file XGB_NF32.rds
#> <gcap> 2022-03-30 17:15:47 info [gcap.runPrediction]: selecting necessary features from input data
#> <gcap> 2022-03-30 17:15:47 info [gcap.runPrediction]: running prediction
#> [17:15:47] WARNING: amalgamation/../src/c_api/c_api.cc:718: `ntree_limit` is deprecated, use `iteration_range` instead.
#> <gcap> 2022-03-30 17:15:47 info [gcap.ASCNworkflow]: ====================================
#> <gcap> 2022-03-30 17:15:47 info [gcap.ASCNworkflow]: Step 3: Run scoring and summarizing
#> <gcap> 2022-03-30 17:15:47 info [gcap.ASCNworkflow]: ====================================
#> <gcap> 2022-03-30 17:15:47 info [gcap.runScoring]: checking input data type
#> <gcap> 2022-03-30 17:15:47 info [gcap.runScoring]: checking columns
#> <gcap> 2022-03-30 17:15:47 info [gcap.runScoring]: filtering out records without prob result
#> <gcap> 2022-03-30 17:15:47 info [gcap.runScoring]: joining extra annotation data
#> <gcap> 2022-03-30 17:15:50 info [gcap.runScoring]: summarizing in gene level
#> <gcap> 2022-03-30 17:15:53 info [gcap.runScoring]: summarizing in sample level
#> <gcap> 2022-03-30 17:15:53 info [gcap.runScoring]: done
#> <gcap> 2022-03-30 17:15:53 info [gcap.ASCNworkflow]: Saving raw prediction result to C:\Users\ADMINI~1\AppData\Local\Temp\Rtmpuer5z4/gcap_ffc11c5c-b009-11ec-8000-132265b5ee3a_prediction_result.rds
#> <gcap> 2022-03-30 17:15:53 info [gcap.ASCNworkflow]: Saving gene result to C:\Users\ADMINI~1\AppData\Local\Temp\Rtmpuer5z4/gcap_ffc11c5c-b009-11ec-8000-132265b5ee3a_by_gene.csv
#> <gcap> 2022-03-30 17:15:53 info [gcap.ASCNworkflow]: Saving sample result to C:\Users\ADMINI~1\AppData\Local\Temp\Rtmpuer5z4/gcap_ffc11c5c-b009-11ec-8000-132265b5ee3a_by_sample.csv
#> <gcap> 2022-03-30 17:15:53 info [gcap.ASCNworkflow]: ===========================================
#> <gcap> 2022-03-30 17:15:53 info [gcap.ASCNworkflow]:  Done! Thanks for using GCAP ASCN workflow
#> <gcap> 2022-03-30 17:15:53 info [gcap.ASCNworkflow]: ===========================================
# If you want to use 'XGB56', you should include 'type' column
data$type <- "LUAD"
rv4 <- gcap.ASCNworkflow(data, outdir = tempdir(), model = "XGB56")
#> <gcap> 2022-03-30 17:15:53 info [gcap.ASCNworkflow]: =========================
#> <gcap> 2022-03-30 17:15:53 info [gcap.ASCNworkflow]:    GCAP ASCN WORKFLOW    
#> <gcap> 2022-03-30 17:15:53 info [gcap.ASCNworkflow]: =========================
#> 
#> <gcap> 2022-03-30 17:15:53 info [gcap.ASCNworkflow]: ============================================================
#> <gcap> 2022-03-30 17:15:53 info [gcap.ASCNworkflow]: Step 1: Extract features and collapse features to gene level
#> <gcap> 2022-03-30 17:15:53 info [gcap.ASCNworkflow]: ============================================================
#> <gcap> 2022-03-30 17:15:53 info [gcap.runASCATBuildflow]: extracting sample-level and region-level features
#> <gcap> 2022-03-30 17:15:53 info [gcap.extractFeatures]: > Extract features from input ascn_data<
#> <gcap> 2022-03-30 17:15:53 info [gcap.extractFeatures]: generating CopyNumber object in sigminer package
#> i [2022-03-30 17:15:53]: Started.
#> i [2022-03-30 17:15:53]: Genome build  : hg38.
#> i [2022-03-30 17:15:53]: Genome measure: called.
#> i [2022-03-30 17:15:53]: When add_loh is TRUE, use_all is forced to TRUE.
#> Please drop columns you don't want to keep before reading.
#> v [2022-03-30 17:15:53]: Chromosome size database for build obtained.
#> i [2022-03-30 17:15:53]: Reading input.
#> v [2022-03-30 17:15:53]: A data frame as input detected.
#> v [2022-03-30 17:15:53]: Column names checked.
#> v [2022-03-30 17:15:53]: Column order set.
#> v [2022-03-30 17:15:53]: Chromosomes unified.
#> v [2022-03-30 17:15:53]: Data imported.
#> i [2022-03-30 17:15:53]: Segments info:
#> i [2022-03-30 17:15:53]:     Keep - 129
#> i [2022-03-30 17:15:53]:     Drop - 0
#> v [2022-03-30 17:15:53]: Segments sorted.
#> i [2022-03-30 17:15:53]: Adding LOH labels...
#> i [2022-03-30 17:15:53]: Skipped joining adjacent segments with same copy number value.
#> v [2022-03-30 17:15:53]: Segmental table cleaned.
#> i [2022-03-30 17:15:53]: Annotating.
#> v [2022-03-30 17:15:53]: Annotation done.
#> i [2022-03-30 17:15:53]: Summarizing per sample.
#> v [2022-03-30 17:15:53]: Summarized.
#> i [2022-03-30 17:15:53]: Generating CopyNumber object.
#> v [2022-03-30 17:15:53]: Generated.
#> i [2022-03-30 17:15:53]: Validating object.
#> v [2022-03-30 17:15:53]: Done.
#> i [2022-03-30 17:15:53]: 0.084 secs elapsed.
#> <gcap> 2022-03-30 17:15:53 info [gcap.extractFeatures]: estimating ploidy from copy number data
#> <gcap> 2022-03-30 17:15:53 info [gcap.extractFeatures]: checking if input data contains ploidy and if there are NAs should be overwritten
#> <gcap> 2022-03-30 17:15:53 info [gcap.extractFeatures]: getting Aneuploidy score
#> <gcap> 2022-03-30 17:15:53 info [gcap.extractFeatures]: getting pLOH score
#> <gcap> 2022-03-30 17:15:53 info [gcap.extractFeatures]: getting CNA burden
#> <gcap> 2022-03-30 17:15:53 info [gcap.extractFeatures]: generating copy number catalog matrix for fitting signature activity
#> i [2022-03-30 17:15:53]: Started.
#> i [2022-03-30 17:15:53]: When you use method 'S', please make sure you have set 'join_adj_seg' to FALSE and 'add_loh' to TRUE in 'read_copynumber() in the previous step!
#> v [2022-03-30 17:15:53]: Matrix generated.
#> i [2022-03-30 17:15:53]: 0.02 secs elapsed.
#> <gcap> 2022-03-30 17:15:53 info [gcap.extractFeatures]: fitting copy number signature activity
#> i [2022-03-30 17:15:53]: Started.
#> v [2022-03-30 17:15:53]: Signature index detected.
#> i [2022-03-30 17:15:53]: Checking signature database in package.
#> i [2022-03-30 17:15:53]: Checking signature index.
#> i [2022-03-30 17:15:53]: Valid index for db 'CNS_TCGA':
#> CN1 CN2 CN3 CN4 CN5 CN6 CN7 CN8 CN9 CN10 CN11 CN12 CN13 CN14 CN15 CN16 CN17 CN18 CN19
#> v [2022-03-30 17:15:53]: Database and index checked.
#> v [2022-03-30 17:15:53]: Signature normalized.
#> i [2022-03-30 17:15:53]: Checking row number for catalog matrix and signature matrix.
#> v [2022-03-30 17:15:53]: Checked.
#> i [2022-03-30 17:15:53]: Checking rownames for catalog matrix and signature matrix.
#> v [2022-03-30 17:15:53]: Checked.
#> v [2022-03-30 17:15:53]: Method 'QP' detected.
#> v [2022-03-30 17:15:53]: Corresponding function generated.
#> i [2022-03-30 17:15:53]: Calling function.
#> i [2022-03-30 17:15:53]: Fitting sample: TCGA-02-2485-01
#> i [2022-03-30 17:15:53]: Fitting sample: TCGA-05-4422-01
#> v [2022-03-30 17:15:53]: Done.
#> i [2022-03-30 17:15:53]: Generating output signature exposures.
#> v [2022-03-30 17:15:53]: Done.
#> i [2022-03-30 17:15:53]: 0.056 secs elapsed.
#> <gcap> 2022-03-30 17:15:53 info [gcap.extractFeatures]: merging data
#> <gcap> 2022-03-30 17:15:53 info [gcap.extractFeatures]: feature extraction done
#> <gcap> 2022-03-30 17:15:53 info [gcap.extractFeatures]: now you can modify the result and append 'age' and 'gender' columns to the 'fts_sample' element of result list
#> <gcap> 2022-03-30 17:15:53 info [gcap.runASCATBuildflow]: checking if input age, gender (and type) columns
#> <gcap> 2022-03-30 17:15:53 info [gcap.runASCATBuildflow]: collapsing all data into gene-level prediction input
#> <gcap> 2022-03-30 17:15:53 info [gcap.collapse2Genes]: please make sure the first 3 columns of `fts$fts_region` are for chr, start, end.
#> <gcap> 2022-03-30 17:15:53 info [gcap.collapse2Genes]: one-hot encoding cancer type
#> <gcap> 2022-03-30 17:15:53 info [gcap.collapse2Genes]: valid cancer types: BLCA,BRCA,CESC,COAD,DLBC,ESCA,GBM,HNSC,KICH,KIRC,KIRP,LGG,LIHC,LUAD,LUSC,OV,PRAD,READ,SARC,SKCM,STAD,THCA,UCEC,UVM
#> <gcap> 2022-03-30 17:15:53 info [gcap.collapse2Genes]: checking data type for age column
#> <gcap> 2022-03-30 17:15:53 info [gcap.collapse2Genes]: checking data type for gender column
#> <gcap> 2022-03-30 17:15:53 warn [gcap.collapse2Genes]: non numeric type found, transforming XY to 1, otherwise 0
#> <gcap> 2022-03-30 17:15:53 info [gcap.collapse2Genes]: collapsing region-level features to gene-level
#> <gcap> 2022-03-30 17:15:53 info [collapse_to_genes]: checking input chromosome names
#> <gcap> 2022-03-30 17:15:53 info [collapse_to_genes]: reading reference file D:/Rlib/gcap/extdata/hg38_target_genes.rds
#> <gcap> 2022-03-30 17:15:53 info [collapse_to_genes]: finding overlaps
#> <gcap> 2022-03-30 17:15:53 info [collapse_to_genes]: calculating intersect size
#> <gcap> 2022-03-30 17:15:53 info [collapse_to_genes]: keeping records with >= 100% overlap ratio with a gene
#> <gcap> 2022-03-30 17:15:53 info [gcap.collapse2Genes]: merging gene-level and sample-level data
#> <gcap> 2022-03-30 17:15:53 info [gcap.collapse2Genes]: merging data and extra info
#> <gcap> 2022-03-30 17:15:53 info [gcap.collapse2Genes]: merging data and prior amplicon frequency data
#> <gcap> 2022-03-30 17:15:53 info [gcap.collapse2Genes]: done
#> <gcap> 2022-03-30 17:15:53 info [gcap.ASCNworkflow]: =======================
#> <gcap> 2022-03-30 17:15:53 info [gcap.ASCNworkflow]: Step 2: Run prediction
#> <gcap> 2022-03-30 17:15:53 info [gcap.ASCNworkflow]: =======================
#> <gcap> 2022-03-30 17:15:53 info [gcap.runPrediction]: using model file XGB_NF56.rds
#> <gcap> 2022-03-30 17:15:53 info [gcap.runPrediction]: selecting necessary features from input data
#> <gcap> 2022-03-30 17:15:53 info [gcap.runPrediction]: running prediction
#> [17:15:53] WARNING: amalgamation/../src/c_api/c_api.cc:718: `ntree_limit` is deprecated, use `iteration_range` instead.
#> <gcap> 2022-03-30 17:15:53 info [gcap.ASCNworkflow]: ====================================
#> <gcap> 2022-03-30 17:15:53 info [gcap.ASCNworkflow]: Step 3: Run scoring and summarizing
#> <gcap> 2022-03-30 17:15:53 info [gcap.ASCNworkflow]: ====================================
#> <gcap> 2022-03-30 17:15:53 info [gcap.runScoring]: checking input data type
#> <gcap> 2022-03-30 17:15:53 info [gcap.runScoring]: checking columns
#> <gcap> 2022-03-30 17:15:53 info [gcap.runScoring]: filtering out records without prob result
#> <gcap> 2022-03-30 17:15:53 info [gcap.runScoring]: joining extra annotation data
#> <gcap> 2022-03-30 17:15:56 info [gcap.runScoring]: summarizing in gene level
#> <gcap> 2022-03-30 17:16:00 info [gcap.runScoring]: summarizing in sample level
#> <gcap> 2022-03-30 17:16:00 info [gcap.runScoring]: done
#> <gcap> 2022-03-30 17:16:00 info [gcap.ASCNworkflow]: Saving raw prediction result to C:\Users\ADMINI~1\AppData\Local\Temp\Rtmpuer5z4/gcap_03eab5cc-b00a-11ec-8000-132265b5ee3a_prediction_result.rds
#> <gcap> 2022-03-30 17:16:00 info [gcap.ASCNworkflow]: Saving gene result to C:\Users\ADMINI~1\AppData\Local\Temp\Rtmpuer5z4/gcap_03eab5cc-b00a-11ec-8000-132265b5ee3a_by_gene.csv
#> <gcap> 2022-03-30 17:16:00 info [gcap.ASCNworkflow]: Saving sample result to C:\Users\ADMINI~1\AppData\Local\Temp\Rtmpuer5z4/gcap_03eab5cc-b00a-11ec-8000-132265b5ee3a_by_sample.csv
#> <gcap> 2022-03-30 17:16:00 info [gcap.ASCNworkflow]: ===========================================
#> <gcap> 2022-03-30 17:16:00 info [gcap.ASCNworkflow]:  Done! Thanks for using GCAP ASCN workflow
#> <gcap> 2022-03-30 17:16:00 info [gcap.ASCNworkflow]: ===========================================
# If you only have total integer copy number
data$minor_cn <- NA
rv5 <- gcap.ASCNworkflow(data, outdir = tempdir(), model = "XGB11")
#> <gcap> 2022-03-30 17:16:00 info [gcap.ASCNworkflow]: =========================
#> <gcap> 2022-03-30 17:16:00 info [gcap.ASCNworkflow]:    GCAP ASCN WORKFLOW    
#> <gcap> 2022-03-30 17:16:00 info [gcap.ASCNworkflow]: =========================
#> 
#> <gcap> 2022-03-30 17:16:00 info [gcap.ASCNworkflow]: ============================================================
#> <gcap> 2022-03-30 17:16:00 info [gcap.ASCNworkflow]: Step 1: Extract features and collapse features to gene level
#> <gcap> 2022-03-30 17:16:00 info [gcap.ASCNworkflow]: ============================================================
#> <gcap> 2022-03-30 17:16:00 info [gcap.runASCATBuildflow]: extracting sample-level and region-level features
#> <gcap> 2022-03-30 17:16:00 info [gcap.extractFeatures]: > Extract features from input ascn_data<
#> <gcap> 2022-03-30 17:16:00 info [gcap.extractFeatures]: generating CopyNumber object in sigminer package
#> i [2022-03-30 17:16:00]: Started.
#> i [2022-03-30 17:16:00]: Genome build  : hg38.
#> i [2022-03-30 17:16:00]: Genome measure: called.
#> v [2022-03-30 17:16:00]: Chromosome size database for build obtained.
#> i [2022-03-30 17:16:00]: Reading input.
#> v [2022-03-30 17:16:00]: A data frame as input detected.
#> v [2022-03-30 17:16:00]: Column names checked.
#> v [2022-03-30 17:16:00]: Column order set.
#> v [2022-03-30 17:16:00]: Chromosomes unified.
#> v [2022-03-30 17:16:00]: Data imported.
#> i [2022-03-30 17:16:00]: Segments info:
#> i [2022-03-30 17:16:00]:     Keep - 129
#> i [2022-03-30 17:16:00]:     Drop - 0
#> v [2022-03-30 17:16:00]: Segments sorted.
#> i [2022-03-30 17:16:00]: Skipped joining adjacent segments with same copy number value.
#> v [2022-03-30 17:16:00]: Segmental table cleaned.
#> i [2022-03-30 17:16:00]: Annotating.
#> v [2022-03-30 17:16:00]: Annotation done.
#> i [2022-03-30 17:16:00]: Summarizing per sample.
#> v [2022-03-30 17:16:00]: Summarized.
#> i [2022-03-30 17:16:00]: Generating CopyNumber object.
#> v [2022-03-30 17:16:00]: Generated.
#> i [2022-03-30 17:16:00]: Validating object.
#> v [2022-03-30 17:16:00]: Done.
#> i [2022-03-30 17:16:00]: 0.087 secs elapsed.
#> <gcap> 2022-03-30 17:16:00 info [gcap.extractFeatures]: estimating ploidy from copy number data
#> <gcap> 2022-03-30 17:16:00 info [gcap.extractFeatures]: checking if input data contains ploidy and if there are NAs should be overwritten
#> <gcap> 2022-03-30 17:16:00 info [gcap.extractFeatures]: getting Aneuploidy score
#> <gcap> 2022-03-30 17:16:00 info [gcap.extractFeatures]: pLOH score is skipped as no LOH data available
#> <gcap> 2022-03-30 17:16:00 info [gcap.extractFeatures]: getting CNA burden
#> <gcap> 2022-03-30 17:16:00 info [gcap.extractFeatures]: copy number signature activity quantification is skipped as no LOH data available
#> <gcap> 2022-03-30 17:16:00 info [gcap.extractFeatures]: merging data
#> <gcap> 2022-03-30 17:16:00 info [gcap.extractFeatures]: feature extraction done
#> <gcap> 2022-03-30 17:16:00 info [gcap.extractFeatures]: now you can modify the result and append 'age' and 'gender' columns to the 'fts_sample' element of result list
#> <gcap> 2022-03-30 17:16:00 info [gcap.runASCATBuildflow]: checking if input age, gender (and type) columns
#> <gcap> 2022-03-30 17:16:00 info [gcap.runASCATBuildflow]: collapsing all data into gene-level prediction input
#> <gcap> 2022-03-30 17:16:00 info [gcap.collapse2Genes]: please make sure the first 3 columns of `fts$fts_region` are for chr, start, end.
#> <gcap> 2022-03-30 17:16:00 info [gcap.collapse2Genes]: one-hot encoding cancer type
#> <gcap> 2022-03-30 17:16:00 info [gcap.collapse2Genes]: valid cancer types: BLCA,BRCA,CESC,COAD,DLBC,ESCA,GBM,HNSC,KICH,KIRC,KIRP,LGG,LIHC,LUAD,LUSC,OV,PRAD,READ,SARC,SKCM,STAD,THCA,UCEC,UVM
#> <gcap> 2022-03-30 17:16:00 info [gcap.collapse2Genes]: checking data type for age column
#> <gcap> 2022-03-30 17:16:00 info [gcap.collapse2Genes]: checking data type for gender column
#> <gcap> 2022-03-30 17:16:00 warn [gcap.collapse2Genes]: non numeric type found, transforming XY to 1, otherwise 0
#> <gcap> 2022-03-30 17:16:00 info [gcap.collapse2Genes]: collapsing region-level features to gene-level
#> <gcap> 2022-03-30 17:16:00 info [collapse_to_genes]: checking input chromosome names
#> <gcap> 2022-03-30 17:16:00 info [collapse_to_genes]: reading reference file D:/Rlib/gcap/extdata/hg38_target_genes.rds
#> <gcap> 2022-03-30 17:16:00 info [collapse_to_genes]: finding overlaps
#> <gcap> 2022-03-30 17:16:00 info [collapse_to_genes]: calculating intersect size
#> <gcap> 2022-03-30 17:16:00 info [collapse_to_genes]: keeping records with >= 100% overlap ratio with a gene
#> <gcap> 2022-03-30 17:16:00 info [gcap.collapse2Genes]: merging gene-level and sample-level data
#> <gcap> 2022-03-30 17:16:00 info [gcap.collapse2Genes]: merging data and extra info
#> <gcap> 2022-03-30 17:16:00 info [gcap.collapse2Genes]: merging data and prior amplicon frequency data
#> <gcap> 2022-03-30 17:16:00 info [gcap.collapse2Genes]: done
#> <gcap> 2022-03-30 17:16:00 info [gcap.ASCNworkflow]: =======================
#> <gcap> 2022-03-30 17:16:00 info [gcap.ASCNworkflow]: Step 2: Run prediction
#> <gcap> 2022-03-30 17:16:00 info [gcap.ASCNworkflow]: =======================
#> <gcap> 2022-03-30 17:16:00 info [gcap.runPrediction]: using model file XGB_NF11.rds
#> <gcap> 2022-03-30 17:16:00 info [gcap.runPrediction]: selecting necessary features from input data
#> <gcap> 2022-03-30 17:16:00 info [gcap.runPrediction]: running prediction
#> [17:16:00] WARNING: amalgamation/../src/c_api/c_api.cc:718: `ntree_limit` is deprecated, use `iteration_range` instead.
#> <gcap> 2022-03-30 17:16:00 info [gcap.ASCNworkflow]: ====================================
#> <gcap> 2022-03-30 17:16:00 info [gcap.ASCNworkflow]: Step 3: Run scoring and summarizing
#> <gcap> 2022-03-30 17:16:00 info [gcap.ASCNworkflow]: ====================================
#> <gcap> 2022-03-30 17:16:00 info [gcap.runScoring]: checking input data type
#> <gcap> 2022-03-30 17:16:00 info [gcap.runScoring]: checking columns
#> <gcap> 2022-03-30 17:16:00 info [gcap.runScoring]: filtering out records without prob result
#> <gcap> 2022-03-30 17:16:00 info [gcap.runScoring]: joining extra annotation data
#> <gcap> 2022-03-30 17:16:03 info [gcap.runScoring]: summarizing in gene level
#> <gcap> 2022-03-30 17:16:05 info [gcap.runScoring]: summarizing in sample level
#> <gcap> 2022-03-30 17:16:05 info [gcap.runScoring]: done
#> <gcap> 2022-03-30 17:16:06 info [gcap.ASCNworkflow]: Saving raw prediction result to C:\Users\ADMINI~1\AppData\Local\Temp\Rtmpuer5z4/gcap_076b2182-b00a-11ec-8000-132265b5ee3a_prediction_result.rds
#> <gcap> 2022-03-30 17:16:06 info [gcap.ASCNworkflow]: Saving gene result to C:\Users\ADMINI~1\AppData\Local\Temp\Rtmpuer5z4/gcap_076b2182-b00a-11ec-8000-132265b5ee3a_by_gene.csv
#> <gcap> 2022-03-30 17:16:06 info [gcap.ASCNworkflow]: Saving sample result to C:\Users\ADMINI~1\AppData\Local\Temp\Rtmpuer5z4/gcap_076b2182-b00a-11ec-8000-132265b5ee3a_by_sample.csv
#> <gcap> 2022-03-30 17:16:06 info [gcap.ASCNworkflow]: ===========================================
#> <gcap> 2022-03-30 17:16:06 info [gcap.ASCNworkflow]:  Done! Thanks for using GCAP ASCN workflow
#> <gcap> 2022-03-30 17:16:06 info [gcap.ASCNworkflow]: ===========================================

# R6 class fCNA --------------------------------
print(rv)
#> $gene
#>              gene_id amplicon_type N total_cn minor_cn cytoband_cn
#>   1: ENSG00000111087      circular 1       35        2           4
#>   2: ENSG00000123329      circular 1       35        2           4
#>   3: ENSG00000135454      circular 1       35        2           4
#>   4: ENSG00000139269      circular 1       35        2           4
#>   5: ENSG00000155980      circular 1       35        2           4
#>  ---                                                              
#> 342: ENSG00000212993       nofocal 1        4        2           4
#> 343: ENSG00000132294      circular 1       15        1           7
#> 344: ENSG00000155897      circular 1       15        1           7
#> 345: ENSG00000132294       nofocal 1        4        2           4
#> 346: ENSG00000155897       nofocal 1        4        2           4
#>          support_IDs        band
#>   1: TCGA-02-2485-01 chr12:q13.3
#>   2: TCGA-02-2485-01 chr12:q13.3
#>   3: TCGA-02-2485-01 chr12:q13.3
#>   4: TCGA-02-2485-01 chr12:q13.3
#>   5: TCGA-02-2485-01 chr12:q13.3
#>  ---                            
#> 342:            <NA> chr8:q24.21
#> 343: TCGA-05-4422-01 chr8:q24.22
#> 344: TCGA-05-4422-01 chr8:q24.22
#> 345:            <NA> chr8:q24.22
#> 346:            <NA> chr8:q24.22
#> 
#> $sample
#>             sample ec_genes ec_cytobands
#> 1: TCGA-05-4422-01      101            9
#> 2: TCGA-02-2485-01       72            8
#>                                                                                          ec_cytobands_detail
#> 1: chr20:q13.33,chr8:q21.11,chr8:q21.13,chr8:q22.1,chr8:q22.2,chr8:q22.3,chr8:q24.13,chr8:q24.21,chr8:q24.22
#> 2:                       chr12:q13.3,chr12:q14.1,chr1:q43,chr1:q44,chr4:q12,chr7:p11.2,chr7:p12.1,chr7:p12.2
#>    ec_possibly_genes    class purity ploidy AScore       pLOH cna_burden
#> 1:                 0 circular      1   2.09     13 0.13041086      0.249
#> 2:                 0 circular      1   4.09     13 0.04766306      0.951
#>         CN1       CN2 CN3 CN4 CN5 CN6      CN7 CN8    CN9 CN10     CN11 CN12
#> 1: 23.84798  2.326143   0   0   0   0 8.939099   0 29.688    0 1.154688    0
#> 2:  0.00000 45.112591   0   0   0   0 9.887409   0  0.000    0 0.000000    0
#>    CN13 CN14     CN15 CN16 CN17     CN18 CN19
#> 1:    0    0 1.178945    0    0 6.865139    0
#> 2:    0    0 0.000000    0    0 0.000000    0
#> 
print(rv$data)
#> NULL
print(rv$sample_summary)
#> NULL
print(rv$gene_summary)
#> NULL
print(rv$cytoband_summary)
#> NULL