GCAP workflow for gene-level amplicon prediction from ASCN input

Unlike gcap.workflow, this function directly uses the allele-specific copy number data along with some extra sample information to infer ecDNA genes.

gcap.ASCNworkflow(
  data,
  genome_build = c("hg38", "hg19"),
  model = "XGB11",
  tightness = 1L,
  gap_cn = 3L,
  overlap = 1,
  only_oncogenes = FALSE,
  outdir = getwd(),
  result_file_prefix = paste0("gcap_", uuid::UUIDgenerate(TRUE))
)

Arguments

data

a data.frame with following columns. The key columns can be obtained from common allele specific CNV calling software, e.g., ASCAT, Sequenza, FACETS.

chromosome: chromosome names starts with 'chr'.

start: start position of the segment.

end: end position of the segment.

total_cn: total integer copy number of the segment.

minor_cn: minor allele integer copy number of the segment. Set it to NA if you don't have this data.

sample: sample identifier.

purity: tumor purity of the sample. Set to 1 if you don't know.

ploidy (optinal): ploidy value of the sample tumor genome.

age (optional): age of the case, use along with gender.

gender (optional): gender of the case, use along with age.

type (optional): cancer type of the case, use along with age and gender. Please refer to gcap.collapse2Genes to see the supported cancer types. This info is only used in 'XGB56' model. If you don't use this model, you don't need to set it.

genome_build

genome build version, should be one of 'hg38', 'hg19'.

model

model name ("XGB11", "XGB32", "XGB56") or a custom model from input. 'toy' can be used for test.

tightness

a coefficient to times to TCGA somatic CN to set a more strict threshold as a circular amplicon. If the value is larger, it is more likely a fCNA assigned to noncircular instead of circular. When it is NA, we don't use TCGA somatic CN data as reference.

gap_cn

a gap copy number value, default is 4L refer to Kim 2020 Nat.Gen. A gene with copy number above ploidy + gap_cn would be treated as focal amplicon. Smaller, more amplicons.

overlap

the overlap percentage on gene.

only_oncogenes

if TRUE, only known oncogenes are kept for fCNA analysis and circular prediction.

outdir

result output path.

result_file_prefix

file name prefix (without directory path) for storing final model prediction file in CSV format. Default a unique file name is generated by UUID approach.

Value

a list of invisible data.table and corresponding files saved to local machine.

Examples

data("ascn")
data <- ascn
rv <- gcap.ASCNworkflow(data, outdir = tempdir(), model = "XGB11")
#> <gcap> 2022-09-19 11:34:39 info [gcap.ASCNworkflow]: =========================
#> <gcap> 2022-09-19 11:34:39 info [gcap.ASCNworkflow]:    GCAP ASCN WORKFLOW    
#> <gcap> 2022-09-19 11:34:39 info [gcap.ASCNworkflow]: =========================
#> 
#> <gcap> 2022-09-19 11:34:39 info [gcap.ASCNworkflow]: ============================================================
#> <gcap> 2022-09-19 11:34:39 info [gcap.ASCNworkflow]: Step 1: Extract features and collapse features to gene level
#> <gcap> 2022-09-19 11:34:39 info [gcap.ASCNworkflow]: ============================================================
#> <gcap> 2022-09-19 11:34:39 info [gcap.runASCATBuildflow]: extracting sample-level and region-level features
#> <gcap> 2022-09-19 11:34:39 info [gcap.extractFeatures]: > Extract features from input ascn_data<
#> <gcap> 2022-09-19 11:34:39 info [gcap.extractFeatures]: generating CopyNumber object in sigminer package
#> ℹ [2022-09-19 11:34:39]: Started.
#> ℹ [2022-09-19 11:34:39]: Genome build  : hg38.
#> ℹ [2022-09-19 11:34:39]: Genome measure: called.
#> ℹ [2022-09-19 11:34:39]: When add_loh is TRUE, use_all is forced to TRUE.
#> Please drop columns you don't want to keep before reading.
#> ✔ [2022-09-19 11:34:39]: Chromosome size database for build obtained.
#> ℹ [2022-09-19 11:34:39]: Reading input.
#> ✔ [2022-09-19 11:34:39]: A data frame as input detected.
#> ✔ [2022-09-19 11:34:39]: Column names checked.
#> ✔ [2022-09-19 11:34:39]: Column order set.
#> ✔ [2022-09-19 11:34:40]: Chromosomes unified.
#> ✔ [2022-09-19 11:34:40]: Data imported.
#> ℹ [2022-09-19 11:34:40]: Segments info:
#> ℹ [2022-09-19 11:34:40]:     Keep - 129
#> ℹ [2022-09-19 11:34:40]:     Drop - 0
#> ✔ [2022-09-19 11:34:40]: Segments sorted.
#> ℹ [2022-09-19 11:34:40]: Adding LOH labels...
#> ℹ [2022-09-19 11:34:40]: Skipped joining adjacent segments with same copy number value.
#> ✔ [2022-09-19 11:34:40]: Segmental table cleaned.
#> ℹ [2022-09-19 11:34:40]: Annotating.
#> ✔ [2022-09-19 11:34:40]: Annotation done.
#> ℹ [2022-09-19 11:34:40]: Summarizing per sample.
#> ✔ [2022-09-19 11:34:40]: Summarized.
#> ℹ [2022-09-19 11:34:40]: Generating CopyNumber object.
#> ✔ [2022-09-19 11:34:40]: Generated.
#> ℹ [2022-09-19 11:34:40]: Validating object.
#> ✔ [2022-09-19 11:34:40]: Done.
#> ℹ [2022-09-19 11:34:40]: 0.142 secs elapsed.
#> <gcap> 2022-09-19 11:34:40 info [gcap.extractFeatures]: estimating ploidy from copy number data
#> <gcap> 2022-09-19 11:34:40 info [gcap.extractFeatures]: checking if input data contains ploidy and if there are NAs should be overwritten
#> <gcap> 2022-09-19 11:34:40 info [gcap.extractFeatures]: getting Aneuploidy score
#> <gcap> 2022-09-19 11:34:40 info [gcap.extractFeatures]: getting pLOH score
#> <gcap> 2022-09-19 11:34:40 info [gcap.extractFeatures]: getting CNA burden
#> <gcap> 2022-09-19 11:34:40 info [gcap.extractFeatures]: generating copy number catalog matrix for fitting signature activity
#> ℹ [2022-09-19 11:34:40]: Started.
#> ℹ [2022-09-19 11:34:40]: When you use method 'S', please make sure you have set 'join_adj_seg' to FALSE and 'add_loh' to TRUE in 'read_copynumber() in the previous step!
#> ✔ [2022-09-19 11:34:40]: Matrix generated.
#> ℹ [2022-09-19 11:34:40]: 0.025 secs elapsed.
#> <gcap> 2022-09-19 11:34:40 info [gcap.extractFeatures]: fitting copy number signature activity
#> ℹ [2022-09-19 11:34:40]: Started.
#> ✔ [2022-09-19 11:34:40]: Signature index detected.
#> ℹ [2022-09-19 11:34:40]: Checking signature database in package.
#> ℹ [2022-09-19 11:34:40]: Checking signature index.
#> ℹ [2022-09-19 11:34:40]: Valid index for db 'CNS_TCGA':
#> CN1 CN2 CN3 CN4 CN5 CN6 CN7 CN8 CN9 CN10 CN11 CN12 CN13 CN14 CN15 CN16 CN17 CN18 CN19
#> ✔ [2022-09-19 11:34:40]: Database and index checked.
#> ✔ [2022-09-19 11:34:40]: Signature normalized.
#> ℹ [2022-09-19 11:34:40]: Checking row number for catalog matrix and signature matrix.
#> ✔ [2022-09-19 11:34:40]: Checked.
#> ℹ [2022-09-19 11:34:40]: Checking rownames for catalog matrix and signature matrix.
#> ✔ [2022-09-19 11:34:40]: Checked.
#> ✔ [2022-09-19 11:34:40]: Method 'QP' detected.
#> ✔ [2022-09-19 11:34:40]: Corresponding function generated.
#> ℹ [2022-09-19 11:34:40]: Calling function.
#> ℹ [2022-09-19 11:34:40]: Fitting sample: TCGA-02-2485-01
#> ℹ [2022-09-19 11:34:40]: Fitting sample: TCGA-05-4422-01
#> ✔ [2022-09-19 11:34:40]: Done.
#> ℹ [2022-09-19 11:34:40]: Generating output signature exposures.
#> ✔ [2022-09-19 11:34:40]: Done.
#> ℹ [2022-09-19 11:34:40]: 0.266 secs elapsed.
#> <gcap> 2022-09-19 11:34:40 info [gcap.extractFeatures]: merging data
#> <gcap> 2022-09-19 11:34:40 info [gcap.extractFeatures]: feature extraction done
#> <gcap> 2022-09-19 11:34:40 info [gcap.extractFeatures]: now you can modify the result and append 'age' and 'gender' columns to the 'fts_sample' element of result list
#> <gcap> 2022-09-19 11:34:40 info [gcap.runASCATBuildflow]: checking if input age, gender (and type) columns
#> <gcap> 2022-09-19 11:34:40 info [gcap.runASCATBuildflow]: None exist, set to NULL.
#> <gcap> 2022-09-19 11:34:40 info [gcap.runASCATBuildflow]: collapsing all data into gene-level prediction input
#> <gcap> 2022-09-19 11:34:40 info [gcap.collapse2Genes]: please make sure the first 3 columns of `fts$fts_region` are for chr, start, end.
#> <gcap> 2022-09-19 11:34:40 info [gcap.collapse2Genes]: collapsing region-level features to gene-level
#> <gcap> 2022-09-19 11:34:40 info [collapse_to_genes]: checking input chromosome names
#> <gcap> 2022-09-19 11:34:40 info [collapse_to_genes]: reading reference file D:/Rlib/gcap/extdata/hg38_target_genes.rds
#> <gcap> 2022-09-19 11:34:40 info [collapse_to_genes]: finding overlaps
#> <gcap> 2022-09-19 11:34:40 info [collapse_to_genes]: calculating intersect size
#> <gcap> 2022-09-19 11:34:40 info [collapse_to_genes]: keeping records with >= 100% overlap ratio with a gene
#> <gcap> 2022-09-19 11:34:40 info [gcap.collapse2Genes]: merging gene-level and sample-level data
#> <gcap> 2022-09-19 11:34:40 info [gcap.collapse2Genes]: merging data and prior amplicon frequency data
#> <gcap> 2022-09-19 11:34:40 info [gcap.collapse2Genes]: done
#> <gcap> 2022-09-19 11:34:40 info [gcap.ASCNworkflow]: =======================
#> <gcap> 2022-09-19 11:34:40 info [gcap.ASCNworkflow]: Step 2: Run prediction
#> <gcap> 2022-09-19 11:34:40 info [gcap.ASCNworkflow]: =======================
#> <gcap> 2022-09-19 11:34:40 info [gcap.runPrediction]: using model file XGB_NF11.rds
#> <gcap> 2022-09-19 11:34:40 info [gcap.runPrediction]: selecting necessary features from input data
#> <gcap> 2022-09-19 11:34:40 info [gcap.runPrediction]: running prediction
#> [11:34:40] WARNING: amalgamation/../src/c_api/c_api.cc:718: `ntree_limit` is deprecated, use `iteration_range` instead.
#> <gcap> 2022-09-19 11:34:40 info [gcap.ASCNworkflow]: ====================================
#> <gcap> 2022-09-19 11:34:40 info [gcap.ASCNworkflow]: Step 3: Run scoring and summarizing
#> <gcap> 2022-09-19 11:34:40 info [gcap.ASCNworkflow]: ====================================
#> <gcap> 2022-09-19 11:34:40 info [gcap.runScoring]: checking input data type
#> <gcap> 2022-09-19 11:34:40 info [gcap.runScoring]: checking columns
#> <gcap> 2022-09-19 11:34:40 info [gcap.runScoring]: filtering out records without prob result
#> <gcap> 2022-09-19 11:34:40 info [gcap.runScoring]: joining extra annotation data
#> <gcap> 2022-09-19 11:34:44 info [gcap.runScoring]: only keep genes with copy number >= ploidy+2 in result fCNA object
#> summarizing sample...
#>   classifying samples with min_n=1
#> done
#> ======================
#> A <fCNA> object
#>   record: 255
#>     case: 2
#>      |__ (0) noncircular
#>      |__ (0) possibly_circular
#>      |__ (2) circular
#> ======================
#> <gcap> 2022-09-19 11:34:44 info [gcap.runScoring]: done
#> <gcap> 2022-09-19 11:34:44 info [gcap.ASCNworkflow]: Saving raw prediction result to C:\Users\ADMINI~1\AppData\Local\Temp\RtmpKEtn9x/gcap_00abd8d8-37cc-11ed-8000-03c38a3f6043_prediction_result.rds
#> <gcap> 2022-09-19 11:34:44 info [gcap.ASCNworkflow]: Saving fCNA records and sample info to C:\Users\ADMINI~1\AppData\Local\Temp\RtmpKEtn9x/gcap_00abd8d8-37cc-11ed-8000-03c38a3f6043_fCNA_records.csv, C:\Users\ADMINI~1\AppData\Local\Temp\RtmpKEtn9x/gcap_00abd8d8-37cc-11ed-8000-03c38a3f6043_sample_info.csv
#> <gcap> 2022-09-19 11:34:44 info [gcap.ASCNworkflow]: ===========================================
#> <gcap> 2022-09-19 11:34:44 info [gcap.ASCNworkflow]:  Done! Thanks for using GCAP ASCN workflow
#> <gcap> 2022-09-19 11:34:44 info [gcap.ASCNworkflow]: ===========================================
data$purity <- 1
rv2 <- gcap.ASCNworkflow(data, outdir = tempdir(), model = "XGB11")
#> <gcap> 2022-09-19 11:34:44 info [gcap.ASCNworkflow]: =========================
#> <gcap> 2022-09-19 11:34:44 info [gcap.ASCNworkflow]:    GCAP ASCN WORKFLOW    
#> <gcap> 2022-09-19 11:34:44 info [gcap.ASCNworkflow]: =========================
#> 
#> <gcap> 2022-09-19 11:34:44 info [gcap.ASCNworkflow]: ============================================================
#> <gcap> 2022-09-19 11:34:44 info [gcap.ASCNworkflow]: Step 1: Extract features and collapse features to gene level
#> <gcap> 2022-09-19 11:34:44 info [gcap.ASCNworkflow]: ============================================================
#> <gcap> 2022-09-19 11:34:44 info [gcap.runASCATBuildflow]: extracting sample-level and region-level features
#> <gcap> 2022-09-19 11:34:44 info [gcap.extractFeatures]: > Extract features from input ascn_data<
#> <gcap> 2022-09-19 11:34:44 info [gcap.extractFeatures]: generating CopyNumber object in sigminer package
#> ℹ [2022-09-19 11:34:44]: Started.
#> ℹ [2022-09-19 11:34:44]: Genome build  : hg38.
#> ℹ [2022-09-19 11:34:44]: Genome measure: called.
#> ℹ [2022-09-19 11:34:44]: When add_loh is TRUE, use_all is forced to TRUE.
#> Please drop columns you don't want to keep before reading.
#> ✔ [2022-09-19 11:34:44]: Chromosome size database for build obtained.
#> ℹ [2022-09-19 11:34:44]: Reading input.
#> ✔ [2022-09-19 11:34:44]: A data frame as input detected.
#> ✔ [2022-09-19 11:34:44]: Column names checked.
#> ✔ [2022-09-19 11:34:44]: Column order set.
#> ✔ [2022-09-19 11:34:44]: Chromosomes unified.
#> ✔ [2022-09-19 11:34:44]: Data imported.
#> ℹ [2022-09-19 11:34:44]: Segments info:
#> ℹ [2022-09-19 11:34:44]:     Keep - 129
#> ℹ [2022-09-19 11:34:44]:     Drop - 0
#> ✔ [2022-09-19 11:34:44]: Segments sorted.
#> ℹ [2022-09-19 11:34:44]: Adding LOH labels...
#> ℹ [2022-09-19 11:34:44]: Skipped joining adjacent segments with same copy number value.
#> ✔ [2022-09-19 11:34:44]: Segmental table cleaned.
#> ℹ [2022-09-19 11:34:44]: Annotating.
#> ✔ [2022-09-19 11:34:44]: Annotation done.
#> ℹ [2022-09-19 11:34:44]: Summarizing per sample.
#> ✔ [2022-09-19 11:34:44]: Summarized.
#> ℹ [2022-09-19 11:34:44]: Generating CopyNumber object.
#> ✔ [2022-09-19 11:34:44]: Generated.
#> ℹ [2022-09-19 11:34:44]: Validating object.
#> ✔ [2022-09-19 11:34:44]: Done.
#> ℹ [2022-09-19 11:34:44]: 0.092 secs elapsed.
#> <gcap> 2022-09-19 11:34:44 info [gcap.extractFeatures]: estimating ploidy from copy number data
#> <gcap> 2022-09-19 11:34:44 info [gcap.extractFeatures]: checking if input data contains ploidy and if there are NAs should be overwritten
#> <gcap> 2022-09-19 11:34:44 info [gcap.extractFeatures]: getting Aneuploidy score
#> <gcap> 2022-09-19 11:34:44 info [gcap.extractFeatures]: getting pLOH score
#> <gcap> 2022-09-19 11:34:44 info [gcap.extractFeatures]: getting CNA burden
#> <gcap> 2022-09-19 11:34:44 info [gcap.extractFeatures]: generating copy number catalog matrix for fitting signature activity
#> ℹ [2022-09-19 11:34:44]: Started.
#> ℹ [2022-09-19 11:34:44]: When you use method 'S', please make sure you have set 'join_adj_seg' to FALSE and 'add_loh' to TRUE in 'read_copynumber() in the previous step!
#> ✔ [2022-09-19 11:34:44]: Matrix generated.
#> ℹ [2022-09-19 11:34:44]: 0.021 secs elapsed.
#> <gcap> 2022-09-19 11:34:44 info [gcap.extractFeatures]: fitting copy number signature activity
#> ℹ [2022-09-19 11:34:44]: Started.
#> ✔ [2022-09-19 11:34:44]: Signature index detected.
#> ℹ [2022-09-19 11:34:44]: Checking signature database in package.
#> ℹ [2022-09-19 11:34:44]: Checking signature index.
#> ℹ [2022-09-19 11:34:44]: Valid index for db 'CNS_TCGA':
#> CN1 CN2 CN3 CN4 CN5 CN6 CN7 CN8 CN9 CN10 CN11 CN12 CN13 CN14 CN15 CN16 CN17 CN18 CN19
#> ✔ [2022-09-19 11:34:44]: Database and index checked.
#> ✔ [2022-09-19 11:34:44]: Signature normalized.
#> ℹ [2022-09-19 11:34:44]: Checking row number for catalog matrix and signature matrix.
#> ✔ [2022-09-19 11:34:44]: Checked.
#> ℹ [2022-09-19 11:34:44]: Checking rownames for catalog matrix and signature matrix.
#> ✔ [2022-09-19 11:34:44]: Checked.
#> ✔ [2022-09-19 11:34:44]: Method 'QP' detected.
#> ✔ [2022-09-19 11:34:44]: Corresponding function generated.
#> ℹ [2022-09-19 11:34:44]: Calling function.
#> ℹ [2022-09-19 11:34:44]: Fitting sample: TCGA-02-2485-01
#> ℹ [2022-09-19 11:34:44]: Fitting sample: TCGA-05-4422-01
#> ✔ [2022-09-19 11:34:44]: Done.
#> ℹ [2022-09-19 11:34:44]: Generating output signature exposures.
#> ✔ [2022-09-19 11:34:44]: Done.
#> ℹ [2022-09-19 11:34:44]: 0.044 secs elapsed.
#> <gcap> 2022-09-19 11:34:44 info [gcap.extractFeatures]: merging data
#> <gcap> 2022-09-19 11:34:44 info [gcap.extractFeatures]: feature extraction done
#> <gcap> 2022-09-19 11:34:44 info [gcap.extractFeatures]: now you can modify the result and append 'age' and 'gender' columns to the 'fts_sample' element of result list
#> <gcap> 2022-09-19 11:34:44 info [gcap.runASCATBuildflow]: checking if input age, gender (and type) columns
#> <gcap> 2022-09-19 11:34:44 info [gcap.runASCATBuildflow]: None exist, set to NULL.
#> <gcap> 2022-09-19 11:34:44 info [gcap.runASCATBuildflow]: collapsing all data into gene-level prediction input
#> <gcap> 2022-09-19 11:34:44 info [gcap.collapse2Genes]: please make sure the first 3 columns of `fts$fts_region` are for chr, start, end.
#> <gcap> 2022-09-19 11:34:44 info [gcap.collapse2Genes]: collapsing region-level features to gene-level
#> <gcap> 2022-09-19 11:34:44 info [collapse_to_genes]: checking input chromosome names
#> <gcap> 2022-09-19 11:34:44 info [collapse_to_genes]: reading reference file D:/Rlib/gcap/extdata/hg38_target_genes.rds
#> <gcap> 2022-09-19 11:34:44 info [collapse_to_genes]: finding overlaps
#> <gcap> 2022-09-19 11:34:44 info [collapse_to_genes]: calculating intersect size
#> <gcap> 2022-09-19 11:34:44 info [collapse_to_genes]: keeping records with >= 100% overlap ratio with a gene
#> <gcap> 2022-09-19 11:34:44 info [gcap.collapse2Genes]: merging gene-level and sample-level data
#> <gcap> 2022-09-19 11:34:44 info [gcap.collapse2Genes]: merging data and prior amplicon frequency data
#> <gcap> 2022-09-19 11:34:44 info [gcap.collapse2Genes]: done
#> <gcap> 2022-09-19 11:34:44 info [gcap.ASCNworkflow]: =======================
#> <gcap> 2022-09-19 11:34:44 info [gcap.ASCNworkflow]: Step 2: Run prediction
#> <gcap> 2022-09-19 11:34:44 info [gcap.ASCNworkflow]: =======================
#> <gcap> 2022-09-19 11:34:44 info [gcap.runPrediction]: using model file XGB_NF11.rds
#> <gcap> 2022-09-19 11:34:44 info [gcap.runPrediction]: selecting necessary features from input data
#> <gcap> 2022-09-19 11:34:44 info [gcap.runPrediction]: running prediction
#> [11:34:44] WARNING: amalgamation/../src/c_api/c_api.cc:718: `ntree_limit` is deprecated, use `iteration_range` instead.
#> <gcap> 2022-09-19 11:34:44 info [gcap.ASCNworkflow]: ====================================
#> <gcap> 2022-09-19 11:34:44 info [gcap.ASCNworkflow]: Step 3: Run scoring and summarizing
#> <gcap> 2022-09-19 11:34:44 info [gcap.ASCNworkflow]: ====================================
#> <gcap> 2022-09-19 11:34:45 info [gcap.runScoring]: checking input data type
#> <gcap> 2022-09-19 11:34:45 info [gcap.runScoring]: checking columns
#> <gcap> 2022-09-19 11:34:45 info [gcap.runScoring]: filtering out records without prob result
#> <gcap> 2022-09-19 11:34:45 info [gcap.runScoring]: joining extra annotation data
#> <gcap> 2022-09-19 11:34:47 info [gcap.runScoring]: only keep genes with copy number >= ploidy+2 in result fCNA object
#> summarizing sample...
#>   classifying samples with min_n=1
#> done
#> ======================
#> A <fCNA> object
#>   record: 255
#>     case: 2
#>      |__ (0) noncircular
#>      |__ (0) possibly_circular
#>      |__ (2) circular
#> ======================
#> <gcap> 2022-09-19 11:34:47 info [gcap.runScoring]: done
#> <gcap> 2022-09-19 11:34:47 info [gcap.ASCNworkflow]: Saving raw prediction result to C:\Users\ADMINI~1\AppData\Local\Temp\RtmpKEtn9x/gcap_02dae75c-37cc-11ed-8000-03c38a3f6043_prediction_result.rds
#> <gcap> 2022-09-19 11:34:48 info [gcap.ASCNworkflow]: Saving fCNA records and sample info to C:\Users\ADMINI~1\AppData\Local\Temp\RtmpKEtn9x/gcap_02dae75c-37cc-11ed-8000-03c38a3f6043_fCNA_records.csv, C:\Users\ADMINI~1\AppData\Local\Temp\RtmpKEtn9x/gcap_02dae75c-37cc-11ed-8000-03c38a3f6043_sample_info.csv
#> <gcap> 2022-09-19 11:34:48 info [gcap.ASCNworkflow]: ===========================================
#> <gcap> 2022-09-19 11:34:48 info [gcap.ASCNworkflow]:  Done! Thanks for using GCAP ASCN workflow
#> <gcap> 2022-09-19 11:34:48 info [gcap.ASCNworkflow]: ===========================================
data$age <- 60
data$gender <- "XY"
rv3 <- gcap.ASCNworkflow(data, outdir = tempdir(), model = "XGB32")
#> <gcap> 2022-09-19 11:34:48 info [gcap.ASCNworkflow]: =========================
#> <gcap> 2022-09-19 11:34:48 info [gcap.ASCNworkflow]:    GCAP ASCN WORKFLOW    
#> <gcap> 2022-09-19 11:34:48 info [gcap.ASCNworkflow]: =========================
#> 
#> <gcap> 2022-09-19 11:34:48 info [gcap.ASCNworkflow]: ============================================================
#> <gcap> 2022-09-19 11:34:48 info [gcap.ASCNworkflow]: Step 1: Extract features and collapse features to gene level
#> <gcap> 2022-09-19 11:34:48 info [gcap.ASCNworkflow]: ============================================================
#> <gcap> 2022-09-19 11:34:48 info [gcap.runASCATBuildflow]: extracting sample-level and region-level features
#> <gcap> 2022-09-19 11:34:48 info [gcap.extractFeatures]: > Extract features from input ascn_data<
#> <gcap> 2022-09-19 11:34:48 info [gcap.extractFeatures]: generating CopyNumber object in sigminer package
#> ℹ [2022-09-19 11:34:48]: Started.
#> ℹ [2022-09-19 11:34:48]: Genome build  : hg38.
#> ℹ [2022-09-19 11:34:48]: Genome measure: called.
#> ℹ [2022-09-19 11:34:48]: When add_loh is TRUE, use_all is forced to TRUE.
#> Please drop columns you don't want to keep before reading.
#> ✔ [2022-09-19 11:34:48]: Chromosome size database for build obtained.
#> ℹ [2022-09-19 11:34:48]: Reading input.
#> ✔ [2022-09-19 11:34:48]: A data frame as input detected.
#> ✔ [2022-09-19 11:34:48]: Column names checked.
#> ✔ [2022-09-19 11:34:48]: Column order set.
#> ✔ [2022-09-19 11:34:48]: Chromosomes unified.
#> ✔ [2022-09-19 11:34:48]: Data imported.
#> ℹ [2022-09-19 11:34:48]: Segments info:
#> ℹ [2022-09-19 11:34:48]:     Keep - 129
#> ℹ [2022-09-19 11:34:48]:     Drop - 0
#> ✔ [2022-09-19 11:34:48]: Segments sorted.
#> ℹ [2022-09-19 11:34:48]: Adding LOH labels...
#> ℹ [2022-09-19 11:34:48]: Skipped joining adjacent segments with same copy number value.
#> ✔ [2022-09-19 11:34:48]: Segmental table cleaned.
#> ℹ [2022-09-19 11:34:48]: Annotating.
#> ✔ [2022-09-19 11:34:48]: Annotation done.
#> ℹ [2022-09-19 11:34:48]: Summarizing per sample.
#> ✔ [2022-09-19 11:34:48]: Summarized.
#> ℹ [2022-09-19 11:34:48]: Generating CopyNumber object.
#> ✔ [2022-09-19 11:34:48]: Generated.
#> ℹ [2022-09-19 11:34:48]: Validating object.
#> ✔ [2022-09-19 11:34:48]: Done.
#> ℹ [2022-09-19 11:34:48]: 0.094 secs elapsed.
#> <gcap> 2022-09-19 11:34:48 info [gcap.extractFeatures]: estimating ploidy from copy number data
#> <gcap> 2022-09-19 11:34:48 info [gcap.extractFeatures]: checking if input data contains ploidy and if there are NAs should be overwritten
#> <gcap> 2022-09-19 11:34:48 info [gcap.extractFeatures]: getting Aneuploidy score
#> <gcap> 2022-09-19 11:34:48 info [gcap.extractFeatures]: getting pLOH score
#> <gcap> 2022-09-19 11:34:48 info [gcap.extractFeatures]: getting CNA burden
#> <gcap> 2022-09-19 11:34:48 info [gcap.extractFeatures]: generating copy number catalog matrix for fitting signature activity
#> ℹ [2022-09-19 11:34:48]: Started.
#> ℹ [2022-09-19 11:34:48]: When you use method 'S', please make sure you have set 'join_adj_seg' to FALSE and 'add_loh' to TRUE in 'read_copynumber() in the previous step!
#> ✔ [2022-09-19 11:34:48]: Matrix generated.
#> ℹ [2022-09-19 11:34:48]: 0.021 secs elapsed.
#> <gcap> 2022-09-19 11:34:48 info [gcap.extractFeatures]: fitting copy number signature activity
#> ℹ [2022-09-19 11:34:48]: Started.
#> ✔ [2022-09-19 11:34:48]: Signature index detected.
#> ℹ [2022-09-19 11:34:48]: Checking signature database in package.
#> ℹ [2022-09-19 11:34:48]: Checking signature index.
#> ℹ [2022-09-19 11:34:48]: Valid index for db 'CNS_TCGA':
#> CN1 CN2 CN3 CN4 CN5 CN6 CN7 CN8 CN9 CN10 CN11 CN12 CN13 CN14 CN15 CN16 CN17 CN18 CN19
#> ✔ [2022-09-19 11:34:48]: Database and index checked.
#> ✔ [2022-09-19 11:34:48]: Signature normalized.
#> ℹ [2022-09-19 11:34:48]: Checking row number for catalog matrix and signature matrix.
#> ✔ [2022-09-19 11:34:48]: Checked.
#> ℹ [2022-09-19 11:34:48]: Checking rownames for catalog matrix and signature matrix.
#> ✔ [2022-09-19 11:34:48]: Checked.
#> ✔ [2022-09-19 11:34:48]: Method 'QP' detected.
#> ✔ [2022-09-19 11:34:48]: Corresponding function generated.
#> ℹ [2022-09-19 11:34:48]: Calling function.
#> ℹ [2022-09-19 11:34:48]: Fitting sample: TCGA-02-2485-01
#> ℹ [2022-09-19 11:34:48]: Fitting sample: TCGA-05-4422-01
#> ✔ [2022-09-19 11:34:48]: Done.
#> ℹ [2022-09-19 11:34:48]: Generating output signature exposures.
#> ✔ [2022-09-19 11:34:48]: Done.
#> ℹ [2022-09-19 11:34:48]: 0.044 secs elapsed.
#> <gcap> 2022-09-19 11:34:48 info [gcap.extractFeatures]: merging data
#> <gcap> 2022-09-19 11:34:48 info [gcap.extractFeatures]: feature extraction done
#> <gcap> 2022-09-19 11:34:48 info [gcap.extractFeatures]: now you can modify the result and append 'age' and 'gender' columns to the 'fts_sample' element of result list
#> <gcap> 2022-09-19 11:34:48 info [gcap.runASCATBuildflow]: checking if input age, gender (and type) columns
#> <gcap> 2022-09-19 11:34:48 info [gcap.runASCATBuildflow]: collapsing all data into gene-level prediction input
#> <gcap> 2022-09-19 11:34:48 info [gcap.collapse2Genes]: please make sure the first 3 columns of `fts$fts_region` are for chr, start, end.
#> <gcap> 2022-09-19 11:34:48 info [gcap.collapse2Genes]: checking data type for age column
#> <gcap> 2022-09-19 11:34:48 info [gcap.collapse2Genes]: checking data type for gender column
#> <gcap> 2022-09-19 11:34:48 warn [gcap.collapse2Genes]: non numeric type found, transforming XY to 1, otherwise 0
#> <gcap> 2022-09-19 11:34:48 info [gcap.collapse2Genes]: collapsing region-level features to gene-level
#> <gcap> 2022-09-19 11:34:48 info [collapse_to_genes]: checking input chromosome names
#> <gcap> 2022-09-19 11:34:48 info [collapse_to_genes]: reading reference file D:/Rlib/gcap/extdata/hg38_target_genes.rds
#> <gcap> 2022-09-19 11:34:48 info [collapse_to_genes]: finding overlaps
#> <gcap> 2022-09-19 11:34:48 info [collapse_to_genes]: calculating intersect size
#> <gcap> 2022-09-19 11:34:48 info [collapse_to_genes]: keeping records with >= 100% overlap ratio with a gene
#> <gcap> 2022-09-19 11:34:48 info [gcap.collapse2Genes]: merging gene-level and sample-level data
#> <gcap> 2022-09-19 11:34:48 info [gcap.collapse2Genes]: merging data and extra info
#> <gcap> 2022-09-19 11:34:48 info [gcap.collapse2Genes]: merging data and prior amplicon frequency data
#> <gcap> 2022-09-19 11:34:48 info [gcap.collapse2Genes]: done
#> <gcap> 2022-09-19 11:34:48 info [gcap.ASCNworkflow]: =======================
#> <gcap> 2022-09-19 11:34:48 info [gcap.ASCNworkflow]: Step 2: Run prediction
#> <gcap> 2022-09-19 11:34:48 info [gcap.ASCNworkflow]: =======================
#> <gcap> 2022-09-19 11:34:48 info [gcap.runPrediction]: using model file XGB_NF32.rds
#> <gcap> 2022-09-19 11:34:48 info [gcap.runPrediction]: selecting necessary features from input data
#> <gcap> 2022-09-19 11:34:48 info [gcap.runPrediction]: running prediction
#> [11:34:48] WARNING: amalgamation/../src/c_api/c_api.cc:718: `ntree_limit` is deprecated, use `iteration_range` instead.
#> <gcap> 2022-09-19 11:34:48 info [gcap.ASCNworkflow]: ====================================
#> <gcap> 2022-09-19 11:34:48 info [gcap.ASCNworkflow]: Step 3: Run scoring and summarizing
#> <gcap> 2022-09-19 11:34:48 info [gcap.ASCNworkflow]: ====================================
#> <gcap> 2022-09-19 11:34:48 info [gcap.runScoring]: checking input data type
#> <gcap> 2022-09-19 11:34:48 info [gcap.runScoring]: checking columns
#> <gcap> 2022-09-19 11:34:48 info [gcap.runScoring]: filtering out records without prob result
#> <gcap> 2022-09-19 11:34:48 info [gcap.runScoring]: joining extra annotation data
#> <gcap> 2022-09-19 11:34:51 info [gcap.runScoring]: only keep genes with copy number >= ploidy+2 in result fCNA object
#> summarizing sample...
#>   classifying samples with min_n=1
#> done
#> ======================
#> A <fCNA> object
#>   record: 255
#>     case: 2
#>      |__ (0) noncircular
#>      |__ (0) possibly_circular
#>      |__ (2) circular
#> ======================
#> <gcap> 2022-09-19 11:34:51 info [gcap.runScoring]: done
#> <gcap> 2022-09-19 11:34:51 info [gcap.ASCNworkflow]: Saving raw prediction result to C:\Users\ADMINI~1\AppData\Local\Temp\RtmpKEtn9x/gcap_04e8c6fe-37cc-11ed-8000-03c38a3f6043_prediction_result.rds
#> <gcap> 2022-09-19 11:34:51 info [gcap.ASCNworkflow]: Saving fCNA records and sample info to C:\Users\ADMINI~1\AppData\Local\Temp\RtmpKEtn9x/gcap_04e8c6fe-37cc-11ed-8000-03c38a3f6043_fCNA_records.csv, C:\Users\ADMINI~1\AppData\Local\Temp\RtmpKEtn9x/gcap_04e8c6fe-37cc-11ed-8000-03c38a3f6043_sample_info.csv
#> <gcap> 2022-09-19 11:34:51 info [gcap.ASCNworkflow]: ===========================================
#> <gcap> 2022-09-19 11:34:51 info [gcap.ASCNworkflow]:  Done! Thanks for using GCAP ASCN workflow
#> <gcap> 2022-09-19 11:34:51 info [gcap.ASCNworkflow]: ===========================================
# If you want to use 'XGB56', you should include 'type' column
data$type <- "LUAD"
rv4 <- gcap.ASCNworkflow(data, outdir = tempdir(), model = "XGB56")
#> <gcap> 2022-09-19 11:34:51 info [gcap.ASCNworkflow]: =========================
#> <gcap> 2022-09-19 11:34:51 info [gcap.ASCNworkflow]:    GCAP ASCN WORKFLOW    
#> <gcap> 2022-09-19 11:34:51 info [gcap.ASCNworkflow]: =========================
#> 
#> <gcap> 2022-09-19 11:34:51 info [gcap.ASCNworkflow]: ============================================================
#> <gcap> 2022-09-19 11:34:51 info [gcap.ASCNworkflow]: Step 1: Extract features and collapse features to gene level
#> <gcap> 2022-09-19 11:34:51 info [gcap.ASCNworkflow]: ============================================================
#> <gcap> 2022-09-19 11:34:51 info [gcap.runASCATBuildflow]: extracting sample-level and region-level features
#> <gcap> 2022-09-19 11:34:51 info [gcap.extractFeatures]: > Extract features from input ascn_data<
#> <gcap> 2022-09-19 11:34:51 info [gcap.extractFeatures]: generating CopyNumber object in sigminer package
#> ℹ [2022-09-19 11:34:51]: Started.
#> ℹ [2022-09-19 11:34:51]: Genome build  : hg38.
#> ℹ [2022-09-19 11:34:51]: Genome measure: called.
#> ℹ [2022-09-19 11:34:51]: When add_loh is TRUE, use_all is forced to TRUE.
#> Please drop columns you don't want to keep before reading.
#> ✔ [2022-09-19 11:34:51]: Chromosome size database for build obtained.
#> ℹ [2022-09-19 11:34:51]: Reading input.
#> ✔ [2022-09-19 11:34:51]: A data frame as input detected.
#> ✔ [2022-09-19 11:34:51]: Column names checked.
#> ✔ [2022-09-19 11:34:51]: Column order set.
#> ✔ [2022-09-19 11:34:51]: Chromosomes unified.
#> ✔ [2022-09-19 11:34:51]: Data imported.
#> ℹ [2022-09-19 11:34:51]: Segments info:
#> ℹ [2022-09-19 11:34:51]:     Keep - 129
#> ℹ [2022-09-19 11:34:51]:     Drop - 0
#> ✔ [2022-09-19 11:34:51]: Segments sorted.
#> ℹ [2022-09-19 11:34:51]: Adding LOH labels...
#> ℹ [2022-09-19 11:34:51]: Skipped joining adjacent segments with same copy number value.
#> ✔ [2022-09-19 11:34:51]: Segmental table cleaned.
#> ℹ [2022-09-19 11:34:51]: Annotating.
#> ✔ [2022-09-19 11:34:51]: Annotation done.
#> ℹ [2022-09-19 11:34:51]: Summarizing per sample.
#> ✔ [2022-09-19 11:34:51]: Summarized.
#> ℹ [2022-09-19 11:34:51]: Generating CopyNumber object.
#> ✔ [2022-09-19 11:34:51]: Generated.
#> ℹ [2022-09-19 11:34:51]: Validating object.
#> ✔ [2022-09-19 11:34:51]: Done.
#> ℹ [2022-09-19 11:34:51]: 0.095 secs elapsed.
#> <gcap> 2022-09-19 11:34:51 info [gcap.extractFeatures]: estimating ploidy from copy number data
#> <gcap> 2022-09-19 11:34:51 info [gcap.extractFeatures]: checking if input data contains ploidy and if there are NAs should be overwritten
#> <gcap> 2022-09-19 11:34:51 info [gcap.extractFeatures]: getting Aneuploidy score
#> <gcap> 2022-09-19 11:34:51 info [gcap.extractFeatures]: getting pLOH score
#> <gcap> 2022-09-19 11:34:51 info [gcap.extractFeatures]: getting CNA burden
#> <gcap> 2022-09-19 11:34:51 info [gcap.extractFeatures]: generating copy number catalog matrix for fitting signature activity
#> ℹ [2022-09-19 11:34:51]: Started.
#> ℹ [2022-09-19 11:34:51]: When you use method 'S', please make sure you have set 'join_adj_seg' to FALSE and 'add_loh' to TRUE in 'read_copynumber() in the previous step!
#> ✔ [2022-09-19 11:34:51]: Matrix generated.
#> ℹ [2022-09-19 11:34:51]: 0.022 secs elapsed.
#> <gcap> 2022-09-19 11:34:51 info [gcap.extractFeatures]: fitting copy number signature activity
#> ℹ [2022-09-19 11:34:51]: Started.
#> ✔ [2022-09-19 11:34:51]: Signature index detected.
#> ℹ [2022-09-19 11:34:51]: Checking signature database in package.
#> ℹ [2022-09-19 11:34:51]: Checking signature index.
#> ℹ [2022-09-19 11:34:51]: Valid index for db 'CNS_TCGA':
#> CN1 CN2 CN3 CN4 CN5 CN6 CN7 CN8 CN9 CN10 CN11 CN12 CN13 CN14 CN15 CN16 CN17 CN18 CN19
#> ✔ [2022-09-19 11:34:51]: Database and index checked.
#> ✔ [2022-09-19 11:34:51]: Signature normalized.
#> ℹ [2022-09-19 11:34:51]: Checking row number for catalog matrix and signature matrix.
#> ✔ [2022-09-19 11:34:51]: Checked.
#> ℹ [2022-09-19 11:34:51]: Checking rownames for catalog matrix and signature matrix.
#> ✔ [2022-09-19 11:34:51]: Checked.
#> ✔ [2022-09-19 11:34:51]: Method 'QP' detected.
#> ✔ [2022-09-19 11:34:51]: Corresponding function generated.
#> ℹ [2022-09-19 11:34:51]: Calling function.
#> ℹ [2022-09-19 11:34:51]: Fitting sample: TCGA-02-2485-01
#> ℹ [2022-09-19 11:34:51]: Fitting sample: TCGA-05-4422-01
#> ✔ [2022-09-19 11:34:51]: Done.
#> ℹ [2022-09-19 11:34:51]: Generating output signature exposures.
#> ✔ [2022-09-19 11:34:51]: Done.
#> ℹ [2022-09-19 11:34:51]: 0.044 secs elapsed.
#> <gcap> 2022-09-19 11:34:51 info [gcap.extractFeatures]: merging data
#> <gcap> 2022-09-19 11:34:51 info [gcap.extractFeatures]: feature extraction done
#> <gcap> 2022-09-19 11:34:51 info [gcap.extractFeatures]: now you can modify the result and append 'age' and 'gender' columns to the 'fts_sample' element of result list
#> <gcap> 2022-09-19 11:34:51 info [gcap.runASCATBuildflow]: checking if input age, gender (and type) columns
#> <gcap> 2022-09-19 11:34:51 info [gcap.runASCATBuildflow]: collapsing all data into gene-level prediction input
#> <gcap> 2022-09-19 11:34:51 info [gcap.collapse2Genes]: please make sure the first 3 columns of `fts$fts_region` are for chr, start, end.
#> <gcap> 2022-09-19 11:34:51 info [gcap.collapse2Genes]: one-hot encoding cancer type
#> <gcap> 2022-09-19 11:34:51 info [gcap.collapse2Genes]: valid cancer types: BLCA,BRCA,CESC,COAD,DLBC,ESCA,GBM,HNSC,KICH,KIRC,KIRP,LGG,LIHC,LUAD,LUSC,OV,PRAD,READ,SARC,SKCM,STAD,THCA,UCEC,UVM
#> <gcap> 2022-09-19 11:34:52 info [gcap.collapse2Genes]: checking data type for age column
#> <gcap> 2022-09-19 11:34:52 info [gcap.collapse2Genes]: checking data type for gender column
#> <gcap> 2022-09-19 11:34:52 warn [gcap.collapse2Genes]: non numeric type found, transforming XY to 1, otherwise 0
#> <gcap> 2022-09-19 11:34:52 info [gcap.collapse2Genes]: collapsing region-level features to gene-level
#> <gcap> 2022-09-19 11:34:52 info [collapse_to_genes]: checking input chromosome names
#> <gcap> 2022-09-19 11:34:52 info [collapse_to_genes]: reading reference file D:/Rlib/gcap/extdata/hg38_target_genes.rds
#> <gcap> 2022-09-19 11:34:52 info [collapse_to_genes]: finding overlaps
#> <gcap> 2022-09-19 11:34:52 info [collapse_to_genes]: calculating intersect size
#> <gcap> 2022-09-19 11:34:52 info [collapse_to_genes]: keeping records with >= 100% overlap ratio with a gene
#> <gcap> 2022-09-19 11:34:52 info [gcap.collapse2Genes]: merging gene-level and sample-level data
#> <gcap> 2022-09-19 11:34:52 info [gcap.collapse2Genes]: merging data and extra info
#> <gcap> 2022-09-19 11:34:52 info [gcap.collapse2Genes]: merging data and prior amplicon frequency data
#> <gcap> 2022-09-19 11:34:52 info [gcap.collapse2Genes]: done
#> <gcap> 2022-09-19 11:34:52 info [gcap.ASCNworkflow]: =======================
#> <gcap> 2022-09-19 11:34:52 info [gcap.ASCNworkflow]: Step 2: Run prediction
#> <gcap> 2022-09-19 11:34:52 info [gcap.ASCNworkflow]: =======================
#> <gcap> 2022-09-19 11:34:52 info [gcap.runPrediction]: using model file XGB_NF56.rds
#> <gcap> 2022-09-19 11:34:52 info [gcap.runPrediction]: selecting necessary features from input data
#> <gcap> 2022-09-19 11:34:52 info [gcap.runPrediction]: running prediction
#> [11:34:52] WARNING: amalgamation/../src/c_api/c_api.cc:718: `ntree_limit` is deprecated, use `iteration_range` instead.
#> <gcap> 2022-09-19 11:34:52 info [gcap.ASCNworkflow]: ====================================
#> <gcap> 2022-09-19 11:34:52 info [gcap.ASCNworkflow]: Step 3: Run scoring and summarizing
#> <gcap> 2022-09-19 11:34:52 info [gcap.ASCNworkflow]: ====================================
#> <gcap> 2022-09-19 11:34:52 info [gcap.runScoring]: checking input data type
#> <gcap> 2022-09-19 11:34:52 info [gcap.runScoring]: checking columns
#> <gcap> 2022-09-19 11:34:52 info [gcap.runScoring]: filtering out records without prob result
#> <gcap> 2022-09-19 11:34:52 info [gcap.runScoring]: joining extra annotation data
#> <gcap> 2022-09-19 11:34:55 info [gcap.runScoring]: only keep genes with copy number >= ploidy+2 in result fCNA object
#> summarizing sample...
#>   classifying samples with min_n=1
#> done
#> ======================
#> A <fCNA> object
#>   record: 255
#>     case: 2
#>      |__ (0) noncircular
#>      |__ (0) possibly_circular
#>      |__ (2) circular
#> ======================
#> <gcap> 2022-09-19 11:34:55 info [gcap.runScoring]: done
#> <gcap> 2022-09-19 11:34:55 info [gcap.ASCNworkflow]: Saving raw prediction result to C:\Users\ADMINI~1\AppData\Local\Temp\RtmpKEtn9x/gcap_0745c302-37cc-11ed-8000-03c38a3f6043_prediction_result.rds
#> <gcap> 2022-09-19 11:34:55 info [gcap.ASCNworkflow]: Saving fCNA records and sample info to C:\Users\ADMINI~1\AppData\Local\Temp\RtmpKEtn9x/gcap_0745c302-37cc-11ed-8000-03c38a3f6043_fCNA_records.csv, C:\Users\ADMINI~1\AppData\Local\Temp\RtmpKEtn9x/gcap_0745c302-37cc-11ed-8000-03c38a3f6043_sample_info.csv
#> <gcap> 2022-09-19 11:34:55 info [gcap.ASCNworkflow]: ===========================================
#> <gcap> 2022-09-19 11:34:55 info [gcap.ASCNworkflow]:  Done! Thanks for using GCAP ASCN workflow
#> <gcap> 2022-09-19 11:34:55 info [gcap.ASCNworkflow]: ===========================================
# If you only have total integer copy number
data$minor_cn <- NA
rv5 <- gcap.ASCNworkflow(data, outdir = tempdir(), model = "XGB11")
#> <gcap> 2022-09-19 11:34:55 info [gcap.ASCNworkflow]: =========================
#> <gcap> 2022-09-19 11:34:55 info [gcap.ASCNworkflow]:    GCAP ASCN WORKFLOW    
#> <gcap> 2022-09-19 11:34:55 info [gcap.ASCNworkflow]: =========================
#> 
#> <gcap> 2022-09-19 11:34:55 info [gcap.ASCNworkflow]: ============================================================
#> <gcap> 2022-09-19 11:34:55 info [gcap.ASCNworkflow]: Step 1: Extract features and collapse features to gene level
#> <gcap> 2022-09-19 11:34:55 info [gcap.ASCNworkflow]: ============================================================
#> <gcap> 2022-09-19 11:34:55 info [gcap.runASCATBuildflow]: extracting sample-level and region-level features
#> <gcap> 2022-09-19 11:34:55 info [gcap.extractFeatures]: > Extract features from input ascn_data<
#> <gcap> 2022-09-19 11:34:55 info [gcap.extractFeatures]: generating CopyNumber object in sigminer package
#> ℹ [2022-09-19 11:34:55]: Started.
#> ℹ [2022-09-19 11:34:55]: Genome build  : hg38.
#> ℹ [2022-09-19 11:34:55]: Genome measure: called.
#> ✔ [2022-09-19 11:34:55]: Chromosome size database for build obtained.
#> ℹ [2022-09-19 11:34:55]: Reading input.
#> ✔ [2022-09-19 11:34:55]: A data frame as input detected.
#> ✔ [2022-09-19 11:34:55]: Column names checked.
#> ✔ [2022-09-19 11:34:55]: Column order set.
#> ✔ [2022-09-19 11:34:55]: Chromosomes unified.
#> ✔ [2022-09-19 11:34:55]: Data imported.
#> ℹ [2022-09-19 11:34:55]: Segments info:
#> ℹ [2022-09-19 11:34:55]:     Keep - 129
#> ℹ [2022-09-19 11:34:55]:     Drop - 0
#> ✔ [2022-09-19 11:34:55]: Segments sorted.
#> ℹ [2022-09-19 11:34:55]: Skipped joining adjacent segments with same copy number value.
#> ✔ [2022-09-19 11:34:55]: Segmental table cleaned.
#> ℹ [2022-09-19 11:34:55]: Annotating.
#> ✔ [2022-09-19 11:34:55]: Annotation done.
#> ℹ [2022-09-19 11:34:55]: Summarizing per sample.
#> ✔ [2022-09-19 11:34:55]: Summarized.
#> ℹ [2022-09-19 11:34:55]: Generating CopyNumber object.
#> ✔ [2022-09-19 11:34:55]: Generated.
#> ℹ [2022-09-19 11:34:55]: Validating object.
#> ✔ [2022-09-19 11:34:55]: Done.
#> ℹ [2022-09-19 11:34:55]: 0.085 secs elapsed.
#> <gcap> 2022-09-19 11:34:55 info [gcap.extractFeatures]: estimating ploidy from copy number data
#> <gcap> 2022-09-19 11:34:55 info [gcap.extractFeatures]: checking if input data contains ploidy and if there are NAs should be overwritten
#> <gcap> 2022-09-19 11:34:55 info [gcap.extractFeatures]: getting Aneuploidy score
#> <gcap> 2022-09-19 11:34:55 info [gcap.extractFeatures]: pLOH score is skipped as no LOH data available
#> <gcap> 2022-09-19 11:34:55 info [gcap.extractFeatures]: getting CNA burden
#> <gcap> 2022-09-19 11:34:55 info [gcap.extractFeatures]: copy number signature activity quantification is skipped as no LOH data available
#> <gcap> 2022-09-19 11:34:55 info [gcap.extractFeatures]: merging data
#> <gcap> 2022-09-19 11:34:55 info [gcap.extractFeatures]: feature extraction done
#> <gcap> 2022-09-19 11:34:55 info [gcap.extractFeatures]: now you can modify the result and append 'age' and 'gender' columns to the 'fts_sample' element of result list
#> <gcap> 2022-09-19 11:34:55 info [gcap.runASCATBuildflow]: checking if input age, gender (and type) columns
#> <gcap> 2022-09-19 11:34:55 info [gcap.runASCATBuildflow]: collapsing all data into gene-level prediction input
#> <gcap> 2022-09-19 11:34:55 info [gcap.collapse2Genes]: please make sure the first 3 columns of `fts$fts_region` are for chr, start, end.
#> <gcap> 2022-09-19 11:34:55 info [gcap.collapse2Genes]: one-hot encoding cancer type
#> <gcap> 2022-09-19 11:34:55 info [gcap.collapse2Genes]: valid cancer types: BLCA,BRCA,CESC,COAD,DLBC,ESCA,GBM,HNSC,KICH,KIRC,KIRP,LGG,LIHC,LUAD,LUSC,OV,PRAD,READ,SARC,SKCM,STAD,THCA,UCEC,UVM
#> <gcap> 2022-09-19 11:34:55 info [gcap.collapse2Genes]: checking data type for age column
#> <gcap> 2022-09-19 11:34:55 info [gcap.collapse2Genes]: checking data type for gender column
#> <gcap> 2022-09-19 11:34:55 warn [gcap.collapse2Genes]: non numeric type found, transforming XY to 1, otherwise 0
#> <gcap> 2022-09-19 11:34:55 info [gcap.collapse2Genes]: collapsing region-level features to gene-level
#> <gcap> 2022-09-19 11:34:55 info [collapse_to_genes]: checking input chromosome names
#> <gcap> 2022-09-19 11:34:55 info [collapse_to_genes]: reading reference file D:/Rlib/gcap/extdata/hg38_target_genes.rds
#> <gcap> 2022-09-19 11:34:55 info [collapse_to_genes]: finding overlaps
#> <gcap> 2022-09-19 11:34:55 info [collapse_to_genes]: calculating intersect size
#> <gcap> 2022-09-19 11:34:55 info [collapse_to_genes]: keeping records with >= 100% overlap ratio with a gene
#> <gcap> 2022-09-19 11:34:55 info [gcap.collapse2Genes]: merging gene-level and sample-level data
#> <gcap> 2022-09-19 11:34:55 info [gcap.collapse2Genes]: merging data and extra info
#> <gcap> 2022-09-19 11:34:55 info [gcap.collapse2Genes]: merging data and prior amplicon frequency data
#> <gcap> 2022-09-19 11:34:55 info [gcap.collapse2Genes]: done
#> <gcap> 2022-09-19 11:34:55 info [gcap.ASCNworkflow]: =======================
#> <gcap> 2022-09-19 11:34:55 info [gcap.ASCNworkflow]: Step 2: Run prediction
#> <gcap> 2022-09-19 11:34:55 info [gcap.ASCNworkflow]: =======================
#> <gcap> 2022-09-19 11:34:55 info [gcap.runPrediction]: using model file XGB_NF11.rds
#> <gcap> 2022-09-19 11:34:55 info [gcap.runPrediction]: selecting necessary features from input data
#> <gcap> 2022-09-19 11:34:56 info [gcap.runPrediction]: running prediction
#> [11:34:56] WARNING: amalgamation/../src/c_api/c_api.cc:718: `ntree_limit` is deprecated, use `iteration_range` instead.
#> <gcap> 2022-09-19 11:34:56 info [gcap.ASCNworkflow]: ====================================
#> <gcap> 2022-09-19 11:34:56 info [gcap.ASCNworkflow]: Step 3: Run scoring and summarizing
#> <gcap> 2022-09-19 11:34:56 info [gcap.ASCNworkflow]: ====================================
#> <gcap> 2022-09-19 11:34:56 info [gcap.runScoring]: checking input data type
#> <gcap> 2022-09-19 11:34:56 info [gcap.runScoring]: checking columns
#> <gcap> 2022-09-19 11:34:56 info [gcap.runScoring]: filtering out records without prob result
#> <gcap> 2022-09-19 11:34:56 info [gcap.runScoring]: joining extra annotation data
#> <gcap> 2022-09-19 11:34:58 info [gcap.runScoring]: only keep genes with copy number >= ploidy+2 in result fCNA object
#> summarizing sample...
#>   classifying samples with min_n=1
#> done
#> ======================
#> A <fCNA> object
#>   record: 255
#>     case: 2
#>      |__ (0) noncircular
#>      |__ (0) possibly_circular
#>      |__ (2) circular
#> ======================
#> <gcap> 2022-09-19 11:34:58 info [gcap.runScoring]: done
#> <gcap> 2022-09-19 11:34:58 info [gcap.ASCNworkflow]: Saving raw prediction result to C:\Users\ADMINI~1\AppData\Local\Temp\RtmpKEtn9x/gcap_0957cb40-37cc-11ed-8000-03c38a3f6043_prediction_result.rds
#> <gcap> 2022-09-19 11:34:59 info [gcap.ASCNworkflow]: Saving fCNA records and sample info to C:\Users\ADMINI~1\AppData\Local\Temp\RtmpKEtn9x/gcap_0957cb40-37cc-11ed-8000-03c38a3f6043_fCNA_records.csv, C:\Users\ADMINI~1\AppData\Local\Temp\RtmpKEtn9x/gcap_0957cb40-37cc-11ed-8000-03c38a3f6043_sample_info.csv
#> <gcap> 2022-09-19 11:34:59 info [gcap.ASCNworkflow]: ===========================================
#> <gcap> 2022-09-19 11:34:59 info [gcap.ASCNworkflow]:  Done! Thanks for using GCAP ASCN workflow
#> <gcap> 2022-09-19 11:34:59 info [gcap.ASCNworkflow]: ===========================================

# R6 class fCNA --------------------------------
print(rv)
#> ======================
#> A <fCNA> object
#>   record: 255
#>     case: 2
#>      |__ (0) noncircular
#>      |__ (0) possibly_circular
#>      |__ (2) circular
#> ======================
print(rv$data)
#>               sample        band         gene_id total_cn minor_cn ploidy
#>   1: TCGA-02-2485-01 chr12:q13.3 ENSG00000111087       35        2   4.09
#>   2: TCGA-02-2485-01 chr12:q13.3 ENSG00000123329       35        2   4.09
#>   3: TCGA-02-2485-01 chr12:q13.3 ENSG00000135454       35        2   4.09
#>   4: TCGA-02-2485-01 chr12:q13.3 ENSG00000139269       35        2   4.09
#>   5: TCGA-02-2485-01 chr12:q13.3 ENSG00000155980       35        2   4.09
#>  ---                                                                     
#> 251: TCGA-05-4422-01  chr8:q24.3 ENSG00000204882        5        1   2.09
#> 252: TCGA-05-4422-01  chr8:q24.3 ENSG00000226490        5        1   2.09
#> 253: TCGA-05-4422-01  chr8:q24.3 ENSG00000234616        5        1   2.09
#> 254: TCGA-05-4422-01  chr8:q24.3 ENSG00000283992        5        1   2.09
#> 255: TCGA-05-4422-01  chr8:q24.3 ENSG00000284505        5        1   2.09
#>           prob thlevel
#>   1: 0.9875240      l3
#>   2: 0.9875240      l3
#>   3: 0.9875240      l3
#>   4: 0.9875240      l3
#>   5: 0.9875240      l3
#>  ---                  
#> 251: 0.2838656      l0
#> 252: 0.2838656      l0
#> 253: 0.2838656      l0
#> 254: 0.2838656      l0
#> 255: 0.2838656      l0
print(rv$sample_summary)
#>             sample purity ploidy AScore       pLOH cna_burden      CN1
#> 1: TCGA-02-2485-01      1   4.09     13 0.04766306      0.951  0.00000
#> 2: TCGA-05-4422-01      1   2.09     13 0.13041086      0.249 23.84798
#>          CN2 CN3 CN4 CN5 CN6      CN7 CN8    CN9 CN10     CN11 CN12 CN13 CN14
#> 1: 45.112591   0   0   0   0 9.887409   0  0.000    0 0.000000    0    0    0
#> 2:  2.326143   0   0   0   0 8.939099   0 29.688    0 1.154688    0    0    0
#>        CN15 CN16 CN17     CN18 CN19    class
#> 1: 0.000000    0    0 0.000000    0 circular
#> 2: 1.178945    0    0 6.865139    0 circular
print(rv$gene_summary)
#> NULL
print(rv$cytoband_summary)
#> NULL

# Create a subset fCNA
rv_subset <- rv$subset(total_cn > 10)
#> summarizing sample...
#>   classifying samples with min_n=1
#>   overlapping columns detected, remove them from input 'pdata'
#> done
nrow(rv$data)
#> [1] 255
nrow(rv_subset$data)
#> [1] 85

rv_subset2 <- rv$subset(sample == "TCGA-02-2485-01")
#> summarizing sample...
#>   classifying samples with min_n=1
#>   overlapping columns detected, remove them from input 'pdata'
#> done
nrow(rv_subset2$data)
#> [1] 75
unique(rv_subset2$data$sample)
#> [1] "TCGA-02-2485-01"

sum_gene <- rv$getGeneSummary()
sum_gene
#>              gene_id noncircular circular Total
#>   1: ENSG00000008513           0        1     1
#>   2: ENSG00000026036           0        1     1
#>   3: ENSG00000034677           0        1     1
#>   4: ENSG00000035687           0        1     1
#>   5: ENSG00000037897           0        1     1
#>  ---                                           
#> 181: ENSG00000274443           0        1     1
#> 182: ENSG00000284188           0        1     1
#> 183: ENSG00000285133           0        1     1
#> 184: ENSG00000285982           0        1     1
#> 185: ENSG00000287908           0        1     1
mat_gene <- rv$getGeneSummary(return_mat = TRUE)
#> Error in rv$getGeneSummary(return_mat = TRUE): unused argument (return_mat = TRUE)
mat_gene
#> Error in eval(expr, envir, enclos): object 'mat_gene' not found

sum_cytoband <- rv$getCytobandSummary()
sum_cytoband
#>             band noncircular circular Total
#>  1:  chr12:q13.3           0       17    17
#>  2:  chr12:q14.1           0       14    14
#>  3:     chr1:q43           0        5     5
#>  4:     chr1:q44           0       10    10
#>  5: chr20:q13.33           0       30    30
#>  6:     chr4:q12           0        4     4
#>  7:   chr7:p11.2           0       13    13
#>  8:   chr7:p12.1           0        4     4
#>  9:   chr7:p12.2           0        5     5
#> 10:  chr8:q21.11           0       11    11
#> 11:  chr8:q21.13           0        4     4
#> 12:   chr8:q22.1           0        8     8
#> 13:   chr8:q22.2           0       16    16
#> 14:   chr8:q22.3           0       19    19
#> 15:  chr8:q24.13           0        5     5
#> 16:  chr8:q24.21           0        6     6
#> 17:  chr8:q24.22           0       13    13
#> 18:  chr8:q24.23           0        1     1
mat_cytoband <- rv$getCytobandSummary(return_mat = TRUE)
#> Error in rv$getCytobandSummary(return_mat = TRUE): unused argument (return_mat = TRUE)
mat_cytoband
#> Error in eval(expr, envir, enclos): object 'mat_cytoband' not found