Draw Oncoprint for fCNA Profile Visualization

See oncoprint and heatmap to modify the result plot in depth.

gcap.plotProfile(
  fCNA,
  genes = NULL,
  samples = NULL,
  merge_circular = FALSE,
  show_column_names = TRUE,
  remove_empty_columns = FALSE,
  remove_empty_rows = TRUE,
  heatmap_legend_param = list(title = "fCNA"),
  col = c(noncircular = "blue", circular = "red", possibly_circular = "magenta"),
  ...
)

Arguments

fCNA

a fCNA object.

genes

a list of genes. Must be subsets of fCNA$data$gene_id.

samples

a vector to subset samples.

merge_circular

if TRUE, merge 'possibly_circular' into 'circular' type.

show_column_names

see ?ComplexHeatmap::oncoPrint.

remove_empty_columns

see ?ComplexHeatmap::oncoPrint.

remove_empty_rows

see ?ComplexHeatmap::oncoPrint.

heatmap_legend_param

see ?ComplexHeatmap::oncoPrint.

col

see ?ComplexHeatmap::oncoPrint.

...

Other parameters passing to ComplexHeatmap::oncoPrint.

Value

a Heatmap object, you can go further modify it.

See also

Examples

# \donttest{
if (require("ComplexHeatmap") && require("IDConverter")) {
  data("ascn")
  data <- ascn

  # Create fake data
  set.seed(1234)
  data$sample <- sample(LETTERS[1:10], nrow(data), replace = TRUE)
  rv <- gcap.ASCNworkflow(data, outdir = tempdir(), model = "XGB11")

  gcap.plotProfile(rv, samples = c("B", "A", "C"))

  rv$convertGeneID()
  ht <- gcap.plotProfile(rv,
    samples = c("B", "A", "C", "D", "F"),
    genes = unique(rv$data$gene_id)[1:10],
    top_annotation = ComplexHeatmap::HeatmapAnnotation(
      cbar = ComplexHeatmap::anno_oncoprint_barplot(),
      foo1 = c("g1", "g1", "g2", "g2", "g3"),
      bar1 = ComplexHeatmap::anno_points(1:5)
    )
  )
  ht
  ComplexHeatmap::draw(ht, merge_legends = TRUE)

  #----------------
  # Plot KM curve
  #----------------
  surv_data <- data.frame(
    sample = rv$sample_summary$sample,
    time = 3000 * abs(rnorm(nrow(rv$sample_summary))),
    status = sample(c(0, 1), nrow(rv$sample_summary), replace = TRUE)
  )
  surv_data
  p <- gcap.plotKMcurve(rv, surv_data)
  p

  p2 <- gcap.plotKMcurve(rv, surv_data, genes = "MYC")
  p2

  # ---------------
  # Plot forest
  # ---------------
  p3 <- gcap.plotForest(rv, surv_data)

  # Fake some data for gene analysis
  rv$data[amplicon_type == "circular"]$gene_id[1:10] <- "MYC"
  rv$data[amplicon_type == "circular"]$gene_id[11:20] <- "EGFR"
  rv$data[amplicon_type == "circular"]$gene_id[31:35] <- "GSX2"
  p4 <- gcap.plotForest(rv, surv_data,
    x = c("TP53", "MYC", "ABC", "EGFR", "GSX2"), x_is_gene = TRUE,
    ref_line = 1, xlim = c(0, 10)
  )
}
#> Loading required package: ComplexHeatmap
#> Loading required package: grid
#> ========================================
#> ComplexHeatmap version 2.8.0
#> Bioconductor page: http://bioconductor.org/packages/ComplexHeatmap/
#> Github page: https://github.com/jokergoo/ComplexHeatmap
#> Documentation: http://jokergoo.github.io/ComplexHeatmap-reference
#> 
#> If you use it in published research, please cite:
#> Gu, Z. Complex heatmaps reveal patterns and correlations in multidimensional 
#>   genomic data. Bioinformatics 2016.
#> 
#> The new InteractiveComplexHeatmap package can directly export static 
#> complex heatmaps into an interactive Shiny app with zero effort. Have a try!
#> 
#> This message can be suppressed by:
#>   suppressPackageStartupMessages(library(ComplexHeatmap))
#> ========================================
#> Loading required package: IDConverter
#> <gcap> 2022-04-06 19:47:48 info [gcap.ASCNworkflow]: =========================
#> <gcap> 2022-04-06 19:47:48 info [gcap.ASCNworkflow]:    GCAP ASCN WORKFLOW    
#> <gcap> 2022-04-06 19:47:48 info [gcap.ASCNworkflow]: =========================
#> 
#> <gcap> 2022-04-06 19:47:48 info [gcap.ASCNworkflow]: ============================================================
#> <gcap> 2022-04-06 19:47:48 info [gcap.ASCNworkflow]: Step 1: Extract features and collapse features to gene level
#> <gcap> 2022-04-06 19:47:48 info [gcap.ASCNworkflow]: ============================================================
#> <gcap> 2022-04-06 19:47:48 info [gcap.runASCATBuildflow]: extracting sample-level and region-level features
#> <gcap> 2022-04-06 19:47:48 info [gcap.extractFeatures]: > Extract features from input ascn_data<
#> <gcap> 2022-04-06 19:47:48 info [gcap.extractFeatures]: generating CopyNumber object in sigminer package
#> i [2022-04-06 19:47:48]: Started.
#> i [2022-04-06 19:47:48]: Genome build  : hg38.
#> i [2022-04-06 19:47:48]: Genome measure: called.
#> i [2022-04-06 19:47:48]: When add_loh is TRUE, use_all is forced to TRUE.
#> Please drop columns you don't want to keep before reading.
#> v [2022-04-06 19:47:48]: Chromosome size database for build obtained.
#> i [2022-04-06 19:47:48]: Reading input.
#> v [2022-04-06 19:47:48]: A data frame as input detected.
#> v [2022-04-06 19:47:48]: Column names checked.
#> v [2022-04-06 19:47:48]: Column order set.
#> v [2022-04-06 19:47:48]: Chromosomes unified.
#> v [2022-04-06 19:47:48]: Data imported.
#> i [2022-04-06 19:47:48]: Segments info:
#> i [2022-04-06 19:47:48]:     Keep - 129
#> i [2022-04-06 19:47:48]:     Drop - 0
#> v [2022-04-06 19:47:48]: Segments sorted.
#> i [2022-04-06 19:47:48]: Adding LOH labels...
#> i [2022-04-06 19:47:48]: Skipped joining adjacent segments with same copy number value.
#> v [2022-04-06 19:47:48]: Segmental table cleaned.
#> i [2022-04-06 19:47:48]: Annotating.
#> v [2022-04-06 19:47:48]: Annotation done.
#> i [2022-04-06 19:47:48]: Summarizing per sample.
#> v [2022-04-06 19:47:48]: Summarized.
#> i [2022-04-06 19:47:48]: Generating CopyNumber object.
#> v [2022-04-06 19:47:48]: Generated.
#> i [2022-04-06 19:47:48]: Validating object.
#> v [2022-04-06 19:47:48]: Done.
#> i [2022-04-06 19:47:48]: 0.115 secs elapsed.
#> <gcap> 2022-04-06 19:47:48 info [gcap.extractFeatures]: estimating ploidy from copy number data
#> <gcap> 2022-04-06 19:47:48 info [gcap.extractFeatures]: checking if input data contains ploidy and if there are NAs should be overwritten
#> <gcap> 2022-04-06 19:47:48 info [gcap.extractFeatures]: getting Aneuploidy score
#> <gcap> 2022-04-06 19:47:48 info [gcap.extractFeatures]: getting pLOH score
#> <gcap> 2022-04-06 19:47:48 info [gcap.extractFeatures]: getting CNA burden
#> <gcap> 2022-04-06 19:47:48 info [gcap.extractFeatures]: generating copy number catalog matrix for fitting signature activity
#> i [2022-04-06 19:47:48]: Started.
#> i [2022-04-06 19:47:48]: When you use method 'S', please make sure you have set 'join_adj_seg' to FALSE and 'add_loh' to TRUE in 'read_copynumber() in the previous step!
#> v [2022-04-06 19:47:48]: Matrix generated.
#> i [2022-04-06 19:47:48]: 0.02 secs elapsed.
#> <gcap> 2022-04-06 19:47:48 info [gcap.extractFeatures]: fitting copy number signature activity
#> i [2022-04-06 19:47:48]: Started.
#> v [2022-04-06 19:47:48]: Signature index detected.
#> i [2022-04-06 19:47:48]: Checking signature database in package.
#> i [2022-04-06 19:47:48]: Checking signature index.
#> i [2022-04-06 19:47:48]: Valid index for db 'CNS_TCGA':
#> CN1 CN2 CN3 CN4 CN5 CN6 CN7 CN8 CN9 CN10 CN11 CN12 CN13 CN14 CN15 CN16 CN17 CN18 CN19
#> v [2022-04-06 19:47:48]: Database and index checked.
#> v [2022-04-06 19:47:48]: Signature normalized.
#> i [2022-04-06 19:47:48]: Checking row number for catalog matrix and signature matrix.
#> v [2022-04-06 19:47:48]: Checked.
#> i [2022-04-06 19:47:48]: Checking rownames for catalog matrix and signature matrix.
#> v [2022-04-06 19:47:48]: Checked.
#> v [2022-04-06 19:47:48]: Method 'QP' detected.
#> v [2022-04-06 19:47:48]: Corresponding function generated.
#> i [2022-04-06 19:47:48]: Calling function.
#> i [2022-04-06 19:47:48]: Fitting sample: A
#> i [2022-04-06 19:47:48]: Fitting sample: B
#> i [2022-04-06 19:47:48]: Fitting sample: C
#> i [2022-04-06 19:47:48]: Fitting sample: D
#> i [2022-04-06 19:47:48]: Fitting sample: E
#> i [2022-04-06 19:47:48]: Fitting sample: F
#> i [2022-04-06 19:47:48]: Fitting sample: G
#> i [2022-04-06 19:47:48]: Fitting sample: H
#> i [2022-04-06 19:47:48]: Fitting sample: I
#> i [2022-04-06 19:47:48]: Fitting sample: J
#> v [2022-04-06 19:47:48]: Done.
#> i [2022-04-06 19:47:48]: Generating output signature exposures.
#> v [2022-04-06 19:47:48]: Done.
#> i [2022-04-06 19:47:48]: 0.05 secs elapsed.
#> <gcap> 2022-04-06 19:47:48 info [gcap.extractFeatures]: merging data
#> <gcap> 2022-04-06 19:47:48 info [gcap.extractFeatures]: feature extraction done
#> <gcap> 2022-04-06 19:47:48 info [gcap.extractFeatures]: now you can modify the result and append 'age' and 'gender' columns to the 'fts_sample' element of result list
#> <gcap> 2022-04-06 19:47:48 info [gcap.runASCATBuildflow]: checking if input age, gender (and type) columns
#> <gcap> 2022-04-06 19:47:48 info [gcap.runASCATBuildflow]: None exist, set to NULL.
#> <gcap> 2022-04-06 19:47:48 info [gcap.runASCATBuildflow]: collapsing all data into gene-level prediction input
#> <gcap> 2022-04-06 19:47:48 info [gcap.collapse2Genes]: please make sure the first 3 columns of `fts$fts_region` are for chr, start, end.
#> <gcap> 2022-04-06 19:47:48 info [gcap.collapse2Genes]: collapsing region-level features to gene-level
#> <gcap> 2022-04-06 19:47:48 info [collapse_to_genes]: checking input chromosome names
#> <gcap> 2022-04-06 19:47:48 info [collapse_to_genes]: reading reference file D:/Rlib/gcap/extdata/hg38_target_genes.rds
#> <gcap> 2022-04-06 19:47:48 info [collapse_to_genes]: finding overlaps
#> <gcap> 2022-04-06 19:47:48 info [collapse_to_genes]: calculating intersect size
#> <gcap> 2022-04-06 19:47:48 info [collapse_to_genes]: keeping records with >= 100% overlap ratio with a gene
#> <gcap> 2022-04-06 19:47:48 info [gcap.collapse2Genes]: merging gene-level and sample-level data
#> <gcap> 2022-04-06 19:47:48 info [gcap.collapse2Genes]: merging data and prior amplicon frequency data
#> <gcap> 2022-04-06 19:47:48 info [gcap.collapse2Genes]: done
#> <gcap> 2022-04-06 19:47:48 info [gcap.ASCNworkflow]: =======================
#> <gcap> 2022-04-06 19:47:48 info [gcap.ASCNworkflow]: Step 2: Run prediction
#> <gcap> 2022-04-06 19:47:48 info [gcap.ASCNworkflow]: =======================
#> <gcap> 2022-04-06 19:47:48 info [gcap.runPrediction]: using model file XGB_NF11.rds
#> <gcap> 2022-04-06 19:47:48 info [gcap.runPrediction]: selecting necessary features from input data
#> <gcap> 2022-04-06 19:47:48 info [gcap.runPrediction]: running prediction
#> [19:47:48] WARNING: amalgamation/../src/c_api/c_api.cc:718: `ntree_limit` is deprecated, use `iteration_range` instead.
#> <gcap> 2022-04-06 19:47:48 info [gcap.ASCNworkflow]: ====================================
#> <gcap> 2022-04-06 19:47:48 info [gcap.ASCNworkflow]: Step 3: Run scoring and summarizing
#> <gcap> 2022-04-06 19:47:48 info [gcap.ASCNworkflow]: ====================================
#> <gcap> 2022-04-06 19:47:48 info [gcap.runScoring]: checking input data type
#> <gcap> 2022-04-06 19:47:48 info [gcap.runScoring]: checking columns
#> <gcap> 2022-04-06 19:47:48 info [gcap.runScoring]: filtering out records without prob result
#> <gcap> 2022-04-06 19:47:48 info [gcap.runScoring]: joining extra annotation data
#> set fixed levels for 'amplicon_type'
#> summarizing sample...
#>   classifying samples with min_n=1
#> summarizing gene and cytoband...
#> done
#> ======================
#> A <fCNA> object
#>     case: 10
#>     gene: 155
#>     fCNA: 155
#>      |__ 104 (3) noncircular
#>      |__ 0 (0) possibly_circular
#>      |__ 51 (4) circular
#> ======================
#> <gcap> 2022-04-06 19:47:51 info [gcap.runScoring]: done
#> <gcap> 2022-04-06 19:47:51 info [gcap.ASCNworkflow]: Saving raw prediction result to C:\Users\ADMINI~1\AppData\Local\Temp\Rtmp61Kg9u/gcap_63b28486-b59f-11ec-8000-cb3394b9d359_prediction_result.rds
#> <gcap> 2022-04-06 19:47:52 info [gcap.ASCNworkflow]: Saving fCNA records and sample info to C:\Users\ADMINI~1\AppData\Local\Temp\Rtmp61Kg9u/gcap_63b28486-b59f-11ec-8000-cb3394b9d359_fCNA_records.csv, C:\Users\ADMINI~1\AppData\Local\Temp\Rtmp61Kg9u/gcap_63b28486-b59f-11ec-8000-cb3394b9d359_sample_info.csv
#> <gcap> 2022-04-06 19:47:52 info [gcap.ASCNworkflow]: ===========================================
#> <gcap> 2022-04-06 19:47:52 info [gcap.ASCNworkflow]:  Done! Thanks for using GCAP ASCN workflow
#> <gcap> 2022-04-06 19:47:52 info [gcap.ASCNworkflow]: ===========================================
#> Error in `[.data.frame`(data, , samples, drop = FALSE): undefined columns selected
# }