Run gene-level amplicon prediction

Run gene-level amplicon prediction

gcap.runPrediction(
  data,
  feature = c("without_type", "with_type"),
  target = c("circle", "nonLinear"),
  use_toy = FALSE,
  custom_model = NULL
)

Arguments

data
data to predict (data.frame/matrix format), from gcap.collapse2Genes() in general.

feature
one of 'with_type' and 'without_type' to select model whether includes cancer type or not.

target
one of 'circle' and 'nonLinear' to select model if predict circle amplicon or non-linear amplicon.

use_toy
if TRUE, use a toy model for showing example, not for practical use.

custom_model
a outer custom xgboost model for prediction.

Value

a numeric vector representing prob.

Examples

data("ec")
y_pred <- gcap.runPrediction(ec)
y_pred