GCAP workflow for gene-level amplicon prediction

GCAP workflow for gene-level amplicon prediction

gcap.workflow(
  tumourseqfile,
  normalseqfile,
  tumourname,
  normalname,
  jobname = tumourname,
  extra_info = NULL,
  include_type = FALSE,
  genome_build = c("hg38", "hg19"),
  feature = "without_type",
  target = c("circle", "nonLinear"),
  prob_cutoff = 0.5,
  outdir = getwd(),
  result_file_prefix = paste0("gcap_", uuid::UUIDgenerate(TRUE)),
  allelecounter_exe = "~/miniconda3/envs/cancerit/bin/alleleCounter",
  g1000allelesprefix = file.path("~/data/snp/1000G_loci_hg38",
    "1kg.phase3.v5a_GRCh38nounref_allele_index_chr"),
  g1000lociprefix = file.path("~/data/snp/1000G_loci_hg38",
    "1kg.phase3.v5a_GRCh38nounref_loci_chrstring_chr"),
  GCcontentfile = "~/data/snp/GC_correction_hg38.txt",
  replictimingfile = "~/data/snp/RT_correction_hg38.txt",
  nthreads = 22,
  minCounts = 10,
  BED_file = NA,
  probloci_file = NA,
  chrom_names = 1:22,
  min_base_qual = 20,
  min_map_qual = 35,
  penalty = 70,
  skip_finished_ASCAT = FALSE
)

Arguments

tumourseqfile
Full path to the tumour BAM file.

normalseqfile
Full path to the normal BAM file.

tumourname
Identifier to be used for tumour output files.

normalname
Identifier to be used for normal output files.

jobname
job name, typically an unique name for a tumor-normal pair.

extra_info
(optional) a (file containing) data.frame with 3 columns 'sample', 'age' and 'gender', for including cancer type, check parameter include_type. For gender, should be 'XX' or 'XY', also could be 0 for 'XX' and 1 for 'XY'.

include_type
if TRUE, a fourth column named 'type' should be included in extra_info, the supported cancer type includes c("BLCA", "BRCA", "CESC", "COAD", "ESCA", "GBM", "HNSC", "KICH", "KIRP", "LGG", "LIHC", "LUAD", "LUSC", "OV", "PRAD", "SARC", "SKCM", "STAD", "UCEC", "UVM").

genome_build
genome build version, should be one of 'hg38', 'hg19'.

feature
'with_type' and 'without_type' to select model whether includes cancer type or not. Can be one or both.

target
'circle' and 'nonLinear' to select model if predict circle amplicon or non-linear amplicon. Can be one or both.

prob_cutoff
a value in [0-1] to determine the cutoff for labeling response.

outdir
result output path.

result_file_prefix
file name prefix (without directory path) for storing final model prediction file in CSV format. Default a unique file name is generated by UUID approach.

allelecounter_exe
Path to the allele counter executable.

g1000allelesprefix
Prefix path to the 1000 Genomes alleles reference files.

g1000lociprefix
Prefix path to the 1000 Genomes SNP reference files.

GCcontentfile
File containing the GC content around every SNP for increasing window sizes

replictimingfile
File containing replication timing at every SNP for various cell lines (optional)

nthreads
The number of parallel processes for getting allele counts (optional, default=1).

minCounts
Minimum depth required in the normal for a SNP to be considered (optional, default=10).

BED_file
A BED file for only looking at SNPs within specific intervals (optional, default=NA).

probloci_file
A file (chromosome <tab> position; no header) containing specific loci to ignore (optional, default=NA).

chrom_names
A vector containing the names of chromosomes to be considered (optional, default=c(1:22,'X')).

min_base_qual
Minimum base quality required for a read to be counted (optional, default=20).

min_map_qual
Minimum mapping quality required for a read to be counted (optional, default=35).

penalty
penalty of introducing an additional ASPCF breakpoint (expert parameter, don't adapt unless you know what you're doing)

skip_finished_ASCAT
if TRUE, skipped finished ASCAT calls to save time.

Value

a list of data.table and corresponding files saved to local machine.